+++ src/samples.js
@@ -1,0 +1,4136 @@
+var samples =[
+    {
+	"name": "1",
+	"points": [ { x: -169.897, y: 110.723, z: 9.71903 },
+		    { x: -170.507, y: 111.062, z: 10.4725 },
+		    { x: -170.617, y: 111.308, z: 11.0072 },
+		    { x: -170.955, y: 111.776, z: 11.1872 },
+		    { x: -171.076, y: 112.004, z: 11.4294 },
+		    { x: -175.522, y: 107.898, z: 5.41265 },
+		    { x: -180.677, y: 108.086, z: 2.8646 },
+		    { x: -182.06, y: 108.587, z: 0.742934 },
+		    { x: -182.121, y: 109.74, z: 1.86046 },
+		    { x: -182.16, y: 110.446, z: 2.65619 },
+		    { x: -180.645, y: 114.183, z: 4.17831 },
+		    { x: -178.684, y: 113.5, z: 3.58282 },
+		    { x: -178.22, y: 112.636, z: 3.19227 },
+		    { x: -175.174, y: 112.583, z: 10.0256 },
+		    { x: -173.106, y: 110.688, z: 5.83266 },
+		    { x: -170.223, y: 111.594, z: 7.03603 },
+		    { x: -167.4, y: 111.88, z: 5.2312 },
+		    { x: -163.993, y: 110.809, z: 3.98124 },
+		    { x: -159.691, y: 110.152, z: 2.01416 },
+		    { x: -155.268, y: 107.407, z: -1.69046 },
+		    { x: -150.912, y: 105.941, z: -3.85341 },
+		    { x: -147.709, y: 105.544, z: -6.39012 },
+		    { x: -143.554, y: 106.632, z: -7.85637 },
+		    { x: -139.914, y: 107.652, z: -9.79192 },
+		    { x: -134.908, y: 110.107, z: -5.5642 },
+		    { x: -129.643, y: 112.352, z: -2.97844 },
+		    { x: -126.066, y: 111.513, z: -4.11704 },
+		    { x: -122.622, y: 110.114, z: -7.16365 },
+		    { x: -119.12, y: 109.51, z: -10.5297 },
+		    { x: -115.24, y: 109.47, z: -12.3013 },
+		    { x: -111.66, y: 109.433, z: -13.4373 },
+		    { x: -108.729, y: 109.616, z: -14.3415 },
+		    { x: -104.512, y: 111.758, z: -10.6368 },
+		    { x: -102.659, y: 112.409, z: -10.3111 },
+		    { x: -100.181, y: 113.121, z: -11.0817 },
+		    { x: -97.5108, y: 113.429, z: -11.8019 },
+		    { x: -94.9407, y: 113.89, z: -12.5438 },
+		    { x: -91.9617, y: 114.12, z: -13.7771 },
+		    { x: -89.8463, y: 114.531, z: -14.6595 },
+		    { x: -87.2707, y: 114.929, z: -15.2296 },
+		    { x: -87.2707, y: 114.929, z: -15.2296 },
+		    { x: -84.5202, y: 115.728, z: -16.9571 },
+		    { x: -82.9907, y: 115.862, z: -17.2954 },
+		    { x: -80.2604, y: 116.242, z: -17.4631 },
+		    { x: -78.2424, y: 116.9, z: -17.4881 },
+		    { x: -77.0874, y: 117.178, z: -17.4794 },
+		    { x: -75.5218, y: 117.8, z: -17.5202 },
+		    { x: -74.9125, y: 117.841, z: -17.7967 },
+		    { x: -73.8302, y: 117.969, z: -18.2984 },
+		    { x: -73.2454, y: 118.027, z: -18.4822 },
+		    { x: -72.1609, y: 118.064, z: -18.6806 },
+		    { x: -71.8731, y: 118.102, z: -18.7919 },
+		    { x: -71.456, y: 118.243, z: -18.8304 },
+		    { x: -71.0769, y: 118.463, z: -18.8985 },
+		    { x: -70.8378, y: 118.54, z: -18.9039 },
+		    { x: -70.5883, y: 118.703, z: -18.8815 },
+		    { x: -70.491, y: 118.809, z: -18.841 },
+		    { x: -70.4999, y: 118.893, z: -18.8262 },
+		    { x: -70.483, y: 119, z: -18.8129 },
+		    { x: -70.4675, y: 119.122, z: -18.8041 },
+		    { x: -70.5464, y: 119.718, z: -18.4996 },
+		    { x: -70.6924, y: 120.004, z: -18.4351 },
+		    { x: -70.8143, y: 120.233, z: -18.4414 },
+		    { x: -71.0289, y: 120.457, z: -18.5468 },
+		    { x: -71.1865, y: 120.621, z: -18.6292 },
+		    { x: -71.2893, y: 120.765, z: -18.642 },
+		    { x: -71.4745, y: 120.864, z: -18.585 },
+		    { x: -71.6967, y: 120.869, z: -18.2453 },
+		    { x: -71.6967, y: 120.869, z: -18.2453 },
+		    { x: -71.9159, y: 120.844, z: -17.8973 },
+		    { x: -72.0703, y: 120.783, z: -17.6801 },
+		    { x: -72.1519, y: 120.547, z: -17.5317 },
+		    { x: -72.2236, y: 120.439, z: -17.5014 },
+		    { x: -72.4173, y: 120.4, z: -17.4984 },
+		    { x: -72.4173, y: 120.4, z: -17.4984 },
+		    { x: -72.7098, y: 120.283, z: -17.5789 },
+		    { x: -72.8828, y: 120.08, z: -17.6608 },
+		    { x: -72.9969, y: 119.935, z: -17.7296 },
+		    { x: -73.1199, y: 119.809, z: -17.854 },
+		    { x: -73.1199, y: 119.809, z: -17.854 },
+		    { x: -73.259, y: 119.713, z: -17.9284 },
+		    { x: -73.3328, y: 119.64, z: -17.9873 },
+		    { x: -73.384, y: 119.584, z: -18.0439 },
+		    { x: -73.384, y: 119.584, z: -18.0439 },
+		    { x: -73.3923, y: 119.53, z: -17.9634 },
+		    { x: -73.4079, y: 119.541, z: -17.9451 },
+		    { x: -73.4185, y: 119.551, z: -17.9238 },
+		    { x: -73.4124, y: 119.579, z: -17.8825 },
+		    { x: -73.4124, y: 119.579, z: -17.8825 },
+		    { x: -73.423, y: 119.583, z: -17.8341 },
+		    { x: -73.4289, y: 119.598, z: -17.8065 },
+		    { x: -73.4184, y: 119.614, z: -17.811 },
+		    { x: -73.4184, y: 119.614, z: -17.811 },
+		    { x: -73.4512, y: 119.651, z: -17.813 },
+		    { x: -73.5201, y: 119.706, z: -17.8435 },
+		    { x: -73.583, y: 119.772, z: -17.8742 },
+		    { x: -73.583, y: 119.772, z: -17.8742 },
+		    { x: -73.6482, y: 119.824, z: -17.9118 },
+		    { x: -73.7011, y: 119.866, z: -17.9392 },
+		    { x: -73.7011, y: 119.866, z: -17.9392 },
+		    { x: -73.7665, y: 119.888, z: -17.9642 },
+		    { x: -73.8689, y: 119.89, z: -17.9572 },
+		    { x: -73.8689, y: 119.89, z: -17.9572 },
+		    { x: -73.9027, y: 119.787, z: -17.784 },
+		    { x: -73.9002, y: 119.756, z: -17.7347 },
+		    { x: -73.9002, y: 119.756, z: -17.7347 },
+		    { x: -73.865, y: 119.677, z: -17.7113 },
+		    { x: -73.828, y: 119.576, z: -17.6664 },
+		    { x: -73.828, y: 119.576, z: -17.6664 },
+		    { x: -73.7904, y: 119.501, z: -17.6383 },
+		    { x: -73.7014, y: 119.48, z: -17.667 },
+		    { x: -73.7014, y: 119.48, z: -17.667 },
+		    { x: -73.7121, y: 119.486, z: -17.7021 },
+		    { x: -73.7325, y: 119.503, z: -17.7239 }]
+    },
+    {
+	"name": "1'",
+	"points":[ { x: -5.49821, y: 78.6334, z: -26.2063 },
+		   { x: -5.52948, y: 78.7125, z: -26.2316 },
+		   { x: -5.57791, y: 78.7756, z: -26.2533 },
+		   { x: -5.77662, y: 78.8193, z: -26.2991 },
+		   { x: -5.77662, y: 78.8193, z: -26.2991 },
+		   { x: -6.2453, y: 78.673, z: -26.3833 },
+		   { x: -6.2453, y: 78.673, z: -26.3833 },
+		   { x: -6.2453, y: 78.673, z: -26.3833 },
+		   { x: -9.28235, y: 78.1665, z: -26.1817 },
+		   { x: -11.7653, y: 77.9391, z: -26.032 },
+		   { x: -15.8093, y: 77.7153, z: -26.1149 },
+		   { x: -19.8867, y: 77.3178, z: -25.9456 },
+		   { x: -23.6965, y: 77.0533, z: -25.9905 },
+		   { x: -27.3371, y: 76.903, z: -26.4744 },
+		   { x: -30.7448, y: 77.1057, z: -26.9653 },
+		   { x: -34.1866, y: 77.1754, z: -27.7272 },
+		   { x: -37.1943, y: 77.2898, z: -28.2607 },
+		   { x: -39.9003, y: 76.9707, z: -28.0312 },
+		   { x: -44.1944, y: 77.1307, z: -27.7477 },
+		   { x: -46.7447, y: 77.3582, z: -28.0361 },
+		   { x: -48.98, y: 77.6201, z: -28.2579 },
+		   { x: -48.98, y: 77.6201, z: -28.2579 },
+		   { x: -54.0152, y: 78.2074, z: -28.3032 },
+		   { x: -55.4722, y: 78.5045, z: -28.3522 },
+		   { x: -58.8505, y: 79.2047, z: -28.2035 },
+		   { x: -60.4914, y: 79.5283, z: -28.1098 },
+		   { x: -62.1869, y: 79.9195, z: -28.0594 },
+		   { x: -66.556, y: 81.8883, z: -27.8667 },
+		   { x: -68.0582, y: 82.4068, z: -27.7089 },
+		   { x: -69.3876, y: 82.6286, z: -27.5596 },
+		   { x: -72.7765, y: 83.3484, z: -27.3354 },
+		   { x: -75.1642, y: 84.1344, z: -27.1247 },
+		   { x: -77.2407, y: 84.5927, z: -26.9357 },
+		   { x: -79.0818, y: 84.8513, z: -26.8392 },
+		   { x: -83.2541, y: 85.6206, z: -26.4423 },
+		   { x: -84.5939, y: 85.3529, z: -26.4086 },
+		   { x: -85.2139, y: 84.6321, z: -26.5788 },
+		   { x: -86.5362, y: 83.8858, z: -27.0837 },
+		   { x: -88.5034, y: 82.1214, z: -27.9099 },
+		   { x: -89.4185, y: 81.7871, z: -28.1833 },
+		   { x: -90.0946, y: 81.4322, z: -28.4012 },
+		   { x: -91.0274, y: 81.2323, z: -28.6589 },
+		   { x: -91.7015, y: 80.8915, z: -28.8159 },
+		   { x: -93.613, y: 80.5233, z: -29.1309 },
+		   { x: -94.6898, y: 80.4877, z: -29.3185 },
+		   { x: -95.6048, y: 80.4677, z: -29.3892 },
+		   { x: -96.3342, y: 80.4055, z: -29.4017 },
+		   { x: -97.1369, y: 80.4424, z: -29.4016 },
+		   { x: -97.8216, y: 80.4443, z: -29.4399 },
+		   { x: -98.4668, y: 80.3403, z: -29.5509 },
+		   { x: -99.6753, y: 80.1554, z: -29.9601 },
+		   { x: -100.326, y: 80.1604, z: -30.2022 },
+		   { x: -100.909, y: 80.1327, z: -30.3774 },
+		   { x: -101.446, y: 80.135, z: -30.5367 },
+		   { x: -101.936, y: 80.1226, z: -30.6626 },
+		   { x: -102.391, y: 80.1389, z: -30.7779 },
+		   { x: -102.979, y: 80.2118, z: -30.8572 },
+		   { x: -103.396, y: 80.2908, z: -30.9061 },
+		   { x: -103.667, y: 80.3634, z: -30.9308 },
+		   { x: -103.869, y: 80.4229, z: -30.9449 },
+		   { x: -104.034, y: 80.4955, z: -30.9497 },
+		   { x: -104.276, y: 80.5695, z: -31.0371 },
+		   { x: -104.276, y: 80.5695, z: -31.0371 },
+		   { x: -104.466, y: 80.6354, z: -31.104 },
+		   { x: -104.621, y: 80.7101, z: -31.2203 },
+		   { x: -104.787, y: 80.7739, z: -31.2982 },
+		   { x: -104.954, y: 80.8763, z: -31.3855 },
+		   { x: -105.064, y: 80.9943, z: -31.4612 },
+		   { x: -105.118, y: 81.1217, z: -31.5125 },
+		   { x: -105.118, y: 81.1217, z: -31.5125 },
+		   { x: -105.17, y: 81.2404, z: -31.5495 },
+		   { x: -105.175, y: 81.3767, z: -31.6001 },
+		   { x: -105.21, y: 81.4705, z: -31.6269 },
+		   { x: -105.21, y: 81.4705, z: -31.6269 },
+		   { x: -105.245, y: 81.5584, z: -31.6476 },
+		   { x: -105.276, y: 81.64, z: -31.6656 },
+		   { x: -105.276, y: 81.64, z: -31.6656 },
+		   { x: -105.317, y: 81.7096, z: -31.6778 },
+		   { x: -105.374, y: 81.7736, z: -31.6836 },
+		   { x: -105.879, y: 81.8755, z: -31.2604 },
+		   { x: -105.879, y: 81.8755, z: -31.2604 },
+		   { x: -106.603, y: 81.9498, z: -30.7864 },
+		   { x: -107.295, y: 81.8419, z: -30.4447 },
+		   { x: -107.295, y: 81.8419, z: -30.4447 },
+		   { x: -107.822, y: 81.8225, z: -30.1519 },
+		   { x: -108.123, y: 81.8869, z: -29.9863 },
+		   { x: -108.423, y: 81.9583, z: -29.8568 },
+		   { x: -108.423, y: 81.9583, z: -29.8568 },
+		   { x: -108.879, y: 82.0299, z: -29.6431 },
+		   { x: -109.195, y: 82.1427, z: -29.4541 },
+		   { x: -109.651, y: 82.4299, z: -29.2034 },
+		   { x: -109.651, y: 82.4299, z: -29.2034 },
+		   { x: -109.935, y: 82.66, z: -29.0306 },
+		   { x: -110.328, y: 82.9797, z: -28.7728 },
+		   { x: -110.852, y: 83.3693, z: -28.3189 },
+		   { x: -110.852, y: 83.3693, z: -28.3189 },
+		   { x: -111.254, y: 83.7424, z: -27.9515 },
+		   { x: -111.533, y: 84.0235, z: -27.6547 },
+		   { x: -111.722, y: 84.3248, z: -27.3609 },
+		   { x: -111.927, y: 84.6619, z: -27.0418 },
+		   { x: -111.927, y: 84.6619, z: -27.0418 },
+		   { x: -112.016, y: 84.88, z: -26.8067 },
+		   { x: -112.026, y: 85.1114, z: -26.5088 },
+		   { x: -112.003, y: 85.3495, z: -26.2596 },
+		   { x: -112.003, y: 85.3495, z: -26.2596 },
+		   { x: -111.946, y: 85.5612, z: -25.9712 },
+		   { x: -111.81, y: 85.7548, z: -25.6604 },
+		   { x: -111.698, y: 85.8947, z: -25.5163 },
+		   { x: -111.698, y: 85.8947, z: -25.5163 },
+		   { x: -111.569, y: 85.9708, z: -25.4331 },
+		   { x: -111.475, y: 86.0035, z: -25.4088 },
+		   { x: -111.353, y: 86.0536, z: -25.4038 },
+		   { x: -111.353, y: 86.0536, z: -25.4038 },
+		   { x: -111.219, y: 86.153, z: -25.392 },
+		   { x: -111.016, y: 86.2548, z: -25.386 },
+		   { x: -110.785, y: 86.4939, z: -25.4113 },
+		   { x: -110.785, y: 86.4939, z: -25.4113 },
+		   { x: -110.474, y: 86.6756, z: -25.3591 },
+		   { x: -110.247, y: 86.7959, z: -25.3581 },
+		   { x: -110.247, y: 86.7959, z: -25.3581 },
+		   { x: -109.98, y: 86.9275, z: -25.3555 },
+		   { x: -109.751, y: 87.11, z: -25.399 },
+		   { x: -109.509, y: 87.2399, z: -25.4808 },
+		   { x: -109.509, y: 87.2399, z: -25.4808 },
+		   { x: -109.333, y: 87.3758, z: -25.5821 },
+		   { x: -109.077, y: 87.5787, z: -25.6851 },
+		   { x: -108.828, y: 87.6949, z: -25.8462 },
+		   { x: -108.828, y: 87.6949, z: -25.8462 },
+		   { x: -108.604, y: 87.7758, z: -25.9818 } ]
+    },
+    {
+	"name":"2",
+	"points":[{ x: -130.382, y: 60.6884, z: -25.1259 },
+		  { x: -130.284, y: 59.9657, z: -25.4692 },
+		  { x: -130.345, y: 59.0702, z: -25.7096 },
+		  { x: -130.298, y: 58.2307, z: -26.0376 },
+		  { x: -130.333, y: 58.0954, z: -25.9724 },
+		  { x: -130.333, y: 58.0954, z: -25.9724 },
+		  { x: -130.294, y: 58.4077, z: -25.9471 },
+		  { x: -130.275, y: 58.4882, z: -25.9173 },
+		  { x: -130.447, y: 58.8889, z: -25.712 },
+		  { x: -130.549, y: 59.2317, z: -25.5667 },
+		  { x: -130.549, y: 59.2317, z: -25.5667 },
+		  { x: -130.614, y: 59.4492, z: -25.2569 },
+		  { x: -130.761, y: 59.5723, z: -24.9668 },
+		  { x: -130.814, y: 60.2464, z: -24.5355 },
+		  { x: -130.766, y: 60.2648, z: -24.3478 },
+		  { x: -130.766, y: 60.2648, z: -24.3478 },
+		  { x: -130.766, y: 60.2648, z: -24.3478 },
+		  { x: -130.766, y: 60.2648, z: -24.3478 },
+		  { x: -130.303, y: 60.2986, z: -24.5937 },
+		  { x: -129.886, y: 60.4198, z: -24.7928 },
+		  { x: -129.18, y: 61.9526, z: -25.7018 },
+		  { x: -127.786, y: 63.4496, z: -27.4843 },
+		  { x: -126.57, y: 67.0598, z: -28.3656 },
+		  { x: -125.785, y: 71.213, z: -30.1469 },
+		  { x: -124.567, y: 73.8108, z: -31.3435 },
+		  { x: -123.5, y: 75.8607, z: -32.1164 },
+		  { x: -117.614, y: 79.4257, z: -32.298 },
+		  { x: -115.245, y: 81.8037, z: -32.3222 },
+		  { x: -111.108, y: 83.2379, z: -32.8824 },
+		  { x: -103.254, y: 87.2311, z: -33.8244 },
+		  { x: -99.6948, y: 89.4513, z: -32.5077 },
+		  { x: -92.2322, y: 90.6416, z: -35.8502 },
+		  { x: -88.1123, y: 92.6509, z: -36.9081 },
+		  { x: -86.2532, y: 93.9331, z: -36.7758 },
+		  { x: -83.5366, y: 96.9547, z: -34.2345 },
+		  { x: -80.3296, y: 97.4469, z: -33.9321 },
+		  { x: -77.0051, y: 97.2943, z: -34.4913 },
+		  { x: -74.6756, y: 97.2866, z: -33.809 },
+		  { x: -71.0388, y: 96.5402, z: -32.9646 },
+		  { x: -68.198, y: 95.6956, z: -33.1434 },
+		  { x: -66.8633, y: 94.7617, z: -33.671 },
+		  { x: -64.1543, y: 93.5486, z: -33.9338 },
+		  { x: -60.57, y: 90.9604, z: -31.7759 },
+		  { x: -59.6126, y: 90.1954, z: -31.4151 },
+		  { x: -56.818, y: 87.9607, z: -29.7639 },
+		  { x: -53.8835, y: 86.5776, z: -28.4129 },
+		  { x: -52.8044, y: 85.3108, z: -27.2957 },
+		  { x: -51.2896, y: 82.8385, z: -25.8611 },
+		  { x: -50.5186, y: 79.631, z: -25.5866 },
+		  { x: -49.7945, y: 77.8317, z: -25.0331 },
+		  { x: -49.1418, y: 76.2037, z: -24.4792 },
+		  { x: -47.8892, y: 73.1082, z: -23.5697 },
+		  { x: -47.1608, y: 71.42, z: -23.0299 },
+		  { x: -45.5248, y: 68.3841, z: -21.9964 },
+		  { x: -44.7484, y: 67.1902, z: -21.7384 },
+		  { x: -43.7063, y: 65.787, z: -21.3726 },
+		  { x: -40.8163, y: 62.4016, z: -20.2965 },
+		  { x: -39.1321, y: 60.8354, z: -19.9632 },
+		  { x: -37.372, y: 59.3952, z: -19.67 },
+		  { x: -32.6235, y: 57.0171, z: -19.0259 },
+		  { x: -30.0197, y: 56.0708, z: -18.7477 },
+		  { x: -26.4354, y: 54.5929, z: -17.6461 },
+		  { x: -20.47, y: 52.4633, z: -16.3442 },
+		  { x: -17.9166, y: 51.8204, z: -16.2537 },
+		  { x: -15.32, y: 51.3812, z: -16.2035 },
+		  { x: -9.80093, y: 51.3942, z: -16.2261 },
+		  { x: -7.07565, y: 51.6363, z: -16.3971 },
+		  { x: -4.45838, y: 52.0284, z: -16.9623 },
+		  { x: 0.288199, y: 53.2732, z: -18.4892 },
+		  { x: 2.62024, y: 54.3049, z: -19.8787 },
+		  { x: 4.62963, y: 55.5311, z: -20.8641 },
+		  { x: 8.1605, y: 58.392, z: -22.5586 },
+		  { x: 10.0035, y: 60.3873, z: -23.8755 },
+		  { x: 13.014, y: 64.1333, z: -25.398 },
+		  { x: 15.1253, y: 67.0244, z: -26.1821 },
+		  { x: 16.7129, y: 69.4125, z: -26.6193 },
+		  { x: 21.1213, y: 77.1728, z: -28.4314 },
+		  { x: 24.0926, y: 81.2212, z: -29.1281 },
+		  { x: 28.4763, y: 87.8802, z: -30.2727 },
+		  { x: 30.0892, y: 91.178, z: -30.6445 },
+		  { x: 33.4131, y: 96.164, z: -31.2457 },
+		  { x: 34.8124, y: 97.9087, z: -31.4924 },
+		  { x: 37.9483, y: 101.545, z: -31.4075 },
+		  { x: 39.5153, y: 103.054, z: -31.5208 },
+		  { x: 40.7654, y: 103.969, z: -31.7814 },
+		  { x: 43.7182, y: 107.055, z: -32.8792 },
+		  { x: 46.2614, y: 108.556, z: -32.9114 },
+		  { x: 47.6838, y: 109.555, z: -33.1784 },
+		  { x: 50.0345, y: 111.106, z: -32.7561 },
+		  { x: 50.2155, y: 110.868, z: -32.7909 },
+		  { x: 52.5523, y: 111.311, z: -32.6446 },
+		  { x: 55.8623, y: 111.763, z: -32.3743 },
+		  { x: 57.0991, y: 111.797, z: -32.1642 },
+		  { x: 60.1912, y: 111.704, z: -31.6771 },
+		  { x: 61.2451, y: 111.474, z: -31.4869 },
+		  { x: 62.2153, y: 110.935, z: -31.2571 },
+		  { x: 64.0579, y: 109.289, z: -29.6299 },
+		  { x: 64.9718, y: 108.499, z: -28.9816 },
+		  { x: 66.8467, y: 106.654, z: -27.8468 },
+		  { x: 68.2396, y: 105.989, z: -27.1872 },
+		  { x: 69.3754, y: 105.038, z: -26.8059 },
+		  { x: 72.2758, y: 101.241, z: -23.7524 },
+		  { x: 73.212, y: 99.7098, z: -23.0424 },
+		  { x: 73.9516, y: 98.401, z: -22.4087 },
+		  { x: 75.9464, y: 94.5122, z: -19.9794 },
+		  { x: 75.9464, y: 94.5122, z: -19.9794 },
+		  { x: 75.9464, y: 94.5122, z: -19.9794 },
+		  { x: 79.2818, y: 89.7382, z: -17.997 },
+		  { x: 79.6914, y: 88.2823, z: -17.6044 },
+		  { x: 81.4922, y: 85.5857, z: -16.26 },
+		  { x: 82.266, y: 84.6134, z: -15.6898 },
+		  { x: 83.4972, y: 83.2214, z: -14.9115 },
+		  { x: 85.6551, y: 80.782, z: -13.4341 },
+		  { x: 86.3165, y: 79.771, z: -12.9199 },
+		  { x: 87.737, y: 77.8491, z: -11.943 },
+		  { x: 88.9444, y: 75.9842, z: -10.2107 },
+		  { x: 90.6611, y: 72.8153, z: -7.26881 },
+		  { x: 91.4034, y: 71.318, z: -6.12351 },
+		  { x: 92.2931, y: 69.7608, z: -5.00147 },
+		  { x: 93.2631, y: 68.443, z: -3.91191 },
+		  { x: 95.1757, y: 65.8408, z: -2.15077 },
+		  { x: 96.5679, y: 64.8884, z: -1.29994 },
+		  { x: 97.835, y: 63.72, z: -0.543963 },
+		  { x: 99.4469, y: 63.1042, z: 0.226746 },
+		  { x: 102.811, y: 62.0585, z: 1.75432 },
+		  { x: 104.863, y: 61.6943, z: 2.65319 },
+		  { x: 106.395, y: 61.4167, z: 2.89493 },
+		  { x: 107.377, y: 60.7992, z: 3.18109 },
+		  { x: 108.003, y: 60.4076, z: 3.38236 },
+		  { x: 112.413, y: 60.8613, z: 4.03899 },
+		  { x: 114.426, y: 61.2223, z: 4.11784 },
+		  { x: 115.418, y: 61.3503, z: 3.83989 },
+		  { x: 118.149, y: 62.206, z: 3.89884 },
+		  { x: 119.331, y: 62.5849, z: 3.98838 },
+		  { x: 120.058, y: 62.8704, z: 4.03365 },
+		  { x: 124.54, y: 65.0527, z: 2.91862 },
+		  { x: 126.574, y: 66.2851, z: 2.14777 },
+		  { x: 128.094, y: 66.8765, z: 1.65742 },
+		  { x: 131.955, y: 68.9018, z: 2.34554 },
+		  { x: 133.274, y: 70.139, z: 2.95247 },
+		  { x: 137.77, y: 72.8698, z: 1.54415 },
+		  { x: 138.587, y: 73.6832, z: 0.340906 },
+		  { x: 140.406, y: 75.262, z: -0.108547 },
+		  { x: 141.631, y: 76.5683, z: -0.219539 },
+		  { x: 144.038, y: 80.2604, z: -0.767319 },
+		  { x: 144.712, y: 81.3866, z: -1.38398 },
+		  { x: 146.424, y: 83.4582, z: -3.01039 },
+		  { x: 148.432, y: 86.0299, z: -2.35742 },
+		  { x: 149.141, y: 86.8772, z: -2.17715 },
+		  { x: 150.476, y: 88.7761, z: -1.7068 },
+		  { x: 151.463, y: 89.9018, z: -1.31568 },
+		  { x: 155.795, y: 96.5916, z: 0.340873 },
+		  { x: 158.473, y: 100.31, z: 1.81431 },
+		  { x: 158.558, y: 101.55, z: 2.41402 },
+		  { x: 149.005, y: 146.087, z: 42.0762 },
+		  { x: 152.597, y: 148.847, z: 41.2229 },
+		  { x: 151.876, y: 148.381, z: 37.7849 },
+		  { x: 151.889, y: 148.904, z: 36.2001 },
+		  { x: 151.552, y: 148.994, z: 34.9842 },
+		  { x: 151.823, y: 149.024, z: 34.7764 },
+		  { x: 152.091, y: 149.852, z: 33.2284 },
+		  { x: 154.081, y: 150.909, z: 34.7266 },
+		  { x: 157.197, y: 151.637, z: 32.6142 },
+		  { x: 160.033, y: 152.526, z: 28.8703 },
+		  { x: 160.525, y: 152.641, z: 27.65 },
+		  { x: 160.761, y: 151.689, z: 27.1135 },
+		  { x: 160.689, y: 151.489, z: 27.3584 },
+		  { x: 162.513, y: 151.229, z: 27.5742 },
+		  { x: 164.405, y: 150.743, z: 27.4592 },
+		  { x: 167.415, y: 150.625, z: 29.6871 },
+		  { x: 168.525, y: 151, z: 35.5181 },
+		  { x: 170.084, y: 150.683, z: 39.2324 },
+		  { x: 172.723, y: 151.468, z: 52.8199 },
+		  { x: 169.88, y: 148.629, z: 65.4089 },
+		  { x: 199.03, y: 154.204, z: 26.7918 },
+		  { x: 215.998, y: 152.744, z: 16.5589 },
+		  { x: 211.929, y: 149.056, z: 1.30937 },
+		  { x: 214.494, y: 139.357, z: -10.5868 },
+		  { x: 217.212, y: 131.784, z: -19.5229 },
+		  { x: 216.906, y: 131.467, z: -23.3245 },
+		  { x: 215.678, y: 130.023, z: -26.0229 },
+		  { x: 215.426, y: 128.425, z: -27.2866 },
+		  { x: 214.439, y: 126.678, z: -29.0332 },
+		  { x: 213.804, y: 126.083, z: -29.841 },
+		  { x: 212.065, y: 125.263, z: -31.4138 },
+		  { x: 211.603, y: 119.102, z: -31.2428 },
+		  { x: 211.467, y: 119.281, z: -30.6697 },
+		  { x: 209.129, y: 120.399, z: -31.0082 },
+		  { x: 207.51, y: 119.145, z: -25.9214 } ]
+    },
+    {
+	"name":"2'",
+	"points":[{ x: 84.709, y: 83.0414, z: -16.9649 },
+		  { x: 84.7107, y: 83.1293, z: -16.9486 },
+		  { x: 84.7164, y: 83.2446, z: -16.9165 },
+		  { x: 84.7164, y: 83.2446, z: -16.9165 },
+		  { x: 84.7072, y: 83.5895, z: -16.7793 },
+		  { x: 84.1488, y: 83.4567, z: -16.3225 },
+		  { x: 84.1488, y: 83.4567, z: -16.3225 },
+		  { x: 83.9025, y: 83.4222, z: -16.0982 },
+		  { x: 83.7962, y: 83.446, z: -15.9807 },
+		  { x: 83.7513, y: 83.5051, z: -15.9049 },
+		  { x: 83.7217, y: 83.5423, z: -15.8508 },
+		  { x: 83.7217, y: 83.5423, z: -15.8508 },
+		  { x: 83.864, y: 83.5059, z: -15.8599 },
+		  { x: 84.0149, y: 83.43, z: -16.216 },
+		  { x: 84.2549, y: 82.9939, z: -16.4128 },
+		  { x: 84.3007, y: 82.4681, z: -16.4199 },
+		  { x: 84.1111, y: 81.5027, z: -16.4198 },
+		  { x: 83.6977, y: 80.4741, z: -16.4005 },
+		  { x: 83.6977, y: 80.4741, z: -16.4005 },
+		  { x: 82.8716, y: 78.7443, z: -16.1448 },
+		  { x: 81.8302, y: 76.6475, z: -15.8848 },
+		  { x: 81.8302, y: 76.6475, z: -15.8848 },
+		  { x: 81.8302, y: 76.6475, z: -15.8848 },
+		  { x: 78.7888, y: 71.5377, z: -14.8853 },
+		  { x: 76.3977, y: 68.6326, z: -13.8564 },
+		  { x: 75.1554, y: 67.193, z: -13.6095 },
+		  { x: 73.9286, y: 65.9638, z: -13.5129 },
+		  { x: 72.8046, y: 64.8847, z: -13.3922 },
+		  { x: 71.3715, y: 63.6875, z: -13.0722 },
+		  { x: 70.0911, y: 62.7116, z: -12.7617 },
+		  { x: 69.1077, y: 61.9936, z: -12.5683 },
+		  { x: 68.543, y: 61.5753, z: -12.4652 },
+		  { x: 67.288, y: 60.8124, z: -12.4098 },
+		  { x: 65.0345, y: 59.6363, z: -12.2212 },
+		  { x: 63.9761, y: 59.235, z: -12.1871 },
+		  { x: 62.9647, y: 58.9322, z: -12.1868 },
+		  { x: 61.6439, y: 58.6945, z: -12.2905 },
+		  { x: 60.4324, y: 58.5685, z: -12.4465 },
+		  { x: 59.1027, y: 58.4108, z: -12.7101 },
+		  { x: 58.2564, y: 58.3442, z: -12.8795 },
+		  { x: 56.9474, y: 58.6054, z: -13.1612 },
+		  { x: 55.5925, y: 58.9587, z: -13.4541 },
+		  { x: 54.4866, y: 59.5914, z: -14.0025 },
+		  { x: 53.4952, y: 60.313, z: -14.4636 },
+		  { x: 52.5043, y: 61.2612, z: -15.0318 },
+		  { x: 51.4773, y: 62.397, z: -15.6435 },
+		  { x: 50.3899, y: 63.8418, z: -16.4154 },
+		  { x: 49.385, y: 65.3187, z: -17.1128 },
+		  { x: 48.3188, y: 66.9041, z: -17.768 },
+		  { x: 46.3336, y: 70.356, z: -18.9614 },
+		  { x: 45.3928, y: 72.1959, z: -19.5149 },
+		  { x: 44.3662, y: 74.1713, z: -20.213 },
+		  { x: 43.4476, y: 75.8862, z: -20.8629 },
+		  { x: 42.5995, y: 77.6281, z: -21.5955 },
+		  { x: 41.8683, y: 79.0428, z: -22.0178 },
+		  { x: 40.8476, y: 80.2212, z: -22.6032 },
+		  { x: 39.7888, y: 81.6247, z: -22.7828 },
+		  { x: 38.5652, y: 83.2091, z: -22.949 },
+		  { x: 37.4973, y: 84.7998, z: -22.9426 },
+		  { x: 35.6653, y: 88.0572, z: -22.7011 },
+		  { x: 34.5365, y: 89.2664, z: -22.1104 },
+		  { x: 33.3034, y: 90.8031, z: -21.694 },
+		  { x: 31.8429, y: 92.6604, z: -21.6977 },
+		  { x: 30.314, y: 94.2415, z: -21.5407 },
+		  { x: 28.8231, y: 95.6768, z: -21.7393 },
+		  { x: 25.9806, y: 97.1684, z: -21.7385 },
+		  { x: 24.6061, y: 97.7101, z: -21.6653 },
+		  { x: 22.7438, y: 98.3528, z: -21.5195 },
+		  { x: 20.4722, y: 98.8184, z: -21.2547 },
+		  { x: 16.2866, y: 99.4169, z: -20.6975 },
+		  { x: 14.3401, y: 99.4084, z: -20.5895 },
+		  { x: 12.6895, y: 99.3633, z: -20.3847 },
+		  { x: 10.9978, y: 99.1243, z: -20.2443 },
+		  { x: 7.7631, y: 97.7869, z: -20.3602 },
+		  { x: 6.20055, y: 96.6146, z: -20.0748 },
+		  { x: 4.67344, y: 95.3877, z: -19.9331 },
+		  { x: 3.44942, y: 94.2517, z: -19.8864 },
+		  { x: 2.45583, y: 92.9041, z: -19.4195 },
+		  { x: 0.622955, y: 89.7538, z: -18.6025 },
+		  { x: -0.275261, y: 87.7753, z: -18.3455 },
+		  { x: -1.0993, y: 85.5079, z: -18.1047 },
+		  { x: -1.79674, y: 83.0511, z: -17.8374 },
+		  { x: -2.41089, y: 80.619, z: -17.5088 },
+		  { x: -3.15914, y: 75.8219, z: -16.6953 },
+		  { x: -3.54167, y: 73.1468, z: -16.1643 },
+		  { x: -3.92171, y: 70.9604, z: -15.6687 },
+		  { x: -4.32428, y: 68.8652, z: -15.0988 },
+		  { x: -5.37323, y: 64.3017, z: -13.6043 },
+		  { x: -6.13641, y: 61.9418, z: -12.9761 },
+		  { x: -6.91499, y: 59.4841, z: -12.3577 },
+		  { x: -7.84419, y: 57.1707, z: -11.633 },
+		  { x: -9.95421, y: 53.4681, z: -10.7554 },
+		  { x: -11.0974, y: 51.6709, z: -10.3654 },
+		  { x: -12.2558, y: 49.9968, z: -10.1484 },
+		  { x: -14.5676, y: 47.1628, z: -9.63174 },
+		  { x: -15.6804, y: 45.9053, z: -9.39821 },
+		  { x: -16.7472, y: 44.836, z: -9.1468 },
+		  { x: -17.7659, y: 43.8554, z: -8.94992 },
+		  { x: -19.7932, y: 42.2046, z: -8.65554 },
+		  { x: -20.8106, y: 41.6463, z: -8.54192 },
+		  { x: -21.7712, y: 41.1813, z: -8.49005 },
+		  { x: -23.6821, y: 40.8081, z: -8.56184 },
+		  { x: -24.846, y: 40.8451, z: -8.87373 },
+		  { x: -25.9821, y: 41.0724, z: -9.17939 },
+		  { x: -27.092, y: 41.5222, z: -9.53466 },
+		  { x: -28.2161, y: 42.1727, z: -9.94824 },
+		  { x: -30.7834, y: 44.1708, z: -11.2456 },
+		  { x: -32.0714, y: 45.3973, z: -11.7795 },
+		  { x: -33.2299, y: 46.6989, z: -12.3693 },
+		  { x: -34.3864, y: 48.9065, z: -13.5597 },
+		  { x: -35.447, y: 50.7763, z: -14.3228 },
+		  { x: -36.4274, y: 52.9254, z: -14.6491 },
+		  { x: -37.4055, y: 54.9412, z: -15.2665 },
+		  { x: -38.2332, y: 56.8088, z: -16.0813 },
+		  { x: -39.9379, y: 61.0051, z: -17.7998 },
+		  { x: -40.7089, y: 63.4027, z: -18.6487 },
+		  { x: -41.439, y: 65.7129, z: -19.5174 },
+		  { x: -41.9065, y: 67.7623, z: -20.168 },
+		  { x: -42.4143, y: 69.9985, z: -20.8225 },
+		  { x: -43.1052, y: 72.4997, z: -21.7289 },
+		  { x: -43.6237, y: 77.8838, z: -22.1969 },
+		  { x: -44.0473, y: 80.565, z: -21.8863 },
+		  { x: -43.6006, y: 82.728, z: -23.5766 },
+		  { x: -44.573, y: 85.3082, z: -24.2888 },
+		  { x: -47.1005, y: 88.394, z: -24.7161 },
+		  { x: -49.0709, y: 89.7678, z: -25.4368 },
+		  { x: -50.6677, y: 92.6851, z: -25.3969 },
+		  { x: -51.6578, y: 94.7741, z: -25.5622 },
+		  { x: -53.0078, y: 97.7236, z: -25.7 },
+		  { x: -53.0078, y: 97.7236, z: -25.7 },
+		  { x: -57.9232, y: 106.948, z: -26.502 },
+		  { x: -58.6992, y: 108.462, z: -26.3252 },
+		  { x: -61.0398, y: 111.679, z: -25.6316 },
+		  { x: -62.2772, y: 112.734, z: -25.6917 },
+		  { x: -63.1954, y: 113.25, z: -25.8289 },
+		  { x: -65.9307, y: 114.525, z: -26.282 },
+		  { x: -67.118, y: 114.922, z: -26.2857 },
+		  { x: -67.7686, y: 113.93, z: -28.331 },
+		  { x: -68.8514, y: 114.275, z: -27.7046 },
+		  { x: -70.0291, y: 114.886, z: -26.5632 },
+		  { x: -70.6482, y: 114.877, z: -26.1963 },
+		  { x: -72.1952, y: 114.485, z: -25.4561 },
+		  { x: -73.4939, y: 113.954, z: -22.8479 },
+		  { x: -74.76, y: 113.034, z: -21.2763 },
+		  { x: -76.5615, y: 110.396, z: -19.7303 },
+		  { x: -77.5266, y: 108.788, z: -18.727 },
+		  { x: -78.0246, y: 106.1, z: -19.6481 },
+		  { x: -77.3765, y: 101.271, z: -20.6612 },
+		  { x: -79.8225, y: 97.0111, z: -17.4031 },
+		  { x: -80.6778, y: 95.2303, z: -16.2772 },
+		  { x: -80.8141, y: 93.2078, z: -15.8447 },
+		  { x: -82.038, y: 88.391, z: -13.7328 },
+		  { x: -82.2452, y: 86.2467, z: -13.2717 },
+		  { x: -82.284, y: 84.3739, z: -12.6705 },
+		  { x: -82.5799, y: 80.4857, z: -11.1943 },
+		  { x: -82.4963, y: 78.0676, z: -10.439 },
+		  { x: -82.2151, y: 73.0384, z: -8.66175 },
+		  { x: -82.1774, y: 70.7968, z: -7.89087 },
+		  { x: -82.1608, y: 68.8039, z: -7.33154 },
+		  { x: -81.7667, y: 64.482, z: -6.31092 },
+		  { x: -81.5444, y: 62.3691, z: -5.83066 },
+		  { x: -81.1865, y: 58.1721, z: -4.82629 },
+		  { x: -81.1238, y: 56.2532, z: -4.29619 },
+		  { x: -81.1707, y: 52.5481, z: -3.42039 },
+		  { x: -81.0644, y: 50.6363, z: -3.03588 },
+		  { x: -81.0282, y: 48.5925, z: -2.64338 },
+		  { x: -80.814, y: 44.5146, z: -2.24009 },
+		  { x: -80.8429, y: 42.4695, z: -1.81699 },
+		  { x: -81.041, y: 40.0356, z: -1.13666 },
+		  { x: -81.0231, y: 33.5966, z: -0.150273 },
+		  { x: -81.3602, y: 30.0963, z: 0.702549 },
+		  { x: -81.9341, y: 25.9131, z: 1.74682 },
+		  { x: -83.2538, y: 19.0668, z: 2.54817 },
+		  { x: -84.2626, y: 17.2241, z: 2.57129 },
+		  { x: -86.8003, y: 14.709, z: 0.378109 },
+		  { x: -88.0616, y: 14.1414, z: -0.697356 },
+		  { x: -89.4592, y: 14.4272, z: -0.811879 },
+		  { x: -91.8516, y: 14.9419, z: -0.906064 },
+		  { x: -93.4016, y: 15.3005, z: -0.727004 },
+		  { x: -97.0903, y: 18.6919, z: -0.692938 },
+		  { x: -98.874, y: 20.7057, z: -0.828964 },
+		  { x: -100.98, y: 23.0138, z: -0.896264 },
+		  { x: -104.822, y: 28.4448, z: -1.66023 },
+		  { x: -106.869, y: 31.4637, z: -2.00162 },
+		  { x: -108.9, y: 37.019, z: -2.84749 },
+		  { x: -110.331, y: 41.0981, z: -2.89389 },
+		  { x: -112.896, y: 50.1971, z: -3.11505 },
+		  { x: -114.015, y: 53.417, z: -3.01968 },
+		  { x: -115.513, y: 57.7918, z: -3.09584 },
+		  { x: -116.102, y: 59.5189, z: -3.2328 },
+		  { x: -117.32, y: 63.1243, z: -3.91794 },
+		  { x: -117.979, y: 65.2653, z: -4.55273 },
+		  { x: -118.624, y: 68.303, z: -5.49243 },
+		  { x: -118.819, y: 69.6373, z: -5.90725 },
+		  { x: -119.782, y: 73.4971, z: -7.46436 },
+		  { x: -120.159, y: 75.3637, z: -8.13945 },
+		  { x: -120.333, y: 78.1635, z: -9.01039 },
+		  { x: -120.358, y: 79.6435, z: -9.31608 },
+		  { x: -120.456, y: 82.9809, z: -10.1068 },
+		  { x: -120.837, y: 84.6398, z: -10.3859 },
+		  { x: -120.722, y: 87.1921, z: -10.8789 },
+		  { x: -120.079, y: 88.2827, z: -10.4155 },
+		  { x: -119.985, y: 89.2802, z: -10.2671 },
+		  { x: -120.049, y: 92.1643, z: -10.4917 },
+		  { x: -120.136, y: 93.6476, z: -10.5363 },
+		  { x: -120.082, y: 94.6394, z: -10.2896 },
+		  { x: -120.078, y: 95.2185, z: -10.2864 },
+		  { x: -120.688, y: 96.8392, z: -11.0942 },
+		  { x: -121.07, y: 97.5083, z: -11.3722 },
+		  { x: -121.196, y: 97.671, z: -11.427 },
+		  { x: -120.89, y: 97.9183, z: -10.6174 },
+		  { x: -120.554, y: 98.0892, z: -9.85099 },
+		  { x: -120.678, y: 98.0333, z: -9.96111 },
+		  { x: -120.698, y: 97.8475, z: -9.97266 },
+		  { x: -120.78, y: 97.5651, z: -9.88381 },
+		  { x: -120.638, y: 96.934, z: -8.95158 },
+		  { x: -120.48, y: 96.2299, z: -7.88871 },
+		  { x: -120.562, y: 94.9228, z: -6.80518 },
+		  { x: -120.562, y: 94.9228, z: -6.80518 },
+		  { x: -120.514, y: 94.3843, z: -6.66538 },
+		  { x: -120.494, y: 93.8599, z: -6.57711 },
+		  { x: -120.266, y: 93.1794, z: -6.2616 },
+		  { x: -120.194, y: 92.7392, z: -6.22629 },
+		  { x: -120.194, y: 92.7392, z: -6.22629 },
+		  { x: -120.194, y: 92.7392, z: -6.22629 },
+		  { x: -120.076, y: 92.321, z: -6.30114 },
+		  { x: -120.019, y: 92.2023, z: -6.3131 },
+		  { x: -119.87, y: 92.0597, z: -6.37326 },
+		  { x: -119.728, y: 92.0275, z: -6.76235 },
+		  { x: -119.545, y: 92.042, z: -7.10802 },
+		  { x: -119.421, y: 92.0807, z: -7.63455 },
+		  { x: -119.278, y: 92.119, z: -7.98742 },
+		  { x: -119.278, y: 92.119, z: -7.98742 },
+		  { x: -119.169, y: 92.1265, z: -8.27037 },
+		  { x: -119.011, y: 92.1651, z: -8.50766 },
+		  { x: -118.786, y: 92.1668, z: -8.63142 },
+		  { x: -118.417, y: 92.1463, z: -8.77355 },
+		  { x: -118.417, y: 92.1463, z: -8.77355 },
+		  { x: -117.96, y: 92.0981, z: -8.88911 },
+		  { x: -117.666, y: 92.116, z: -8.94738 },
+		  { x: -117.443, y: 92.1307, z: -8.99128 },
+		  { x: -117.443, y: 92.1307, z: -8.99128 },
+		  { x: -117.326, y: 92.0872, z: -9.04723 },
+		  { x: -117.326, y: 92.0872, z: -9.04723 },
+		  { x: -117.326, y: 92.0872, z: -9.04723 },
+		  { x: -116.81, y: 92.037, z: -8.87799 },
+		  { x: -116.632, y: 92.0122, z: -8.81713 },
+		  { x: -116.441, y: 91.9816, z: -8.80541 },
+		  { x: -116.303, y: 91.923, z: -8.80048 },
+		  { x: -116.303, y: 91.923, z: -8.80048 },
+		  { x: -116.214, y: 91.8722, z: -8.91166 },
+		  { x: -116.14, y: 91.7973, z: -8.97983 },
+		  { x: -115.951, y: 91.8261, z: -8.97455 },
+		  { x: -115.951, y: 91.8261, z: -8.97455 },
+		  { x: -115.849, y: 91.759, z: -8.98872 },
+		  { x: -115.772, y: 91.6797, z: -9.00002 },
+		  { x: -115.772, y: 91.6797, z: -9.00002 },
+		  { x: -115.746, y: 91.6191, z: -9.06445 },
+		  { x: -115.666, y: 91.573, z: -9.08385 },
+		  { x: -115.6, y: 91.557, z: -9.10852 },
+		  { x: -115.6, y: 91.557, z: -9.10852 },
+		  { x: -115.541, y: 91.5504, z: -9.1198 },
+		  { x: -115.49, y: 91.5634, z: -9.12343 },
+		  { x: -115.441, y: 91.5736, z: -9.11842 },
+		  { x: -115.441, y: 91.5736, z: -9.11842 },
+		  { x: -115.392, y: 91.5592, z: -9.09283 },
+		  { x: -115.273, y: 91.5855, z: -9.04546 },
+		  { x: -115.144, y: 91.6246, z: -9.00913 },
+		  { x: -115.144, y: 91.6246, z: -9.00913 },
+		  { x: -115.019, y: 91.6645, z: -9.0044 },
+		  { x: -114.942, y: 91.6916, z: -9.05621 },
+		  { x: -114.942, y: 91.6916, z: -9.05621 },
+		  { x: -114.771, y: 91.7127, z: -9.0235 },
+		  { x: -114.577, y: 91.7722, z: -9.04261 },
+		  { x: -114.415, y: 91.8432, z: -9.08924 },
+		  { x: -114.415, y: 91.8432, z: -9.08924 },
+		  { x: -114.111, y: 91.9035, z: -9.07359 },
+		  { x: -113.852, y: 91.9552, z: -9.08701 },
+		  { x: -113.852, y: 91.9552, z: -9.08701 },
+		  { x: -113.622, y: 91.9872, z: -9.10293 },
+		  { x: -113.407, y: 92.021, z: -9.1335 },
+		  { x: -113.407, y: 92.021, z: -9.1335 },
+		  { x: -113.207, y: 92.0501, z: -9.17336 },
+		  { x: -113.007, y: 92.0321, z: -9.31928 },
+		  { x: -112.843, y: 92.0403, z: -9.41943 },
+		  { x: -112.843, y: 92.0403, z: -9.41943 },
+		  { x: -112.699, y: 92.057, z: -9.50873 },
+		  { x: -112.604, y: 92.0618, z: -9.579 } ]
+    },
+    {
+	"name": "3",
+	"points":[ { x: -94.5389, y: 96.2864, z: -21.7056 },
+		   { x: -94.4825, y: 96.2919, z: -21.765 },
+		   { x: -94.4825, y: 96.2919, z: -21.765 },
+		   { x: -94.4378, y: 96.2998, z: -21.8334 },
+		   { x: -94.4261, y: 96.3231, z: -21.8963 },
+		   { x: -94.4261, y: 96.3231, z: -21.8963 },
+		   { x: -94.4302, y: 96.341, z: -21.9373 },
+		   { x: -94.4344, y: 96.356, z: -21.9716 },
+		   { x: -94.4344, y: 96.356, z: -21.9716 },
+		   { x: -94.4085, y: 96.4292, z: -22.0027 },
+		   { x: -94.3913, y: 96.4966, z: -22.0275 },
+		   { x: -94.3843, y: 96.6059, z: -22.083 },
+		   { x: -94.3843, y: 96.6059, z: -22.083 },
+		   { x: -94.3774, y: 96.6981, z: -22.1295 },
+		   { x: -94.3659, y: 96.8214, z: -22.1817 },
+		   { x: -94.3659, y: 96.8214, z: -22.1817 },
+		   { x: -94.309, y: 96.9417, z: -22.2088 },
+		   { x: -94.1432, y: 97.1127, z: -22.2604 },
+		   { x: -94.1432, y: 97.1127, z: -22.2604 },
+		   { x: -93.6994, y: 97.2522, z: -22.333 },
+		   { x: -92.6194, y: 97.3351, z: -21.9599 },
+		   { x: -92.6194, y: 97.3351, z: -21.9599 },
+		   { x: -91.3409, y: 97.3844, z: -22.0092 },
+		   { x: -89.4243, y: 97.3681, z: -22.0465 },
+		   { x: -87.4462, y: 97.2713, z: -22.3692 },
+		   { x: -87.4462, y: 97.2713, z: -22.3692 },
+		   { x: -85.2203, y: 97.2307, z: -22.5738 },
+		   { x: -82.8597, y: 96.81, z: -22.7656 },
+		   { x: -80.5924, y: 96.5421, z: -22.8023 },
+		   { x: -78.41, y: 96.3816, z: -22.6417 },
+		   { x: -76.275, y: 96.2026, z: -22.8586 },
+		   { x: -74.2126, y: 96.133, z: -23.1136 },
+		   { x: -72.1117, y: 95.9651, z: -23.3902 },
+		   { x: -69.6503, y: 95.8335, z: -23.1851 },
+		   { x: -67.1925, y: 95.6276, z: -23.3234 },
+		   { x: -64.9972, y: 95.181, z: -23.7134 },
+		   { x: -62.8731, y: 94.8021, z: -23.9185 },
+		   { x: -58.3214, y: 94.3626, z: -24.7697 },
+		   { x: -55.8102, y: 93.9659, z: -25.3374 },
+		   { x: -53.3856, y: 93.722, z: -25.7494 },
+		   { x: -50.0672, y: 92.9541, z: -25.6684 },
+		   { x: -47.1506, y: 92.8593, z: -25.483 },
+		   { x: -44.5269, y: 92.9825, z: -25.2428 },
+		   { x: -38.8187, y: 93.4634, z: -24.7377 },
+		   { x: -36.1402, y: 93.7555, z: -24.5482 },
+		   { x: -33.5823, y: 94.22, z: -24.2733 },
+		   { x: -28.866, y: 94.753, z: -24.3855 },
+		   { x: -26.8071, y: 94.8538, z: -24.4249 },
+		   { x: -24.6069, y: 94.9991, z: -24.6505 },
+		   { x: -19.5052, y: 94.4498, z: -26.2843 },
+		   { x: -16.9104, y: 94.2752, z: -26.8564 },
+		   { x: -14.2318, y: 93.8184, z: -27.6865 },
+		   { x: -8.36906, y: 93.1833, z: -28.6465 },
+		   { x: -5.2912, y: 92.8957, z: -28.6487 },
+		   { x: -2.65485, y: 92.6185, z: -28.6111 },
+		   { x: 1.89016, y: 92.7198, z: -28.835 },
+		   { x: 4.09416, y: 92.8061, z: -29.0167 },
+		   { x: 8.4368, y: 93.1435, z: -29.5212 },
+		   { x: 10.7121, y: 93.2676, z: -29.6263 },
+		   { x: 13.0045, y: 93.7418, z: -29.7034 },
+		   { x: 17.3212, y: 94.2266, z: -29.8048 },
+		   { x: 19.3959, y: 94.5358, z: -29.85 },
+		   { x: 21.0388, y: 94.5011, z: -30.0054 },
+		   { x: 24.8039, y: 94.9609, z: -30.1919 },
+		   { x: 24.8039, y: 94.9609, z: -30.1919 },
+		   { x: 24.8039, y: 94.9609, z: -30.1919 },
+		   { x: 32.4819, y: 96.1604, z: -29.6411 },
+		   { x: 34.2721, y: 96.3758, z: -29.5174 },
+		   { x: 35.7181, y: 96.4294, z: -29.5321 },
+		   { x: 39.0192, y: 96.742, z: -29.5619 },
+		   { x: 40.9394, y: 97.1704, z: -29.5829 },
+		   { x: 42.3351, y: 97.3733, z: -29.5987 },
+		   { x: 45.1993, y: 97.649, z: -29.5244 },
+		   { x: 46.8054, y: 97.6, z: -29.5162 },
+		   { x: 48.3071, y: 97.5551, z: -29.5881 },
+		   { x: 51.6788, y: 97.6244, z: -30.0651 },
+		   { x: 53.7359, y: 98.0925, z: -31.3589 },
+		   { x: 55.7253, y: 98.4487, z: -32.7273 },
+		   { x: 57.5741, y: 99.1454, z: -32.8514 },
+		   { x: 59.1198, y: 99.632, z: -32.8442 },
+		   { x: 62.0913, y: 99.9091, z: -33.2278 },
+		   { x: 63.9709, y: 99.7368, z: -33.684 },
+		   { x: 65.4639, y: 100.048, z: -33.75 },
+		   { x: 67.2251, y: 100.321, z: -34.0859 },
+		   { x: 70.5906, y: 101.355, z: -34.3674 },
+		   { x: 72.2494, y: 102.014, z: -34.3296 },
+		   { x: 74.3371, y: 102.976, z: -34.2433 },
+		   { x: 75.9927, y: 103.204, z: -33.1144 },
+		   { x: 78.7768, y: 103.304, z: -32.4574 },
+		   { x: 80.0615, y: 103.478, z: -32.3736 },
+		   { x: 81.8424, y: 104.615, z: -31.6948 },
+		   { x: 83.4093, y: 105.195, z: -31.3861 },
+		   { x: 84.4586, y: 105.543, z: -31.1795 },
+		   { x: 88.3765, y: 107.073, z: -30.8971 },
+		   { x: 89.7823, y: 107.442, z: -30.7281 },
+		   { x: 91.1075, y: 107.636, z: -30.3452 },
+		   { x: 91.8404, y: 107.667, z: -30.2037 },
+		   { x: 93.4299, y: 107.611, z: -30.0828 },
+		   { x: 94.8685, y: 107.828, z: -30.0768 },
+		   { x: 95.8741, y: 107.785, z: -30.0886 },
+		   { x: 96.4761, y: 107.69, z: -30.0892 },
+		   { x: 98.4525, y: 107.248, z: -29.9464 },
+		   { x: 104.88, y: 107.79, z: -29.1844 },
+		   { x: 108.108, y: 108.149, z: -28.483 },
+		   { x: 111.28, y: 108.793, z: -28.1886 },
+		   { x: 113.536, y: 109.078, z: -27.9942 },
+		   { x: 116.919, y: 109.353, z: -27.6724 },
+		   { x: 118.409, y: 109.583, z: -27.5025 },
+		   { x: 120.012, y: 109.854, z: -27.2287 },
+		   { x: 121.426, y: 109.997, z: -27.0572 },
+		   { x: 124.278, y: 110.514, z: -27.1279 },
+		   { x: 124.971, y: 110.218, z: -27.1814 },
+		   { x: 125.826, y: 110.021, z: -27.2146 },
+		   { x: 127.942, y: 109.856, z: -27.4419 },
+		   { x: 130.175, y: 109.829, z: -27.6075 },
+		   { x: 132.934, y: 109.904, z: -27.3687 },
+		   { x: 134.104, y: 110.189, z: -27.3005 },
+		   { x: 135.192, y: 110.031, z: -26.9597 },
+		   { x: 136.089, y: 109.76, z: -27.1764 },
+		   { x: 139.674, y: 110.803, z: -27.318 },
+		   { x: 140.836, y: 111.258, z: -27.0369 },
+		   { x: 141.345, y: 111.407, z: -26.922 },
+		   { x: 142.778, y: 111.471, z: -26.414 },
+		   { x: 144.412, y: 109.729, z: -26.1103 },
+		   { x: 143.769, y: 108.725, z: -25.9757 },
+		   { x: 144.048, y: 107.003, z: -25.4348 },
+		   { x: 144.643, y: 106.628, z: -25.1531 },
+		   { x: 145.92, y: 106.565, z: -24.6384 },
+		   { x: 146.471, y: 106.033, z: -24.2063 },
+		   { x: 147.259, y: 106.13, z: -24.2004 },
+		   { x: 147.84, y: 106.116, z: -24.1143 },
+		   { x: 149.859, y: 106.105, z: -24.159 },
+		   { x: 149.082, y: 104.775, z: -24.6606 },
+		   { x: 150.68, y: 105.283, z: -24.5738 },
+		   { x: 151.523, y: 105.72, z: -24.5623 },
+		   { x: 153.581, y: 106.586, z: -24.506 },
+		   { x: 158.603, y: 109.115, z: -24.263 },
+		   { x: 160.144, y: 110.063, z: -24.399 },
+		   { x: 160.534, y: 110.139, z: -24.5174 },
+		   { x: 161.044, y: 110.369, z: -24.5609 },
+		   { x: 160.796, y: 110.324, z: -24.6941 },
+		   { x: 160.505, y: 110.412, z: -24.8229 },
+		   { x: 163.377, y: 112.337, z: -24.8902 },
+		   { x: 164.522, y: 112.638, z: -25.2555 },
+		   { x: 164.918, y: 111.57, z: -25.9367 },
+		   { x: 165.834, y: 111.167, z: -26.3124 },
+		   { x: 166.988, y: 110.951, z: -26.4875 },
+		   { x: 168.735, y: 110.895, z: -26.6813 },
+		   { x: 169.528, y: 110.923, z: -26.6907 },
+		   { x: 171.469, y: 108.737, z: -26.478 },
+		   { x: 171.254, y: 104.615, z: -26.1429 },
+		   { x: 171.149, y: 100.672, z: -25.5525 },
+		   { x: 172.905, y: 98.3252, z: -25.0961 },
+		   { x: 176.205, y: 95.208, z: -21.5049 },
+		   { x: 177.744, y: 94.6452, z: -20.2544 },
+		   { x: 180.877, y: 95.6037, z: -19.0599 },
+		   { x: 181.343, y: 93.1645, z: -17.2105 },
+		   { x: 182.667, y: 92.5424, z: -16.0073 },
+		   { x: 183.385, y: 91.3586, z: -13.4781 },
+		   { x: 183.71, y: 90.2627, z: -12.9359 },
+		   { x: 180.52, y: 85.54, z: -13.8458 },
+		   { x: 178.175, y: 83.3667, z: -14.0851 },
+		   { x: 169.135, y: 83.4013, z: -9.49324 },
+		   { x: 169.002, y: 83.2877, z: -5.24128 },
+		   { x: 169.002, y: 83.2877, z: -5.24128 },
+		   { x: 168.713, y: 81.7423, z: -4.54178 },
+		   { x: 169.185, y: 83.7326, z: -4.62498 },
+		   { x: 160.771, y: 79.2834, z: -0.00679269 },
+		   { x: 169.737, y: 92.7513, z: -0.61072 },
+		   { x: 174.335, y: 98.2324, z: -1.63701 },
+		   { x: 176.288, y: 104.25, z: -4.98615 },
+		   { x: 179.502, y: 108.455, z: -6.69978 },
+		   { x: 179.77, y: 109.194, z: -7.95535 },
+		   { x: 180.755, y: 107.207, z: -9.09203 },
+		   { x: 179.834, y: 94.6027, z: -9.67626 },
+		   { x: 179.741, y: 90.3906, z: -11.7049 },
+		   { x: 183.304, y: 92.4794, z: -15.7062 },
+		   { x: 192.881, y: 84.8614, z: -13.8744 } ]
+    },
+    {
+	"name": "3'",
+	"points":[ { x: 142.396, y: 123.484, z: -26.1569 },
+		   { x: 142.253, y: 123.44, z: -25.9945 },
+		   { x: 142.152, y: 123.383, z: -25.8735 },
+		   { x: 142.117, y: 123.185, z: -25.9467 },
+		   { x: 141.719, y: 122.973, z: -25.6041 },
+		   { x: 142.984, y: 118.804, z: -23.7519 },
+		   { x: 144.777, y: 117.363, z: -22.0375 },
+		   { x: 145.442, y: 117.171, z: -21.4683 },
+		   { x: 145.642, y: 117.22, z: -21.3021 },
+		   { x: 145.735, y: 117.394, z: -21.1826 },
+		   { x: 145.733, y: 117.492, z: -21.1526 },
+		   { x: 145.71, y: 117.583, z: -21.1327 },
+		   { x: 146.434, y: 117.614, z: -21.5346 },
+		   { x: 146.965, y: 117.697, z: -21.8723 },
+		   { x: 147.222, y: 117.673, z: -22.0567 },
+		   { x: 148.955, y: 117.701, z: -21.7956 },
+		   { x: 149.345, y: 118.006, z: -21.7805 },
+		   { x: 149.456, y: 118.217, z: -21.7998 },
+		   { x: 149.43, y: 118.375, z: -21.8384 },
+		   { x: 148.569, y: 117.835, z: -21.9988 },
+		   { x: 148.09, y: 117.666, z: -22.0197 },
+		   { x: 148.09, y: 117.666, z: -22.0197 },
+		   { x: 148.09, y: 117.666, z: -22.0197 },
+		   { x: 147.419, y: 117.811, z: -22.1097 },
+		   { x: 145.749, y: 117.921, z: -23.3244 },
+		   { x: 144.255, y: 117.509, z: -24.5066 },
+		   { x: 142.632, y: 116.934, z: -26.0037 },
+		   { x: 140.171, y: 116.718, z: -26.9335 },
+		   { x: 139.06, y: 116.737, z: -26.972 },
+		   { x: 137.205, y: 117.193, z: -27.026 },
+		   { x: 133.686, y: 117.572, z: -27.7819 },
+		   { x: 131.987, y: 117.862, z: -28.3913 },
+		   { x: 129.734, y: 118.191, z: -28.846 },
+		   { x: 128.743, y: 118.289, z: -29.064 },
+		   { x: 126.355, y: 117.843, z: -29.5705 },
+		   { x: 123.05, y: 117.883, z: -30.7069 },
+		   { x: 121.519, y: 118.149, z: -31.4018 },
+		   { x: 118.672, y: 118.424, z: -32.4624 },
+		   { x: 116.615, y: 118.633, z: -33.0199 },
+		   { x: 108.559, y: 117.443, z: -34.677 },
+		   { x: 106.258, y: 117.512, z: -34.8818 },
+		   { x: 102.039, y: 117.856, z: -35.8898 },
+		   { x: 99.8427, y: 118.468, z: -36.5377 },
+		   { x: 97.4063, y: 118.284, z: -37.689 },
+		   { x: 95.7659, y: 118.033, z: -38.4906 },
+		   { x: 94.6973, y: 118.012, z: -38.9609 },
+		   { x: 91.0773, y: 117.476, z: -40.0195 },
+		   { x: 88.1043, y: 117.833, z: -40.7059 },
+		   { x: 86.317, y: 118.007, z: -41.3943 },
+		   { x: 83.1316, y: 118.2, z: -41.967 },
+		   { x: 81.5503, y: 118.619, z: -42.2904 },
+		   { x: 75.9806, y: 119.746, z: -42.8692 },
+		   { x: 74.6958, y: 121.324, z: -41.3589 },
+		   { x: 72.8541, y: 121.652, z: -40.9734 },
+		   { x: 70.8136, y: 121.238, z: -41.1521 },
+		   { x: 67.4028, y: 120.44, z: -42.2172 },
+		   { x: 64.7341, y: 120.679, z: -42.9756 },
+		   { x: 61.6758, y: 119.943, z: -44.0437 },
+		   { x: 58.9334, y: 119.478, z: -44.5304 },
+		   { x: 56.7089, y: 119.577, z: -44.2105 },
+		   { x: 55.7795, y: 119.523, z: -44.0976 },
+		   { x: 53.7879, y: 119.987, z: -44.1131 },
+		   { x: 51.5073, y: 120.016, z: -44.1914 },
+		   { x: 48.9183, y: 120.243, z: -44.2787 },
+		   { x: 46.3119, y: 120.536, z: -44.2565 },
+		   { x: 43.6722, y: 121.101, z: -44.441 },
+		   { x: 42.6117, y: 120.99, z: -44.7121 },
+		   { x: 40.4228, y: 120.941, z: -44.8566 },
+		   { x: 37.5684, y: 121.33, z: -45.1875 },
+		   { x: 35.6603, y: 121.287, z: -45.3078 },
+		   { x: 33.7454, y: 120.462, z: -45.7557 },
+		   { x: 31.8176, y: 120.265, z: -45.9942 },
+		   { x: 29.6991, y: 120.318, z: -46.1584 },
+		   { x: 25.6054, y: 120.065, z: -46.3566 },
+		   { x: 23.2677, y: 119.334, z: -46.8413 },
+		   { x: 21.2221, y: 118.837, z: -47.1037 },
+		   { x: 19.053, y: 118.762, z: -47.6938 },
+		   { x: 17.0195, y: 119.188, z: -47.7852 },
+		   { x: 13.6576, y: 119.874, z: -47.8476 },
+		   { x: 11.9459, y: 120.2, z: -47.8586 },
+		   { x: 10.2835, y: 120.614, z: -47.7924 },
+		   { x: 8.59285, y: 120.791, z: -47.9169 },
+		   { x: 6.9659, y: 120.914, z: -48.0327 },
+		   { x: 3.44374, y: 121.351, z: -48.1181 },
+		   { x: 1.70358, y: 121.335, z: -48.1827 },
+		   { x: 0.096869, y: 121.293, z: -48.3812 },
+		   { x: -1.43025, y: 120.857, z: -48.5832 },
+		   { x: -2.86551, y: 120.79, z: -48.6511 },
+		   { x: -6.11374, y: 121.047, z: -48.8706 },
+		   { x: -7.70054, y: 121.11, z: -48.9606 },
+		   { x: -9.13957, y: 121.297, z: -48.9957 },
+		   { x: -10.9546, y: 121.402, z: -49.1831 },
+		   { x: -12.7058, y: 121.176, z: -49.616 },
+		   { x: -16.1851, y: 120.872, z: -50.1919 },
+		   { x: -17.389, y: 120.716, z: -50.3718 },
+		   { x: -19.1899, y: 120.838, z: -50.5206 },
+		   { x: -21.3125, y: 120.668, z: -50.8709 },
+		   { x: -23.5865, y: 120.285, z: -51.6593 },
+		   { x: -25.5978, y: 120.186, z: -51.8383 },
+		   { x: -27.3365, y: 120.183, z: -51.942 },
+		   { x: -31.1648, y: 120.26, z: -51.9442 },
+		   { x: -33.1385, y: 120.362, z: -51.8396 },
+		   { x: -34.6964, y: 120.328, z: -51.752 },
+		   { x: -36.5735, y: 120.223, z: -51.6808 },
+		   { x: -38.7245, y: 119.968, z: -51.7551 },
+		   { x: -41.023, y: 119.627, z: -51.9392 },
+		   { x: -43.0858, y: 119.393, z: -51.9322 },
+		   { x: -47.4527, y: 118.68, z: -51.8177 },
+		   { x: -49.497, y: 118.228, z: -51.9626 },
+		   { x: -51.5901, y: 117.532, z: -52.1832 },
+		   { x: -53.383, y: 116.942, z: -52.2658 },
+		   { x: -55.3004, y: 116.743, z: -51.9704 },
+		   { x: -58.3738, y: 115.186, z: -51.3823 },
+		   { x: -58.3738, y: 115.186, z: -51.3823 },
+		   { x: -58.3738, y: 115.186, z: -51.3823 },
+		   { x: -63.8352, y: 113.516, z: -52.0024 },
+		   { x: -65.4893, y: 113.226, z: -51.9005 },
+		   { x: -67.1523, y: 112.989, z: -51.7116 },
+		   { x: -70.1532, y: 112.319, z: -51.9089 },
+		   { x: -71.656, y: 111.956, z: -51.9388 },
+		   { x: -73.3439, y: 111.949, z: -51.7934 },
+		   { x: -75.354, y: 112.218, z: -51.7048 },
+		   { x: -77.1095, y: 112.224, z: -51.9108 },
+		   { x: -78.6853, y: 112.152, z: -52.4537 },
+		   { x: -81.4739, y: 111.327, z: -53.7248 },
+		   { x: -82.9931, y: 110.899, z: -54.2942 },
+		   { x: -84.8547, y: 110.963, z: -54.8477 },
+		   { x: -86.7139, y: 111.3, z: -55.1387 },
+		   { x: -88.5111, y: 111.568, z: -55.2089 },
+		   { x: -90.0367, y: 111.678, z: -55.2005 },
+		   { x: -91.8317, y: 112.414, z: -55.3014 },
+		   { x: -94.9502, y: 112.646, z: -55.5915 },
+		   { x: -96.7503, y: 113.117, z: -56.0734 },
+		   { x: -98.495, y: 113.69, z: -56.3594 },
+		   { x: -100.094, y: 113.901, z: -56.5752 },
+		   { x: -101.347, y: 113.922, z: -56.4153 },
+		   { x: -102.549, y: 113.884, z: -56.3708 },
+		   { x: -103.933, y: 114.076, z: -56.5903 },
+		   { x: -105.457, y: 114.243, z: -56.6131 },
+		   { x: -107.783, y: 114.834, z: -56.2464 },
+		   { x: -108.774, y: 114.865, z: -56.2733 },
+		   { x: -110.004, y: 114.948, z: -56.1302 },
+		   { x: -111.029, y: 115.125, z: -55.9921 },
+		   { x: -112.74, y: 115.66, z: -56.0644 },
+		   { x: -114.918, y: 115.857, z: -56.1526 },
+		   { x: -116.053, y: 116.33, z: -55.5746 },
+		   { x: -117.022, y: 116.766, z: -55.0593 },
+		   { x: -117.94, y: 116.868, z: -54.7558 },
+		   { x: -119.372, y: 117.025, z: -54.6162 },
+		   { x: -122.489, y: 116.992, z: -54.8598 },
+		   { x: -123.716, y: 117.018, z: -54.4573 },
+		   { x: -124.54, y: 116.93, z: -54.0953 },
+		   { x: -125.651, y: 116.377, z: -54.5365 },
+		   { x: -127.287, y: 116.129, z: -54.6593 },
+		   { x: -128.836, y: 116.461, z: -54.785 },
+		   { x: -131.436, y: 116.762, z: -54.7743 },
+		   { x: -133.021, y: 116.437, z: -54.8104 },
+		   { x: -135.038, y: 116.943, z: -55.0879 },
+		   { x: -136.001, y: 117.286, z: -55.2057 },
+		   { x: -136.935, y: 116.899, z: -55.2634 },
+		   { x: -136.935, y: 116.899, z: -55.2634 },
+		   { x: -139, y: 117.221, z: -55.0745 },
+		   { x: -139.997, y: 116.79, z: -55.1721 },
+		   { x: -141.359, y: 116.592, z: -55.3269 },
+		   { x: -142.638, y: 116.379, z: -55.2605 },
+		   { x: -143.92, y: 116.289, z: -54.9002 },
+		   { x: -145.167, y: 116.421, z: -55.0335 },
+		   { x: -146.322, y: 116.425, z: -55.0492 },
+		   { x: -147.528, y: 116.483, z: -55.0821 },
+		   { x: -148.595, y: 116.086, z: -54.9885 },
+		   { x: -149.762, y: 115.718, z: -55.2469 },
+		   { x: -150.922, y: 115.382, z: -55.1418 },
+		   { x: -152.447, y: 115.09, z: -55.0689 },
+		   { x: -154.098, y: 115.506, z: -54.9467 },
+		   { x: -155.712, y: 115.09, z: -54.9842 },
+		   { x: -157.036, y: 114.365, z: -55.6367 },
+		   { x: -158.197, y: 114.851, z: -55.9191 },
+		   { x: -159.627, y: 115.052, z: -55.677 },
+		   { x: -161.527, y: 115.872, z: -55.2258 },
+		   { x: -160.989, y: 114.625, z: -55.2224 },
+		   { x: -161.556, y: 114.327, z: -55.0539 },
+		   { x: -164.137, y: 115.006, z: -55.5215 },
+		   { x: -165.243, y: 113.855, z: -56.9594 },
+		   { x: -166.074, y: 113.242, z: -58.4765 },
+		   { x: -167.036, y: 114.645, z: -58.6842 },
+		   { x: -167.143, y: 113.123, z: -59.8071 },
+		   { x: -167.382, y: 112.857, z: -59.8301 },
+		   { x: -167.611, y: 112.568, z: -59.8201 },
+		   { x: -168.029, y: 111.96, z: -60.1749 },
+		   { x: -168.207, y: 111.767, z: -60.2808 },
+		   { x: -168.373, y: 111.584, z: -60.3512 },
+		   { x: -170.238, y: 112.215, z: -59.6768 },
+		   { x: -172.78, y: 113.249, z: -59.1739 },
+		   { x: -172.406, y: 113.079, z: -59.603 },
+		   { x: -172.627, y: 112.952, z: -59.5706 },
+		   { x: -173.517, y: 112.717, z: -59.702 },
+		   { x: -177.16, y: 115.245, z: -59.393 },
+		   { x: -176.418, y: 113.433, z: -59.6578 },
+		   { x: -177.006, y: 112.982, z: -60.0262 },
+		   { x: -178.644, y: 114.336, z: -59.6688 },
+		   { x: -179.632, y: 114.388, z: -59.5438 },
+		   { x: -178.537, y: 111.231, z: -61.0791 },
+		   { x: -177.873, y: 109.736, z: -61.4871 },
+		   { x: -178.308, y: 108.363, z: -61.309 },
+		   { x: -180.701, y: 109.589, z: -61.0346 },
+		   { x: -182.018, y: 110.732, z: -60.8746 },
+		   { x: -182.231, y: 110.671, z: -60.8304 },
+		   { x: -177.168, y: 105.828, z: -63.8588 },
+		   { x: -179.462, y: 107.754, z: -62.6031 },
+		   { x: -179.833, y: 107.813, z: -62.3632 },
+		   { x: -184.447, y: 109.448, z: -61.5062 },
+		   { x: -184.813, y: 109.206, z: -61.6884 },
+		   { x: -184.451, y: 108.935, z: -62.133 },
+		   { x: -184.494, y: 108.805, z: -62.1353 },
+		   { x: -183.943, y: 108.451, z: -62.3885 },
+		   { x: -184.069, y: 108.589, z: -62.0395 },
+		   { x: -183.991, y: 108.038, z: -61.9051 },
+		   { x: -183.775, y: 108.378, z: -61.3802 },
+		   { x: -184.134, y: 114.225, z: -59.3388 },
+		   { x: -188.95, y: 121.952, z: -57.9683 },
+		   { x: -191.5, y: 124.873, z: -57.0942 },
+		   { x: -191.5, y: 124.873, z: -57.0942 },
+		   { x: -191.803, y: 125.162, z: -56.9352 },
+		   { x: -191.852, y: 125.249, z: -56.9156 },
+		   { x: -192.698, y: 125.868, z: -56.4636 },
+		   { x: -192.716, y: 125.877, z: -56.452 },
+		   { x: -192.545, y: 125.971, z: -56.4079 },
+		   { x: -186.277, y: 123.238, z: -57.828 },
+		   { x: -187.731, y: 123.902, z: -56.9437 },
+		   { x: -187.731, y: 123.902, z: -56.9437 },
+		   { x: -188.76, y: 123.66, z: -56.6024 },
+		   { x: -188.087, y: 123.135, z: -56.8705 },
+		   { x: -187.697, y: 122.877, z: -56.8229 },
+		   { x: -186.991, y: 122.171, z: -56.5196 },
+		   { x: -186.793, y: 122.403, z: -56.4389 },
+		   { x: -187.759, y: 123.27, z: -56.1603 },
+		   { x: -189.276, y: 123.676, z: -59.0797 },
+		   { x: -189.608, y: 122.817, z: -60.3209 },
+		   { x: -186.865, y: 121.483, z: -59.2453 },
+		   { x: -187.294, y: 120.894, z: -60.0189 },
+		   { x: -187.738, y: 121.616, z: -58.5949 },
+		   { x: -187.252, y: 122.351, z: -58.8166 },
+		   { x: -187.012, y: 122.642, z: -58.7994 },
+		   { x: -186.896, y: 122.759, z: -58.8498 },
+		   { x: -186.561, y: 122.936, z: -58.9623 },
+		   { x: -186.503, y: 123.411, z: -59.0528 },
+		   { x: -186.466, y: 124.01, z: -59.1895 },
+		   { x: -186.443, y: 124.021, z: -59.2021 },
+		   { x: -186.418, y: 124.023, z: -59.2196 },
+		   { x: -185.731, y: 123.373, z: -60.1367 },
+		   { x: -184.877, y: 123.1, z: -60.9638 },
+		   { x: -184.228, y: 123.181, z: -61.259 },
+		   { x: -182.994, y: 123.276, z: -61.4644 },
+		   { x: -182.706, y: 123.304, z: -61.5947 },
+		   { x: -182.833, y: 123.389, z: -61.4734 },
+		   { x: -182.859, y: 123.539, z: -61.3437 },
+		   { x: -178.195, y: 122.801, z: -63.4249 },
+		   { x: -174.153, y: 122.98, z: -64.0014 },
+		   { x: -172.14, y: 123.16, z: -64.0967 },
+		   { x: -171.05, y: 123.029, z: -63.9523 },
+		   { x: -170.385, y: 123.186, z: -63.7547 },
+		   { x: -168.306, y: 122.137, z: -63.9 },
+		   { x: -165.469, y: 119.527, z: -63.3551 },
+		   { x: -163.262, y: 118.334, z: -61.8665 },
+		   { x: -161.162, y: 117.095, z: -60.9613 },
+		   { x: -160.537, y: 115.844, z: -60.4372 },
+		   { x: -160.624, y: 113.94, z: -59.84 },
+		   { x: -160.276, y: 112.567, z: -59.0653 } ]
+    },
+    {    "name": "4",
+	 "points": [ { x: -56.2428, y: 54.5958, z: -35.241 },
+		     { x: -56.1163, y: 54.9552, z: -35.1402 },
+		     { x: -55.8536, y: 55.3817, z: -34.9047 },
+		     { x: -55.4212, y: 55.807, z: -34.4179 },
+		     { x: -54.8481, y: 56.1838, z: -33.8291 },
+		     { x: -54.4299, y: 56.5155, z: -33.5058 },
+		     { x: -54.1124, y: 56.8474, z: -33.3269 },
+		     { x: -53.8802, y: 57.1578, z: -33.1734 },
+		     { x: -53.7908, y: 57.3154, z: -33.1183 },
+		     { x: -53.6353, y: 57.618, z: -33.0451 },
+		     { x: -53.5986, y: 57.7231, z: -33.0204 },
+		     { x: -53.501, y: 58.0639, z: -32.933 },
+		     { x: -53.4298, y: 58.353, z: -32.851 },
+		     { x: -53.395, y: 58.5653, z: -32.7715 },
+		     { x: -53.358, y: 58.7376, z: -32.7078 },
+		     { x: -53.3666, y: 58.8997, z: -32.6488 },
+		     { x: -53.3733, y: 58.9784, z: -32.5983 },
+		     { x: -53.3761, y: 59.0416, z: -32.5592 },
+		     { x: -53.3449, y: 59.0751, z: -32.5376 },
+		     { x: -53.2037, y: 59.0609, z: -32.5543 },
+		     { x: -52.8507, y: 58.9595, z: -32.6305 },
+		     { x: -52.1228, y: 58.7028, z: -32.7873 },
+		     { x: -51.2374, y: 58.4327, z: -33.0074 },
+		     { x: -49.6372, y: 58.0663, z: -33.3921 },
+		     { x: -48.5529, y: 57.8618, z: -33.6 },
+		     { x: -46.437, y: 57.5283, z: -34.1929 },
+		     { x: -43.6622, y: 57.3212, z: -34.9384 },
+		     { x: -40.8274, y: 57.1235, z: -35.5831 },
+		     { x: -37.8729, y: 57.1529, z: -36.4202 },
+		     { x: -35.0863, y: 57.7118, z: -36.9771 },
+		     { x: -32.3399, y: 58.2919, z: -37.5388 },
+		     { x: -29.2387, y: 58.3218, z: -38.5172 },
+		     { x: -26.5467, y: 58.4099, z: -39.3908 },
+		     { x: -23.8138, y: 58.7436, z: -39.9908 },
+		     { x: -20.9147, y: 59.321, z: -40.523 },
+		     { x: -17.7608, y: 59.8479, z: -41.1103 },
+		     { x: -14.5203, y: 60.2929, z: -41.7777 },
+		     { x: -12.9468, y: 60.5388, z: -42.0722 },
+		     { x: -9.63206, y: 61.0349, z: -42.6114 },
+		     { x: -6.12759, y: 61.5062, z: -43.0812 },
+		     { x: -2.65483, y: 62.0143, z: -43.4216 },
+		     { x: 0.495749, y: 62.6837, z: -43.6635 },
+		     { x: 3.98219, y: 63.4847, z: -43.9275 },
+		     { x: 7.50245, y: 64.5501, z: -44.2097 },
+		     { x: 10.61, y: 65.2752, z: -44.2317 },
+		     { x: 13.745, y: 66.0213, z: -44.1055 },
+		     { x: 17.2102, y: 66.4333, z: -43.5209 },
+		     { x: 20.3645, y: 66.6651, z: -43.1251 },
+		     { x: 22.9324, y: 67.1844, z: -43.0472 },
+		     { x: 25.4553, y: 67.8224, z: -43.2916 },
+		     { x: 28.1582, y: 68.4921, z: -43.8311 },
+		     { x: 31.0467, y: 69.7316, z: -44.566 },
+		     { x: 32.4301, y: 70.0695, z: -44.7736 },
+		     { x: 35.152, y: 70.7934, z: -44.6826 },
+		     { x: 37.6322, y: 70.6843, z: -44.8404 },
+		     { x: 40.2758, y: 71.1208, z: -44.8536 },
+		     { x: 42.7442, y: 71.1742, z: -44.9288 },
+		     { x: 45.203, y: 71.1419, z: -45.1764 },
+		     { x: 47.7412, y: 70.7695, z: -45.3167 },
+		     { x: 50.2923, y: 70.4719, z: -45.363 },
+		     { x: 52.353, y: 69.9718, z: -45.273 },
+		     { x: 54.213, y: 69.3884, z: -45.2888 },
+		     { x: 55.4457, y: 68.838, z: -45.3711 },
+		     { x: 56.6413, y: 68.3884, z: -45.5191 },
+		     { x: 57.5535, y: 68.0575, z: -45.4972 },
+		     { x: 58.1397, y: 67.8762, z: -45.4547 },
+		     { x: 58.3242, y: 67.8321, z: -45.4328 },
+		     { x: 58.561, y: 67.7432, z: -45.3927 },
+		     { x: 58.6298, y: 67.6916, z: -45.3477 },
+		     { x: 58.5806, y: 67.8281, z: -45.2343 },
+		     { x: 58.3432, y: 68.1661, z: -45.124 },
+		     { x: 57.5005, y: 69.0629, z: -44.9844 },
+		     { x: 56.283, y: 70.405, z: -44.8394 },
+		     { x: 54.3236, y: 72.3184, z: -44.2097 },
+		     { x: 53.4669, y: 73.4734, z: -44.1066 },
+		     { x: 51.8235, y: 76.9302, z: -44.7712 },
+		     { x: 51.4931, y: 79.1714, z: -45.3001 },
+		     { x: 51.4001, y: 83.4525, z: -45.0964 },
+		     { x: 52.2566, y: 88.8049, z: -43.8872 },
+		     { x: 53.6345, y: 93.4011, z: -42.3705 },
+		     { x: 54.5135, y: 95.9511, z: -41.6124 },
+		     { x: 56.471, y: 100.193, z: -40.7125 },
+		     { x: 59.2889, y: 104.65, z: -40.8478 },
+		     { x: 61.6403, y: 108.262, z: -39.9885 },
+		     { x: 64.013, y: 111.294, z: -38.4609 },
+		     { x: 66.8769, y: 112.962, z: -37.5078 },
+		     { x: 69.7738, y: 114.736, z: -34.8788 },
+		     { x: 73.2709, y: 115.296, z: -31.9743 },
+		     { x: 76.9312, y: 115.06, z: -34.6873 },
+		     { x: 80.036, y: 114.078, z: -35.3617 },
+		     { x: 83.2585, y: 112.671, z: -35.8385 },
+		     { x: 86.371, y: 110.583, z: -34.5316 },
+		     { x: 89.0404, y: 107.61, z: -34.3448 },
+		     { x: 90.1434, y: 105.889, z: -34.2334 },
+		     { x: 92.5464, y: 101.543, z: -34.5437 },
+		     { x: 94.7229, y: 97.1389, z: -34.3533 },
+		     { x: 96.3051, y: 92.4502, z: -33.8043 },
+		     { x: 97.7033, y: 88.3825, z: -32.6105 },
+		     { x: 98.3707, y: 84.2992, z: -31.3393 },
+		     { x: 99.2514, y: 80.1217, z: -31.4432 },
+		     { x: 100.206, y: 75.4819, z: -31.6385 },
+		     { x: 100.386, y: 71.7566, z: -31.5448 },
+		     { x: 100.952, y: 68.0023, z: -31.0943 },
+		     { x: 101.071, y: 65.5032, z: -30.9509 },
+		     { x: 101.152, y: 63.5472, z: -30.4645 },
+		     { x: 101.565, y: 61.6858, z: -30.0374 },
+		     { x: 101.671, y: 60.6924, z: -29.8118 },
+		     { x: 101.691, y: 60.6162, z: -29.8068 },
+		     { x: 101.67, y: 60.6555, z: -29.8592 },
+		     { x: 101.868, y: 61.5284, z: -30.6312 },
+		     { x: 101.614, y: 62.3481, z: -30.6568 },
+		     { x: 100.686, y: 63.6912, z: -31.2614 },
+		     { x: 101.127, y: 65.8775, z: -32.8238 },
+		     { x: 102.06, y: 68.6651, z: -33.2241 },
+		     { x: 102.456, y: 71.7474, z: -33.4168 },
+		     { x: 103.847, y: 76.1779, z: -34.7377 },
+		     { x: 104.966, y: 80.2479, z: -34.9034 },
+		     { x: 106.927, y: 83.3289, z: -34.6301 },
+		     { x: 109.154, y: 87.9372, z: -34.8195 },
+		     { x: 111.407, y: 91.3379, z: -36.9772 },
+		     { x: 113.046, y: 93.4252, z: -38.7163 },
+		     { x: 113.613, y: 94.3587, z: -39.3709 },
+		     { x: 115.963, y: 96.2614, z: -40.5291 },
+		     { x: 120.491, y: 99.4501, z: -40.9164 },
+		     { x: 123.439, y: 101.086, z: -41.1993 },
+		     { x: 126.07, y: 102.091, z: -41.3783 },
+		     { x: 129.575, y: 102.171, z: -40.674 },
+		     { x: 133.052, y: 102.442, z: -40.4554 },
+		     { x: 136.092, y: 102.564, z: -40.5949 },
+		     { x: 133.969, y: 99.426, z: -40.0704 },
+		     { x: 139.378, y: 99.7482, z: -38.8182 },
+		     { x: 144.977, y: 98.4571, z: -37.8139 },
+		     { x: 152.773, y: 98.4338, z: -35.8846 },
+		     { x: 157.172, y: 99.6822, z: -35.902 },
+		     { x: 160.04, y: 99.6525, z: -35.8545 },
+		     { x: 163.307, y: 99.415, z: -37.3715 },
+		     { x: 158.458, y: 91.4968, z: -39.5818 },
+		     { x: 156.508, y: 87.0245, z: -39.5767 },
+		     { x: 157.286, y: 84.9152, z: -39.0986 },
+		     { x: 161.749, y: 85.265, z: -37.5708 },
+		     { x: 164.356, y: 86.1265, z: -36.62 },
+		     { x: 164.329, y: 83.6955, z: -36.7747 },
+		     { x: 162.843, y: 77.8299, z: -35.2144 },
+		     { x: 163.519, y: 74.6497, z: -32.6129 },
+		     { x: 166.295, y: 74.9028, z: -32.3573 } ]
+    },
+    {
+	"name": "4'",
+	"points": [ { x: 78.6581, y: 57.4065, z: -71.3868 },
+		    { x: 72.7614, y: 54.0803, z: -71.2273 },
+		    { x: 69.8536, y: 52.9168, z: -69.4426 },
+		    { x: 70.2136, y: 52.9931, z: -68.4017 },
+		    { x: 70.2879, y: 52.9388, z: -68.0689 },
+		    { x: 69.7529, y: 52.6035, z: -67.9564 },
+		    { x: 70.0592, y: 52.5963, z: -67.9543 },
+		    { x: 70.4866, y: 52.4009, z: -68.0404 },
+		    { x: 70.2563, y: 51.9587, z: -67.8125 },
+		    { x: 69.0847, y: 51.1093, z: -67.9528 },
+		    { x: 69.7731, y: 51.0241, z: -67.8275 },
+		    { x: 71.1272, y: 51.6153, z: -67.4949 },
+		    { x: 71.5492, y: 51.47, z: -67.3486 },
+		    { x: 71.632, y: 51.521, z: -67.3277 },
+		    { x: 71.7935, y: 51.5619, z: -67.2942 },
+		    { x: 71.7918, y: 51.3848, z: -67.0872 },
+		    { x: 71.9381, y: 51.4583, z: -66.9044 },
+		    { x: 72.2089, y: 51.8204, z: -66.7785 },
+		    { x: 72.1876, y: 51.9722, z: -66.7751 },
+		    { x: 71.8172, y: 52.2491, z: -66.9674 },
+		    { x: 72.4325, y: 54.2637, z: -66.7628 },
+		    { x: 72.124, y: 56.0139, z: -66.9361 },
+		    { x: 70.8098, y: 57.2888, z: -67.2803 },
+		    { x: 70.0454, y: 60.1544, z: -66.9571 },
+		    { x: 68.8262, y: 63.8548, z: -66.0576 },
+		    { x: 68.1788, y: 68.9381, z: -65.2753 },
+		    { x: 66.5571, y: 72.8004, z: -65.7953 },
+		    { x: 65.9196, y: 74.7742, z: -66.4795 },
+		    { x: 64.2782, y: 80.6266, z: -69.0002 },
+		    { x: 62.1643, y: 85.0498, z: -71.933 },
+		    { x: 60.1332, y: 89.9743, z: -74.3091 },
+		    { x: 57.6897, y: 95.8502, z: -75.6524 },
+		    { x: 54.9493, y: 101.785, z: -76.7291 },
+		    { x: 51.0489, y: 106.65, z: -77.8274 },
+		    { x: 46.8185, y: 111.338, z: -78.2613 },
+		    { x: 42.1527, y: 115.326, z: -77.9489 },
+		    { x: 36.9653, y: 117.656, z: -77.9048 },
+		    { x: 32.2928, y: 119.813, z: -77.9692 },
+		    { x: 27.7583, y: 121.453, z: -78.2072 },
+		    { x: 23.2678, y: 122.449, z: -77.678 },
+		    { x: 21.2823, y: 122.836, z: -77.3847 },
+		    { x: 17.5458, y: 122.532, z: -76.9765 },
+		    { x: 12.6619, y: 121.258, z: -76.1464 },
+		    { x: 7.85267, y: 118.563, z: -74.8851 },
+		    { x: 3.72838, y: 114.722, z: -74.0684 },
+		    { x: -0.234335, y: 109.317, z: -72.3113 },
+		    { x: -3.10875, y: 103.788, z: -70.4872 },
+		    { x: -5.1967, y: 98.1557, z: -69.8432 },
+		    { x: -6.75534, y: 92.1706, z: -69.7218 },
+		    { x: -7.01135, y: 86.2032, z: -70.2811 },
+		    { x: -6.97761, y: 79.1457, z: -70.6265 },
+		    { x: -6.91772, y: 76.0101, z: -70.1317 },
+		    { x: -7.03826, y: 66.1873, z: -69.2002 },
+		    { x: -5.8499, y: 61.5641, z: -69.5126 },
+		    { x: -5.29623, y: 57.3897, z: -70.0487 },
+		    { x: -5.07211, y: 53.8681, z: -70.6039 },
+		    { x: -5.03206, y: 52.193, z: -70.7392 },
+		    { x: -4.82373, y: 49.3058, z: -70.5532 },
+		    { x: -4.14064, y: 47.7183, z: -70.102 },
+		    { x: -2.54446, y: 48.0597, z: -68.834 },
+		    { x: -1.86521, y: 49.2045, z: -68.1064 },
+		    { x: -1.58636, y: 49.574, z: -67.8559 },
+		    { x: -1.24045, y: 54.5041, z: -66.2441 },
+		    { x: -0.613432, y: 56.6365, z: -66.2765 },
+		    { x: -0.0959971, y: 60.3643, z: -67.3134 },
+		    { x: 0.506553, y: 65.2991, z: -70.3454 },
+		    { x: 0.248154, y: 70.2357, z: -73.6972 },
+		    { x: -0.477026, y: 75.0581, z: -75.8527 },
+		    { x: -0.561923, y: 77.8265, z: -76.7723 },
+		    { x: -1.95148, y: 83.007, z: -77.7341 },
+		    { x: -3.69835, y: 87.9836, z: -79.0017 },
+		    { x: -5.5913, y: 94.1298, z: -80.0067 },
+		    { x: -8.02862, y: 98.9448, z: -80.004 },
+		    { x: -10.5552, y: 102.034, z: -79.9298 },
+		    { x: -12.8235, y: 104.913, z: -79.7194 },
+		    { x: -15.1361, y: 107.479, z: -79.3802 },
+		    { x: -17.6088, y: 109.168, z: -78.9615 },
+		    { x: -20.7089, y: 110.095, z: -78.4244 },
+		    { x: -23.3734, y: 110.062, z: -77.8486 },
+		    { x: -26.9726, y: 110.238, z: -76.5411 },
+		    { x: -30.4016, y: 109.878, z: -74.4631 },
+		    { x: -33.3501, y: 107.893, z: -72.6716 },
+		    { x: -34.885, y: 105.743, z: -72.0833 },
+		    { x: -38.1918, y: 101.901, z: -70.3736 },
+		    { x: -40.9933, y: 95.6841, z: -68.0441 },
+		    { x: -42.8428, y: 90.528, z: -67.2986 },
+		    { x: -45.1945, y: 84.114, z: -64.4847 },
+		    { x: -45.8909, y: 79.3509, z: -62.0973 },
+		    { x: -45.5924, y: 73.9695, z: -62.4723 },
+		    { x: -46.2116, y: 69.42, z: -63.2025 },
+		    { x: -45.9026, y: 65.0674, z: -63.879 },
+		    { x: -46.7134, y: 60.4156, z: -65.8516 },
+		    { x: -45.6368, y: 56.7291, z: -67.5664 },
+		    { x: -45.251, y: 54.4219, z: -68.1407 },
+		    { x: -45.122, y: 53.3897, z: -68.386 },
+		    { x: -45.2856, y: 51.8317, z: -68.5715 },
+		    { x: -45.3532, y: 50.7729, z: -68.6263 },
+		    { x: -44.8292, y: 49.0485, z: -68.7354 },
+		    { x: -44.1694, y: 47.3305, z: -68.7044 },
+		    { x: -43.6508, y: 46.2262, z: -68.7206 },
+		    { x: -44.3122, y: 45.4283, z: -69.0685 },
+		    { x: -45.9353, y: 45.4167, z: -69.7118 },
+		    { x: -48.2429, y: 46.6819, z: -69.8598 },
+		    { x: -51.4231, y: 48.0187, z: -69.8374 },
+		    { x: -54.1783, y: 50.0075, z: -69.0998 },
+		    { x: -57.4843, y: 51.3684, z: -66.9559 },
+		    { x: -60.5702, y: 51.5189, z: -64.5248 },
+		    { x: -63.2296, y: 52.1824, z: -61.5585 },
+		    { x: -66.6546, y: 52.5453, z: -60.5428 },
+		    { x: -76.6132, y: 52.7512, z: -60.8217 },
+		    { x: -79.3579, y: 57.8792, z: -51.1674 },
+		    { x: -82.4832, y: 62.7241, z: -46.7705 },
+		    { x: -85.8027, y: 63.4993, z: -44.8534 },
+		    { x: -89.7929, y: 63.9835, z: -42.7157 },
+		    { x: -94.4323, y: 64.2501, z: -41.4448 },
+		    { x: -98.2288, y: 64.0493, z: -40.3644 },
+		    { x: -109.446, y: 64.2836, z: -35.8675 },
+		    { x: -110.952, y: 62.6844, z: -32.8058 },
+		    { x: -110.965, y: 59.7634, z: -29.7826 },
+		    { x: -109.58, y: 59.168, z: -25.3716 },
+		    { x: -110.826, y: 59.9899, z: -23.6033 },
+		    { x: -117.682, y: 61.768, z: -29.6027 },
+		    { x: -122.905, y: 63.0018, z: -30.4916 },
+		    { x: -126.117, y: 63.1485, z: -30.6911 },
+		    { x: -128.768, y: 61.5732, z: -29.9666 },
+		    { x: -141.401, y: 67.743, z: -32.0506 },
+		    { x: -145.95, y: 69.9792, z: -37.5504 },
+		    { x: -136.736, y: 62.7556, z: -45.2782 },
+		    { x: -131.437, y: 57.8147, z: -47.3852 },
+		    { x: -128.679, y: 53.0378, z: -48.0588 },
+		    { x: -127.432, y: 52.473, z: -47.8239 },
+		    { x: -127.333, y: 51.8987, z: -46.755 },
+		    { x: -139.217, y: 53.7077, z: -38.961 },
+		    { x: -154.47, y: 58.1993, z: -32.5204 },
+		    { x: -180.555, y: 68.6949, z: -20.0403 } ]
+    },
+    {
+	"name": "5",
+	"points": [ { x: -79.4077, y: 53.3224, z: -50.6046 },
+		    { x: -79.3446, y: 53.4074, z: -50.6101 },
+		    { x: -79.3193, y: 53.451, z: -50.5993 },
+		    { x: -79.3883, y: 53.5611, z: -50.5053 },
+		    { x: -79.659, y: 53.6821, z: -50.172 },
+		    { x: -80.0106, y: 53.9157, z: -49.7295 },
+		    { x: -80.1831, y: 54.0586, z: -49.4939 },
+		    { x: -80.4054, y: 54.3968, z: -49.1917 },
+		    { x: -80.4856, y: 54.6211, z: -49.0684 },
+		    { x: -80.504, y: 54.7223, z: -49.0315 },
+		    { x: -80.5434, y: 54.7909, z: -49.0125 },
+		    { x: -80.5606, y: 54.8618, z: -48.9606 },
+		    { x: -80.5458, y: 54.901, z: -48.9573 },
+		    { x: -80.6256, y: 55.0539, z: -48.9156 },
+		    { x: -80.6208, y: 55.1389, z: -48.9844 },
+		    { x: -80.3539, y: 55.1974, z: -49.3194 },
+		    { x: -80.0266, y: 55.1544, z: -49.5919 },
+		    { x: -79.0239, y: 55.3548, z: -50.3368 },
+		    { x: -77.4409, y: 55.834, z: -50.8433 },
+		    { x: -76.3235, y: 56.1529, z: -50.8656 },
+		    { x: -73.7862, y: 56.2567, z: -51.4327 },
+		    { x: -70.9344, y: 56.4386, z: -52.8474 },
+		    { x: -68.967, y: 56.5516, z: -53.3989 },
+		    { x: -65.1422, y: 57.2422, z: -54.6715 },
+		    { x: -60.985, y: 57.907, z: -55.8765 },
+		    { x: -56.7669, y: 57.8403, z: -57.3286 },
+		    { x: -55.0754, y: 58.5061, z: -57.9613 },
+		    { x: -51.3564, y: 59.2676, z: -59.7542 },
+		    { x: -47.6761, y: 58.8569, z: -61.8787 },
+		    { x: -45.5808, y: 58.4692, z: -62.9012 },
+		    { x: -40.4765, y: 58.3087, z: -62.5596 },
+		    { x: -36.1887, y: 56.8775, z: -62.9707 },
+		    { x: -34.4031, y: 56.8271, z: -63.3673 },
+		    { x: -29.5597, y: 57.2962, z: -65.8309 },
+		    { x: -26.0837, y: 55.9364, z: -66.3148 },
+		    { x: -23.9289, y: 55.433, z: -65.8835 },
+		    { x: -20.1243, y: 55.0274, z: -63.3096 },
+		    { x: -15.8897, y: 56.2495, z: -57.7479 },
+		    { x: -13.9786, y: 55.5302, z: -56.9483 },
+		    { x: -13.1106, y: 54.8837, z: -56.4702 },
+		    { x: -11.2295, y: 53.6025, z: -55.759 },
+		    { x: -9.49, y: 51.7756, z: -55.2132 },
+		    { x: -8.79634, y: 50.7256, z: -54.8099 },
+		    { x: -7.99408, y: 48.4049, z: -53.3079 },
+		    { x: -7.97176, y: 46.0502, z: -52.1199 },
+		    { x: -7.16828, y: 44.0147, z: -51.3489 },
+		    { x: -6.88915, y: 42.8406, z: -50.4907 },
+		    { x: -6.54985, y: 41.1031, z: -49.3499 },
+		    { x: -6.56768, y: 38.9333, z: -47.9585 },
+		    { x: -6.75143, y: 37.8144, z: -46.6625 },
+		    { x: -7.51541, y: 36.2902, z: -43.7326 },
+		    { x: -8.09215, y: 35.6021, z: -42.1924 },
+		    { x: -9.67318, y: 34.8424, z: -40.7264 },
+		    { x: -10.8047, y: 34.3103, z: -39.7958 },
+		    { x: -12.8652, y: 33.7609, z: -38.4582 },
+		    { x: -15.1778, y: 33.2398, z: -36.6551 },
+		    { x: -16.1261, y: 33.2833, z: -36.061 },
+		    { x: -19.4397, y: 33.2768, z: -34.3904 },
+		    { x: -22.9578, y: 33.5887, z: -33.0436 },
+		    { x: -24.9456, y: 33.8876, z: -32.6512 },
+		    { x: -28.9943, y: 34.2749, z: -31.992 },
+		    { x: -31.8591, y: 34.9473, z: -31.7713 },
+		    { x: -34.869, y: 35.745, z: -31.4638 },
+		    { x: -36.0956, y: 36.0043, z: -31.0635 },
+		    { x: -38.252, y: 36.5712, z: -30.639 },
+		    { x: -41.8292, y: 36.8092, z: -30.1989 },
+		    { x: -43.2711, y: 36.8262, z: -30.1627 },
+		    { x: -48.2801, y: 37.2232, z: -32.8102 },
+		    { x: -49.9214, y: 37.473, z: -34.152 },
+		    { x: -51.3242, y: 37.9526, z: -36.5693 },
+		    { x: -53.1635, y: 39.8564, z: -39.9382 },
+		    { x: -54.1908, y: 41.7929, z: -40.6051 },
+		    { x: -54.9949, y: 46.9734, z: -41.3947 },
+		    { x: -54.9702, y: 48.8692, z: -42.4559 },
+		    { x: -53.8982, y: 52.4114, z: -44.0827 },
+		    { x: -53.2983, y: 57.6801, z: -45.9429 },
+		    { x: -52.6286, y: 59.9855, z: -46.7303 },
+		    { x: -50.722, y: 63.7412, z: -48.7699 },
+		    { x: -50.722, y: 63.7412, z: -48.7699 },
+		    { x: -47.8849, y: 75.2707, z: -60.252 },
+		    { x: -46.9023, y: 77.7046, z: -62.1912 },
+		    { x: -43.8448, y: 80.3891, z: -64.7384 },
+		    { x: -41.628, y: 81.4355, z: -66.5038 },
+		    { x: -36.8018, y: 83.9964, z: -69.732 },
+		    { x: -32.4253, y: 86.1058, z: -72.3813 },
+		    { x: -28.0471, y: 86.9187, z: -74.8422 },
+		    { x: -25.9115, y: 87.587, z: -75.9239 },
+		    { x: -21.0788, y: 88.397, z: -78.3712 },
+		    { x: -15.7893, y: 88.724, z: -80.0598 },
+		    { x: -12.8957, y: 88.5344, z: -81.1249 },
+		    { x: -7.37823, y: 87.3982, z: -81.9749 },
+		    { x: -2.2465, y: 86.9422, z: -82.4096 },
+		    { x: 3.77119, y: 87.0301, z: -81.7295 },
+		    { x: 6.45704, y: 86.6943, z: -81.5852 },
+		    { x: 11.978, y: 86.9751, z: -81.2322 },
+		    { x: 16.7925, y: 84.8334, z: -81.7728 },
+		    { x: 20.0497, y: 83.1944, z: -82.2855 },
+		    { x: 25.8277, y: 81.1956, z: -82.4611 },
+		    { x: 32.111, y: 78.7428, z: -82.2306 },
+		    { x: 35.197, y: 77.3158, z: -81.516 },
+		    { x: 40.4325, y: 74.6695, z: -79.9298 },
+		    { x: 45.5619, y: 73.8924, z: -77.8728 },
+		    { x: 51.9285, y: 73.8938, z: -75.4619 },
+		    { x: 54.7225, y: 72.9725, z: -73.8057 },
+		    { x: 61.1619, y: 73.83, z: -68.2403 },
+		    { x: 66.3869, y: 74.6318, z: -67.7188 },
+		    { x: 68.0267, y: 74.7582, z: -67.5451 },
+		    { x: 71.8802, y: 74.4797, z: -66.7115 },
+		    { x: 73.5299, y: 74.3801, z: -66.3813 },
+		    { x: 74.6818, y: 73.687, z: -65.2969 },
+		    { x: 76.0194, y: 72.8379, z: -64.0391 },
+		    { x: 76.3111, y: 71.8191, z: -63.1828 },
+		    { x: 76.1903, y: 70.7391, z: -62.7233 },
+		    { x: 75.8015, y: 70.1453, z: -62.4997 },
+		    { x: 75.4315, y: 68.8789, z: -61.8683 },
+		    { x: 72.0959, y: 65.6428, z: -59.9241 },
+		    { x: 70.3159, y: 64.9303, z: -59.2593 },
+		    { x: 67.5603, y: 63.1187, z: -58.2828 },
+		    { x: 65.2677, y: 62.1099, z: -57.1875 },
+		    { x: 60.7805, y: 59.7478, z: -56.1448 },
+		    { x: 56.2435, y: 57.9311, z: -54.214 },
+		    { x: 51.5723, y: 56.1766, z: -53.6197 },
+		    { x: 48.3728, y: 55.1838, z: -53.0187 },
+		    { x: 46.3283, y: 54.474, z: -52.7346 },
+		    { x: 42.4584, y: 52.5055, z: -51.9723 },
+		    { x: 39.1603, y: 51.9422, z: -51.0804 },
+		    { x: 36.5096, y: 50.6904, z: -50.5419 },
+		    { x: 31.6481, y: 47.6134, z: -49.9642 },
+		    { x: 28.0105, y: 46.826, z: -49.3799 },
+		    { x: 26.3859, y: 46.6499, z: -49.1948 },
+		    { x: 23.3189, y: 46.2158, z: -49.3854 },
+		    { x: 19.842, y: 46.0155, z: -50.0502 },
+		    { x: 17.0542, y: 46.3395, z: -50.7615 },
+		    { x: 15.2089, y: 46.6017, z: -51.917 },
+		    { x: 12.1014, y: 47.5194, z: -54.2099 },
+		    { x: 9.53476, y: 48.2768, z: -55.4918 },
+		    { x: 8.50418, y: 48.5881, z: -55.9471 },
+		    { x: 5.45529, y: 49.2956, z: -57.4222 },
+		    { x: 3.40848, y: 48.4486, z: -59.31 },
+		    { x: 2.81181, y: 49.1504, z: -59.7663 },
+		    { x: 1.54113, y: 50.8016, z: -60.657 },
+		    { x: 0.473556, y: 51.7516, z: -62.5656 },
+		    { x: 2.26765, y: 53.8536, z: -66.5135 },
+		    { x: 4.30911, y: 56.6589, z: -68.9805 },
+		    { x: 6.60568, y: 59.9905, z: -71.3163 },
+		    { x: 7.85814, y: 61.4982, z: -72.4683 },
+		    { x: 10.5301, y: 64.0273, z: -74.7372 },
+		    { x: 14.2829, y: 66.9457, z: -77.2855 },
+		    { x: 15.693, y: 68.1835, z: -78.1922 },
+		    { x: 19.762, y: 71.2398, z: -80.2989 },
+		    { x: 24.9092, y: 75.8728, z: -82.4687 },
+		    { x: 26.8233, y: 77.4745, z: -83.0902 },
+		    { x: 26.8233, y: 77.4745, z: -83.0902 },
+		    { x: 37.0066, y: 82.6643, z: -84.7606 },
+		    { x: 39.6356, y: 83.4957, z: -84.8679 },
+		    { x: 45.6018, y: 85.5425, z: -84.1608 },
+		    { x: 51.773, y: 86.896, z: -84.2374 },
+		    { x: 57.3125, y: 87.0869, z: -83.4546 },
+		    { x: 59.9339, y: 87.0579, z: -83.0806 },
+		    { x: 65.534, y: 86.4384, z: -81.997 },
+		    { x: 71.6467, y: 85.9353, z: -81.3088 },
+		    { x: 74.4863, y: 85.2878, z: -80.866 },
+		    { x: 80.9217, y: 84.225, z: -80.6132 },
+		    { x: 87.4503, y: 83.3674, z: -80.1859 },
+		    { x: 93.6736, y: 82.7098, z: -78.5903 },
+		    { x: 96.7721, y: 82.175, z: -77.9465 },
+		    { x: 102.286, y: 82.1279, z: -76.8818 },
+		    { x: 108.807, y: 81.0219, z: -78.9879 },
+		    { x: 111.85, y: 80.3796, z: -78.494 },
+		    { x: 116.775, y: 79.9792, z: -77.1354 },
+		    { x: 121.77, y: 78.7458, z: -76.5411 },
+		    { x: 124.392, y: 77.8515, z: -75.3947 },
+		    { x: 128.822, y: 77.8651, z: -72.2067 },
+		    { x: 132.836, y: 77.3298, z: -67.9326 },
+		    { x: 135.321, y: 76.3474, z: -64.9497 },
+		    { x: 135.671, y: 75.7154, z: -64.2781 },
+		    { x: 135.847, y: 74.5993, z: -63.4266 },
+		    { x: 134.223, y: 71.7672, z: -66.3248 },
+		    { x: 135.42, y: 71.3772, z: -67.0808 },
+		    { x: 144.605, y: 40.1277, z: -19.7135 },
+		    { x: 152.72, y: 37.5434, z: -18.4134 },
+		    { x: 155.762, y: 38.4838, z: -19.0305 },
+		    { x: 157.025, y: 38.5155, z: -18.8289 },
+		    { x: 158.972, y: 38.8375, z: -19.328 },
+		    { x: 159.468, y: 39.97, z: -20.6884 },
+		    { x: 159.917, y: 40.0403, z: -20.6147 },
+		    { x: 160.318, y: 40.111, z: -19.9513 },
+		    { x: 161.259, y: 42.0327, z: -20.2775 },
+		    { x: 162.13, y: 41.5717, z: -19.3712 },
+		    { x: 166.693, y: 43.0166, z: -17.8864 },
+		    { x: 167.79, y: 45.2215, z: -20.0591 },
+		    { x: 167.243, y: 49.181, z: -26.0414 },
+		    { x: 166.743, y: 52.0377, z: -32.7253 },
+		    { x: 167.544, y: 51.6948, z: -33.9987 },
+		    { x: 165.007, y: 53.4777, z: -37.2105 },
+		    { x: 164.288, y: 53.902, z: -37.8728 },
+		    { x: 160.133, y: 51.2654, z: -37.0749 },
+		    { x: 158.356, y: 49.8361, z: -36.8886 },
+		    { x: 157.727, y: 49.8479, z: -37.1915 },
+		    { x: 157.14, y: 49.8553, z: -37.5094 },
+		    { x: 156.776, y: 49.9188, z: -37.8714 },
+		    { x: 156.882, y: 48.9631, z: -36.7971 },
+		    { x: 156.837, y: 48.9107, z: -36.5313 },
+		    { x: 158.73, y: 47.5466, z: -33.6347 },
+		    { x: 158.79, y: 49.5829, z: -33.5394 },
+		    { x: 158.849, y: 50.6596, z: -33.5182 },
+		    { x: 158.828, y: 51.1086, z: -33.2418 },
+		    { x: 158.471, y: 49.8814, z: -31.482 },
+		    { x: 158.416, y: 49.2326, z: -30.3818 },
+		    { x: 158.182, y: 49.097, z: -29.7739 },
+		    { x: 156.75, y: 48.2212, z: -29.8297 },
+		    { x: 156.485, y: 48.6863, z: -30.4725 },
+		    { x: 155.668, y: 51.1285, z: -33.5974 },
+		    { x: 155.276, y: 52.1742, z: -35.0104 },
+		    { x: 155.399, y: 54.4092, z: -38.8656 },
+		    { x: 155.526, y: 55.4125, z: -40.7983 },
+		    { x: 155.645, y: 56.6215, z: -42.6084 },
+		    { x: 155.315, y: 57.5863, z: -44.1508 },
+		    { x: 154.948, y: 58.0395, z: -45.2511 },
+		    { x: 154.601, y: 59.1931, z: -46.5975 },
+		    { x: 154.16, y: 60.8239, z: -47.973 },
+		    { x: 153.478, y: 61.9735, z: -49.8941 },
+		    { x: 153.476, y: 62.1684, z: -51.1406 },
+		    { x: 153.86, y: 62.0478, z: -52.6925 },
+		    { x: 153.932, y: 62.2054, z: -53.6727 },
+		    { x: 153.854, y: 62.1534, z: -54.0108 },
+		    { x: 153.779, y: 61.8553, z: -54.1989 },
+		    { x: 153.45, y: 61.9267, z: -54.3939 },
+		    { x: 153.362, y: 61.9485, z: -54.4427 },
+		    { x: 153.136, y: 62.119, z: -54.6359 },
+		    { x: 152.995, y: 62.3409, z: -54.9033 },
+		    { x: 152.569, y: 63.5348, z: -55.3772 },
+		    { x: 152.415, y: 63.8882, z: -55.5474 },
+		    { x: 151.811, y: 64.3188, z: -55.6914 },
+		    { x: 151.36, y: 64.7215, z: -55.7461 },
+		    { x: 151.166, y: 64.9545, z: -55.7686 },
+		    { x: 151.145, y: 65.4737, z: -55.291 },
+		    { x: 150.467, y: 66.6726, z: -54.3861 },
+		    { x: 150.533, y: 66.7606, z: -54.1824 },
+		    { x: 150.103, y: 66.3003, z: -53.8846 },
+		    { x: 148.305, y: 65.3636, z: -53.8362 },
+		    { x: 147.083, y: 65.3506, z: -53.9807 },
+		    { x: 146.937, y: 65.0909, z: -53.9084 },
+		    { x: 146.744, y: 65.059, z: -53.7928 },
+		    { x: 146.707, y: 65.0807, z: -53.7631 },
+		    { x: 146.414, y: 65.1907, z: -53.9035 },
+		    { x: 145.865, y: 65.0724, z: -53.8906 },
+		    { x: 145.256, y: 64.762, z: -53.5969 },
+		    { x: 144.881, y: 64.3502, z: -53.4084 },
+		    { x: 144.403, y: 63.8069, z: -53.1922 },
+		    { x: 144.199, y: 63.6155, z: -53.0063 },
+		    { x: 144.419, y: 63.2206, z: -52.7635 },
+		    { x: 144.898, y: 62.8348, z: -52.4885 },
+		    { x: 145.508, y: 63.1365, z: -51.4988 },
+		    { x: 145.357, y: 63.046, z: -51.2857 },
+		    { x: 145.124, y: 62.9787, z: -51.0305 },
+		    { x: 147.791, y: 61.5621, z: -53.0735 },
+		    { x: 148.859, y: 61.3106, z: -54.4537 },
+		    { x: 149.637, y: 61.4724, z: -55.3418 },
+		    { x: 189.183, y: 110.713, z: 14.9502 },
+		    { x: 189.638, y: 106.395, z: 15.996 },
+		    { x: 189.965, y: 103.443, z: 15.6064 },
+		    { x: 190.445, y: 104.498, z: 14.9585 },
+		    { x: 191.133, y: 105.012, z: 14.8377 },
+		    { x: 191.581, y: 106.577, z: 14.8509 },
+		    { x: 192.303, y: 108.347, z: 14.3928 },
+		    { x: 192.774, y: 110.521, z: 14.1576 },
+		    { x: 192.745, y: 109.82, z: 14.2381 },
+		    { x: 192.681, y: 110.182, z: 14.1473 },
+		    { x: 192.566, y: 110.325, z: 13.8946 },
+		    { x: 192.547, y: 110.12, z: 13.8152 },
+		    { x: 192.568, y: 108.956, z: 13.9966 },
+		    { x: 192.634, y: 106.599, z: 14.3508 },
+		    { x: 193.072, y: 106.886, z: 14.1006 },
+		    { x: 193.124, y: 107.276, z: 14.1249 },
+		    { x: 193.571, y: 107.079, z: 14.0266 },
+		    { x: 193.808, y: 106.692, z: 13.9374 },
+		    { x: 193.68, y: 107.637, z: 14.0762 },
+		    { x: 193.61, y: 108.503, z: 14.2351 } ]
+    },
+    {
+	"name": "5'",
+	"points": [ { x: 106.613, y: 57.5143, z: -99.4655 },
+		    { x: 106.719, y: 57.5731, z: -99.4291 },
+		    { x: 106.945, y: 58.124, z: -99.4593 },
+		    { x: 107.069, y: 58.3967, z: -99.4692 },
+		    { x: 107.102, y: 58.5826, z: -99.418 },
+		    { x: 107.263, y: 58.7192, z: -99.3451 },
+		    { x: 107.377, y: 58.7563, z: -99.2146 },
+		    { x: 107.384, y: 58.7385, z: -99.1892 },
+		    { x: 107.404, y: 58.7385, z: -99.1667 },
+		    { x: 107.427, y: 58.7085, z: -99.137 },
+		    { x: 107.607, y: 58.869, z: -99.1723 },
+		    { x: 107.656, y: 58.8883, z: -99.1925 },
+		    { x: 107.573, y: 58.9435, z: -99.2219 },
+		    { x: 107.348, y: 59.0112, z: -99.3097 },
+		    { x: 107.217, y: 59.0615, z: -99.3135 },
+		    { x: 106.356, y: 58.5497, z: -99.1703 },
+		    { x: 105.734, y: 58.6192, z: -99.1409 },
+		    { x: 105.202, y: 58.8026, z: -99.1519 },
+		    { x: 103.226, y: 58.8012, z: -98.9239 },
+		    { x: 100.772, y: 58.1999, z: -98.3578 },
+		    { x: 98.9335, y: 58.6612, z: -98.1116 },
+		    { x: 97.6802, y: 59.0805, z: -98.3292 },
+		    { x: 96.9715, y: 62.9741, z: -99.6455 },
+		    { x: 95.3581, y: 65.9678, z: -102.283 },
+		    { x: 94.2511, y: 66.8243, z: -103.347 },
+		    { x: 87.6032, y: 64.3946, z: -102.003 },
+		    { x: 83.9148, y: 65.2779, z: -101.06 },
+		    { x: 81.6266, y: 65.7791, z: -101.351 },
+		    { x: 77.7879, y: 65.1637, z: -101.345 },
+		    { x: 74.1671, y: 66.2212, z: -102.15 },
+		    { x: 71.0953, y: 64.1559, z: -102.575 },
+		    { x: 66.438, y: 62.6988, z: -106.264 },
+		    { x: 63.4804, y: 63.299, z: -109.135 },
+		    { x: 60.3657, y: 65.9938, z: -107.916 },
+		    { x: 59.0789, y: 66.5725, z: -108.547 },
+		    { x: 56.7329, y: 67.7361, z: -108.534 },
+		    { x: 53.7295, y: 68.3048, z: -108.095 },
+		    { x: 52.0323, y: 68.0706, z: -107.869 },
+		    { x: 48.7391, y: 67.0153, z: -107.16 },
+		    { x: 44.3619, y: 65.8143, z: -106.822 },
+		    { x: 42.1861, y: 64.8656, z: -106.613 },
+		    { x: 37.5951, y: 63.4708, z: -105.957 },
+		    { x: 33.3228, y: 62.5053, z: -103.744 },
+		    { x: 29.4946, y: 62.9635, z: -102.394 },
+		    { x: 27.8904, y: 63.1179, z: -101.621 },
+		    { x: 23.0542, y: 60.5063, z: -99.4244 },
+		    { x: 18.5192, y: 56.6529, z: -97.2221 },
+		    { x: 17.1061, y: 56.0887, z: -96.363 },
+		    { x: 11.8446, y: 53.3386, z: -93.5286 },
+		    { x: 9.29506, y: 53.3428, z: -91.1121 },
+		    { x: 8.24075, y: 52.7764, z: -90.2611 },
+		    { x: 5.22695, y: 53.4783, z: -87.0327 },
+		    { x: 3.1702, y: 52.9878, z: -84.0529 },
+		    { x: 1.89343, y: 51.2555, z: -82.6557 },
+		    { x: 1.36957, y: 50.2788, z: -81.7259 },
+		    { x: 1.0238, y: 48.3385, z: -79.8926 },
+		    { x: 0.571126, y: 46.2773, z: -77.7278 },
+		    { x: 0.42862, y: 45.2669, z: -76.861 },
+		    { x: -0.139143, y: 44.7673, z: -75.1451 },
+		    { x: -1.03322, y: 44.5562, z: -73.5915 },
+		    { x: -0.699778, y: 44.4608, z: -72.7588 },
+		    { x: -0.609519, y: 42.778, z: -71.3493 },
+		    { x: 0.0819411, y: 41.2067, z: -70.4924 },
+		    { x: 1.40828, y: 39.2129, z: -70.2425 },
+		    { x: 2.45564, y: 37.9041, z: -70.2486 },
+		    { x: 6.09931, y: 35.2006, z: -70.181 },
+		    { x: 10.607, y: 33.5977, z: -69.9518 },
+		    { x: 14.5699, y: 32.4992, z: -69.13 },
+		    { x: 14.5699, y: 32.4992, z: -69.13 },
+		    { x: 25.4301, y: 30.2955, z: -69.753 },
+		    { x: 27.6284, y: 30.061, z: -70.1721 },
+		    { x: 32.1486, y: 29.6539, z: -70.9177 },
+		    { x: 36.2085, y: 28.7, z: -71.2942 },
+		    { x: 40.3006, y: 28.4445, z: -71.6752 },
+		    { x: 42.3213, y: 28.789, z: -72.0102 },
+		    { x: 45.941, y: 29.2991, z: -72.8016 },
+		    { x: 50.4957, y: 30.6779, z: -74.5737 },
+		    { x: 52.0804, y: 30.8998, z: -75.11 },
+		    { x: 54.895, y: 31.3661, z: -76.558 },
+		    { x: 58.0995, y: 32.6099, z: -77.3394 },
+		    { x: 60.5847, y: 33.9666, z: -77.6069 },
+		    { x: 62.8193, y: 36.9879, z: -77.3822 },
+		    { x: 64.8691, y: 41.1615, z: -76.3281 },
+		    { x: 64.1825, y: 43.7822, z: -76.9291 },
+		    { x: 65.3573, y: 46.9839, z: -78.9016 },
+		    { x: 65.1714, y: 50.9782, z: -85.5503 },
+		    { x: 64.2424, y: 53.0722, z: -88.0035 },
+		    { x: 63.3711, y: 54.3114, z: -88.7634 },
+		    { x: 61.9751, y: 59.8856, z: -93.4182 },
+		    { x: 58.8965, y: 63.4669, z: -98.0926 },
+		    { x: 57.1099, y: 64.7207, z: -99.412 },
+		    { x: 52.9132, y: 65.626, z: -101.361 },
+		    { x: 48.0173, y: 67.2664, z: -102.949 },
+		    { x: 42.7386, y: 71.4507, z: -104.436 },
+		    { x: 40.0835, y: 73.7329, z: -104.856 },
+		    { x: 34.184, y: 76.5761, z: -105.049 },
+		    { x: 31.1116, y: 77.4416, z: -105.352 },
+		    { x: 24.8941, y: 77.9714, z: -105.599 },
+		    { x: 18.2898, y: 77.683, z: -105.178 },
+		    { x: 11.1757, y: 77.2558, z: -104.004 },
+		    { x: 7.80903, y: 77.8061, z: -103.362 },
+		    { x: 1.18267, y: 77.9489, z: -102.18 },
+		    { x: -5.50235, y: 79.1487, z: -97.7046 },
+		    { x: -8.10589, y: 78.718, z: -95.9727 },
+		    { x: -12.8419, y: 75.7231, z: -95.0624 },
+		    { x: -18.6378, y: 75.2488, z: -95.7189 },
+		    { x: -26.0538, y: 76.7833, z: -96.8401 },
+		    { x: -29.0262, y: 76.4312, z: -96.7435 },
+		    { x: -34.6045, y: 74.9721, z: -96.1904 },
+		    { x: -40.3673, y: 74.5413, z: -94.5136 },
+		    { x: -42.2403, y: 74.2462, z: -94.1515 },
+		    { x: -45.8904, y: 73.9503, z: -93.2552 },
+		    { x: -49.6272, y: 71.9913, z: -92.3367 },
+		    { x: -51.5628, y: 70.7853, z: -91.7734 },
+		    { x: -55.4165, y: 68.4799, z: -90.5544 },
+		    { x: -58.7617, y: 66.3406, z: -89.5738 },
+		    { x: -61.9006, y: 65.236, z: -88.1535 },
+		    { x: -63.1982, y: 65.5129, z: -87.326 },
+		    { x: -64.9566, y: 62.8221, z: -85.8818 },
+		    { x: -66.4465, y: 60.584, z: -84.9271 },
+		    { x: -67.0884, y: 59.7782, z: -84.4067 },
+		    { x: -68.043, y: 57.8781, z: -83.026 },
+		    { x: -68.2848, y: 55.3468, z: -80.9584 },
+		    { x: -68.5985, y: 54.8079, z: -79.2659 },
+		    { x: -69.0372, y: 52.5589, z: -76.7654 },
+		    { x: -69.1397, y: 50.5445, z: -74.9142 },
+		    { x: -69.7646, y: 47.3861, z: -74.9209 },
+		    { x: -69.6528, y: 47.8158, z: -73.6182 },
+		    { x: -67.7052, y: 48.4369, z: -72.1454 },
+		    { x: -65.669, y: 48.8736, z: -70.7363 },
+		    { x: -62.0272, y: 49.2039, z: -68.0206 },
+		    { x: -58.6363, y: 51.9198, z: -65.423 },
+		    { x: -54.0432, y: 50.5097, z: -64.5413 },
+		    { x: -52.2276, y: 49.6026, z: -64.8882 },
+		    { x: -46.5764, y: 54.3352, z: -62.8579 },
+		    { x: -43.0579, y: 50.9128, z: -61.9071 },
+		    { x: -41.403, y: 50.3692, z: -63.4542 },
+		    { x: -40.4457, y: 50.0546, z: -64.6557 },
+		    { x: -37.245, y: 49.3537, z: -67.7608 },
+		    { x: -32.6447, y: 49.9348, z: -71.1492 },
+		    { x: -30.3071, y: 50.2057, z: -72.7847 },
+		    { x: -25.6554, y: 49.9988, z: -75.6438 },
+		    { x: -23.8563, y: 50.4668, z: -77.0349 },
+		    { x: -23.8563, y: 50.4668, z: -77.0349 },
+		    { x: -17.2923, y: 53.9884, z: -82.8862 },
+		    { x: -15.0981, y: 56.292, z: -86.0758 },
+		    { x: -14.5423, y: 57.884, z: -89.8966 },
+		    { x: -14.517, y: 58.3838, z: -91.0093 },
+		    { x: -14.6191, y: 58.5993, z: -93.6733 },
+		    { x: -15.9913, y: 57.7035, z: -96.0941 },
+		    { x: -17.3154, y: 57.5251, z: -97.2574 },
+		    { x: -20.6959, y: 57.3409, z: -99.8938 },
+		    { x: -24.0474, y: 57.6173, z: -101.628 },
+		    { x: -26.47, y: 58.835, z: -101.803 },
+		    { x: -30.8988, y: 60.1754, z: -102.469 },
+		    { x: -34.6372, y: 63.907, z: -104.363 },
+		    { x: -40.495, y: 64.3158, z: -104.947 },
+		    { x: -45.7472, y: 62.0055, z: -103.803 },
+		    { x: -54.3168, y: 56.8139, z: -102.648 },
+		    { x: -65.077, y: 54.292, z: -98.4202 },
+		    { x: -66.2477, y: 57.3845, z: -99.8502 },
+		    { x: -70.5393, y: 61.975, z: -102.503 },
+		    { x: -79.6271, y: 61.6183, z: -98.2038 },
+		    { x: -72.3947, y: 52.3252, z: -39.0185 },
+		    { x: -69.7099, y: 45.7619, z: -35.5654 },
+		    { x: -69.9594, y: 44.6388, z: -36.6749 },
+		    { x: -71.416, y: 43.8915, z: -35.2567 },
+		    { x: -72.5448, y: 43.6843, z: -34.307 },
+		    { x: -75.4034, y: 43.2261, z: -32.5742 },
+		    { x: -78.731, y: 42.8634, z: -31.4116 },
+		    { x: -80.6441, y: 42.7013, z: -30.6522 },
+		    { x: -84.1475, y: 42.6361, z: -29.4104 },
+		    { x: -87.4613, y: 42.69, z: -28.1342 },
+		    { x: -92.8103, y: 44.2509, z: -25.0956 },
+		    { x: -94.7452, y: 44.7517, z: -23.9015 },
+		    { x: -98.2062, y: 45.7705, z: -22.7506 },
+		    { x: -102.085, y: 47.5666, z: -21.5202 },
+		    { x: -103.766, y: 48.2858, z: -20.7884 },
+		    { x: -107.221, y: 49.7669, z: -18.767 },
+		    { x: -110.613, y: 51.7962, z: -16.9099 },
+		    { x: -112.228, y: 52.7437, z: -16.1923 },
+		    { x: -115.041, y: 54.275, z: -14.9968 },
+		    { x: -118.035, y: 56.4895, z: -15.4442 },
+		    { x: -119.138, y: 56.6655, z: -15.4694 },
+		    { x: -120.288, y: 56.4267, z: -14.9397 },
+		    { x: -121.126, y: 56.2971, z: -14.7234 },
+		    { x: -125.365, y: 59.4232, z: -21.6274 },
+		    { x: -125.054, y: 58.7678, z: -20.9572 },
+		    { x: -121.608, y: 80.9437, z: -51.8284 },
+		    { x: -128.372, y: 83.8873, z: -59.3655 },
+		    { x: -129.456, y: 83.6178, z: -59.2725 },
+		    { x: -130.829, y: 82.8465, z: -59.1103 },
+		    { x: -131.353, y: 83.4559, z: -59.567 },
+		    { x: -132.223, y: 83.8701, z: -60.103 },
+		    { x: -132.616, y: 83.6992, z: -60.4528 },
+		    { x: -133.92, y: 84.9091, z: -60.6427 },
+		    { x: -134.814, y: 85.8532, z: -60.7355 },
+		    { x: -135.539, y: 86.6797, z: -60.7236 },
+		    { x: -136.769, y: 87.7534, z: -60.6737 },
+		    { x: -138.881, y: 89.0667, z: -60.3161 },
+		    { x: -139.712, y: 89.4742, z: -60.0313 },
+		    { x: -140.853, y: 89.9544, z: -59.8631 },
+		    { x: -142.004, y: 90.3952, z: -59.4977 },
+		    { x: -142.9, y: 90.6065, z: -59.0163 },
+		    { x: -143.331, y: 91.0633, z: -58.666 },
+		    { x: -143.792, y: 90.7348, z: -58.168 },
+		    { x: -144.257, y: 90.8453, z: -57.6824 },
+		    { x: -144.378, y: 90.9023, z: -57.4257 },
+		    { x: -144.372, y: 91.1262, z: -57.0801 },
+		    { x: -144.281, y: 90.7291, z: -56.816 },
+		    { x: -144.209, y: 90.6587, z: -56.7521 },
+		    { x: -144.208, y: 90.4675, z: -56.5027 },
+		    { x: -143.583, y: 90.4737, z: -56.4959 },
+		    { x: -143.04, y: 90.3934, z: -56.4759 },
+		    { x: -142.948, y: 90.1406, z: -56.4142 },
+		    { x: -143.012, y: 89.5239, z: -56.2515 },
+		    { x: -143.028, y: 89.3439, z: -56.0953 },
+		    { x: -143.028, y: 89.3439, z: -56.0953 },
+		    { x: -142.781, y: 89.2426, z: -56.1051 },
+		    { x: -142.385, y: 88.9594, z: -56.2301 },
+		    { x: -142.31, y: 88.8311, z: -56.2468 },
+		    { x: -142.084, y: 88.708, z: -56.2362 },
+		    { x: -141.728, y: 88.5095, z: -56.2187 },
+		    { x: -141.027, y: 88.1234, z: -56.2631 },
+		    { x: -140.77, y: 87.9593, z: -56.2661 },
+		    { x: -140.603, y: 87.5, z: -56.2541 },
+		    { x: -140.351, y: 87.0024, z: -56.3184 },
+		    { x: -140.22, y: 86.7168, z: -56.3695 },
+		    { x: -140.043, y: 87.0618, z: -56.3459 },
+		    { x: -139.813, y: 87.2286, z: -56.2147 },
+		    { x: -139.644, y: 87.2284, z: -56.16 },
+		    { x: -139.355, y: 87.2582, z: -56.0554 },
+		    { x: -139.08, y: 87.4057, z: -55.8138 },
+		    { x: -138.722, y: 87.356, z: -55.6445 },
+		    { x: -138.63, y: 87.3428, z: -55.5438 },
+		    { x: -138.286, y: 87.2718, z: -55.2809 },
+		    { x: -138.26, y: 87.2998, z: -55.0102 },
+		    { x: -138.097, y: 87.2661, z: -54.9086 },
+		    { x: -138.053, y: 87.1784, z: -54.7878 },
+		    { x: -137.952, y: 86.8325, z: -54.8787 },
+		    { x: -137.876, y: 86.6278, z: -54.9417 },
+		    { x: -137.755, y: 86.2911, z: -55.0401 },
+		    { x: -137.461, y: 85.7267, z: -55.4334 },
+		    { x: -137.266, y: 85.2968, z: -55.7131 },
+		    { x: -136.998, y: 84.7835, z: -56.024 },
+		    { x: -136.934, y: 84.7919, z: -56.1504 },
+		    { x: -136.903, y: 84.5982, z: -56.4699 },
+		    { x: -136.859, y: 84.4601, z: -56.5392 },
+		    { x: -136.809, y: 84.2706, z: -56.6358 },
+		    { x: -136.778, y: 84.244, z: -56.6952 },
+		    { x: -136.694, y: 83.998, z: -56.97 },
+		    { x: -136.999, y: 85.2216, z: -57.0088 },
+		    { x: -136.524, y: 85.6866, z: -57.1262 },
+		    { x: -136.384, y: 85.7144, z: -57.1773 },
+		    { x: -136.015, y: 86.0543, z: -57.379 },
+		    { x: -135.624, y: 86.1772, z: -57.3701 },
+		    { x: -135.567, y: 86.4149, z: -57.3144 },
+		    { x: -135.551, y: 86.3702, z: -57.354 },
+		    { x: -135.558, y: 86.402, z: -57.4173 },
+		    { x: -135.615, y: 86.788, z: -57.4892 },
+		    { x: -135.661, y: 86.9815, z: -57.5202 },
+		    { x: -135.729, y: 87.1916, z: -57.6094 },
+		    { x: -135.656, y: 86.9864, z: -57.7706 },
+		    { x: -135.615, y: 87.0639, z: -57.7985 },
+		    { x: -135.521, y: 87.3964, z: -57.8404 },
+		    { x: -135.405, y: 87.7533, z: -57.8877 },
+		    { x: -135.437, y: 87.9359, z: -57.9163 },
+		    { x: -135.456, y: 88.0113, z: -57.9372 },
+		    { x: -135.463, y: 88.0029, z: -57.9857 },
+		    { x: -135.523, y: 88.0831, z: -58.0228 },
+		    { x: -135.572, y: 88.1703, z: -58.041 },
+		    { x: -135.642, y: 88.2553, z: -58.0861 },
+		    { x: -135.771, y: 88.5792, z: -58.1281 },
+		    { x: -135.859, y: 88.8239, z: -58.1731 },
+		    { x: -135.998, y: 89.1688, z: -58.2502 } ]
+    },
+    {
+	"name": "6",
+	"points": [ { x: -63.9776, y: 101.159, z: -38.8082 },
+		    { x: -63.9141, y: 101.306, z: -38.788 },
+		    { x: -63.8549, y: 101.488, z: -38.7598 },
+		    { x: -63.8196, y: 101.626, z: -38.7108 },
+		    { x: -63.8196, y: 101.626, z: -38.7108 },
+		    { x: -63.7472, y: 101.761, z: -38.7231 },
+		    { x: -63.673, y: 101.86, z: -38.7414 },
+		    { x: -63.673, y: 101.86, z: -38.7414 },
+		    { x: -63.603, y: 101.937, z: -38.7605 },
+		    { x: -63.5243, y: 101.979, z: -38.7879 },
+		    { x: -63.5243, y: 101.979, z: -38.7879 },
+		    { x: -63.4667, y: 102.009, z: -38.7985 },
+		    { x: -63.4458, y: 102.036, z: -38.7859 },
+		    { x: -63.4458, y: 102.036, z: -38.7859 },
+		    { x: -63.3886, y: 102.092, z: -38.7458 },
+		    { x: -63.3016, y: 102.031, z: -38.6846 },
+		    { x: -63.3016, y: 102.031, z: -38.6846 },
+		    { x: -63.1266, y: 101.861, z: -38.675 },
+		    { x: -62.7474, y: 101.51, z: -38.698 },
+		    { x: -62.7474, y: 101.51, z: -38.698 },
+		    { x: -62.2219, y: 100.985, z: -38.6566 },
+		    { x: -60.999, y: 99.7457, z: -38.3083 },
+		    { x: -60.999, y: 99.7457, z: -38.3083 },
+		    { x: -59.2895, y: 98.2245, z: -37.822 },
+		    { x: -57.6907, y: 96.5686, z: -37.8458 },
+		    { x: -56.1812, y: 94.2947, z: -38.0197 },
+		    { x: -53.8878, y: 92.1203, z: -37.6801 },
+		    { x: -51.183, y: 89.7558, z: -37.0184 },
+		    { x: -51.183, y: 89.7558, z: -37.0184 },
+		    { x: -46.0947, y: 86.2979, z: -35.3195 },
+		    { x: -44.2619, y: 84.9698, z: -34.681 },
+		    { x: -42.6611, y: 82.5332, z: -34.9676 },
+		    { x: -41.1512, y: 80.0093, z: -35.4429 },
+		    { x: -39.936, y: 78.4965, z: -35.7505 },
+		    { x: -38.7096, y: 77.2062, z: -35.8456 },
+		    { x: -37.508, y: 76.0434, z: -35.963 },
+		    { x: -35.8337, y: 74.563, z: -35.7542 },
+		    { x: -32.3082, y: 71.2411, z: -34.1695 },
+		    { x: -31.3085, y: 70.2907, z: -33.5509 },
+		    { x: -30.7607, y: 69.7519, z: -33.3332 },
+		    { x: -30.1081, y: 68.4832, z: -33.0818 },
+		    { x: -28.4495, y: 65.7519, z: -32.7955 },
+		    { x: -26.9163, y: 64.1831, z: -32.4841 },
+		    { x: -25.3121, y: 62.5842, z: -32.2768 },
+		    { x: -24.003, y: 61.196, z: -32.3765 },
+		    { x: -22.0844, y: 59.1873, z: -32.2887 },
+		    { x: -20.9546, y: 58.3135, z: -32.2611 },
+		    { x: -19.9881, y: 57.6539, z: -32.3236 },
+		    { x: -19.3461, y: 56.8111, z: -32.4449 },
+		    { x: -19.1996, y: 56.4976, z: -32.501 },
+		    { x: -19.2463, y: 56.4118, z: -32.5195 },
+		    { x: -18.6027, y: 55.5861, z: -32.2752 },
+		    { x: -18.6027, y: 55.5861, z: -32.2752 },
+		    { x: -16.1465, y: 52.632, z: -30.9304 },
+		    { x: -15.4662, y: 51.6929, z: -30.4902 },
+		    { x: -14.8437, y: 50.7752, z: -29.9995 },
+		    { x: -14.3232, y: 49.9305, z: -29.5089 },
+		    { x: -14.4342, y: 49.0328, z: -29.4051 },
+		    { x: -13.6779, y: 48.6346, z: -28.7577 },
+		    { x: -13.2065, y: 48.7071, z: -28.4252 },
+		    { x: -13.0681, y: 48.9691, z: -28.2906 },
+		    { x: -13.028, y: 49.276, z: -28.2618 },
+		    { x: -13.2813, y: 49.5479, z: -28.2767 },
+		    { x: -13.3782, y: 49.9818, z: -28.3511 },
+		    { x: -13.4844, y: 50.4271, z: -28.416 },
+		    { x: -13.6279, y: 50.9403, z: -28.471 },
+		    { x: -13.6748, y: 51.5946, z: -28.5304 },
+		    { x: -13.6615, y: 52.3868, z: -28.8434 },
+		    { x: -13.8615, y: 52.9218, z: -28.8784 },
+		    { x: -13.7633, y: 53.8777, z: -29.655 },
+		    { x: -12.9684, y: 55.4845, z: -31.8279 },
+		    { x: -12.6859, y: 57.8021, z: -33.2733 },
+		    { x: -12.6608, y: 58.6336, z: -33.725 },
+		    { x: -12.7824, y: 59.1636, z: -33.9083 },
+		    { x: -13.0414, y: 59.5012, z: -34.1099 },
+		    { x: -13.3006, y: 59.7662, z: -34.2222 },
+		    { x: -13.3006, y: 59.7662, z: -34.2222 },
+		    { x: -13.6577, y: 60.0506, z: -34.3145 },
+		    { x: -13.8276, y: 60.1101, z: -34.3831 },
+		    { x: -14.0477, y: 60.175, z: -34.4859 },
+		    { x: -14.2967, y: 60.2868, z: -34.5879 },
+		    { x: -14.6872, y: 60.2458, z: -34.7985 },
+		    { x: -15.0733, y: 60.1465, z: -35.0064 },
+		    { x: -15.4327, y: 60.0829, z: -35.1609 },
+		    { x: -15.6756, y: 60.1834, z: -35.2151 },
+		    { x: -15.8889, y: 60.3491, z: -35.2695 },
+		    { x: -16.0816, y: 60.4828, z: -35.3616 },
+		    { x: -16.0816, y: 60.4828, z: -35.3616 },
+		    { x: -16.2605, y: 60.6308, z: -35.4875 },
+		    { x: -16.4087, y: 60.9078, z: -35.618 },
+		    { x: -16.5615, y: 61.1871, z: -35.7507 },
+		    { x: -16.7561, y: 61.4562, z: -35.8316 },
+		    { x: -16.7561, y: 61.4562, z: -35.8316 },
+		    { x: -16.959, y: 61.6865, z: -35.8814 },
+		    { x: -17.1553, y: 61.9376, z: -35.9152 },
+		    { x: -17.1553, y: 61.9376, z: -35.9152 },
+		    { x: -17.3447, y: 62.2001, z: -35.9394 },
+		    { x: -17.5563, y: 62.4453, z: -35.9428 },
+		    { x: -17.7553, y: 62.691, z: -35.9508 },
+		    { x: -17.7553, y: 62.691, z: -35.9508 },
+		    { x: -17.9215, y: 62.915, z: -35.9541 },
+		    { x: -17.9215, y: 62.915, z: -35.9541 },
+		    { x: -18.1118, y: 63.2217, z: -35.9805 },
+		    { x: -18.1118, y: 63.2217, z: -35.9805 },
+		    { x: -18.2077, y: 63.4019, z: -35.9891 },
+		    { x: -18.4121, y: 63.5702, z: -35.9469 },
+		    { x: -18.5905, y: 63.7668, z: -35.9119 },
+		    { x: -18.5905, y: 63.7668, z: -35.9119 },
+		    { x: -18.7502, y: 63.9635, z: -35.8763 },
+		    { x: -18.9078, y: 64.1096, z: -35.8298 },
+		    { x: -18.9078, y: 64.1096, z: -35.8298 },
+		    { x: -18.9196, y: 64.3214, z: -35.8153 },
+		    { x: -18.9544, y: 64.5227, z: -35.7758 },
+		    { x: -18.9544, y: 64.5227, z: -35.7758 },
+		    { x: -19.0054, y: 64.7174, z: -35.7205 },
+		    { x: -18.6227, y: 65.6423, z: -35.3696 },
+		    { x: -18.6227, y: 65.6423, z: -35.3696 },
+		    { x: -18.5782, y: 65.901, z: -35.3005 },
+		    { x: -18.5881, y: 66.0591, z: -35.2386 },
+		    { x: -18.5881, y: 66.0591, z: -35.2386 },
+		    { x: -18.5871, y: 66.2347, z: -35.2041 },
+		    { x: -18.6006, y: 66.4118, z: -35.1705 },
+		    { x: -18.6052, y: 66.5742, z: -35.1505 } ]
+    },
+    {
+	"name": "6'",
+	"points": [ { x: 16.8848, y: 40.3609, z: -23.9701 },
+		    { x: 16.7535, y: 40.445, z: -23.9742 },
+		    { x: 16.6723, y: 40.4916, z: -23.9825 },
+		    { x: 16.6186, y: 40.5121, z: -23.9961 },
+		    { x: 16.6186, y: 40.5121, z: -23.9961 },
+		    { x: 16.5854, y: 40.5297, z: -24.0132 },
+		    { x: 16.563, y: 40.537, z: -24.0435 },
+		    { x: 16.5544, y: 40.5448, z: -24.0684 },
+		    { x: 16.5686, y: 40.5284, z: -24.068 },
+		    { x: 16.5686, y: 40.5284, z: -24.068 },
+		    { x: 16.5986, y: 40.4965, z: -24.0572 },
+		    { x: 16.63, y: 40.4462, z: -24.0433 },
+		    { x: 16.63, y: 40.4462, z: -24.0433 },
+		    { x: 16.6632, y: 40.4096, z: -24.0245 },
+		    { x: 16.7133, y: 40.3625, z: -24.0044 },
+		    { x: 16.7344, y: 40.3444, z: -23.9963 },
+		    { x: 16.7344, y: 40.3444, z: -23.9963 },
+		    { x: 16.648, y: 40.4068, z: -24.0255 },
+		    { x: 16.5525, y: 40.4409, z: -24.2015 },
+		    { x: 16.5525, y: 40.4409, z: -24.2015 },
+		    { x: 16.2046, y: 40.6054, z: -24.7513 },
+		    { x: 15.5308, y: 41.0564, z: -25.2449 },
+		    { x: 15.5308, y: 41.0564, z: -25.2449 },
+		    { x: 14.682, y: 41.9049, z: -26.2283 },
+		    { x: 13.2681, y: 43.338, z: -27.5849 },
+		    { x: 11.6315, y: 45.3222, z: -29.4976 },
+		    { x: 9.9725, y: 47.1297, z: -30.9947 },
+		    { x: 8.18144, y: 48.8158, z: -32.3406 },
+		    { x: 6.29655, y: 50.6231, z: -33.6949 },
+		    { x: 4.56312, y: 52.2425, z: -34.7317 },
+		    { x: 2.76385, y: 53.9091, z: -35.8971 },
+		    { x: 0.555895, y: 58.3243, z: -38.7002 },
+		    { x: -0.929823, y: 59.2197, z: -39.3087 },
+		    { x: -1.47794, y: 60.914, z: -40.1248 },
+		    { x: -4.54779, y: 63.5853, z: -43.1194 },
+		    { x: -6.01896, y: 64.1628, z: -43.8444 },
+		    { x: -8.06804, y: 64.7556, z: -43.8956 },
+		    { x: -10.0709, y: 65.3377, z: -44.2104 },
+		    { x: -11.1884, y: 65.8363, z: -44.4208 },
+		    { x: -13.3842, y: 66.8892, z: -45.1373 },
+		    { x: -15.5194, y: 67.713, z: -45.8841 },
+		    { x: -17.4575, y: 69.085, z: -46.6668 },
+		    { x: -18.5605, y: 69.703, z: -47.1517 },
+		    { x: -20.6373, y: 71.1108, z: -48.25 },
+		    { x: -22.3892, y: 72.3822, z: -49.1866 },
+		    { x: -23.2085, y: 73.1583, z: -49.4855 },
+		    { x: -24.9333, y: 73.9848, z: -50.2292 },
+		    { x: -26.4258, y: 74.8286, z: -50.8975 },
+		    { x: -27.1545, y: 75.3808, z: -51.2559 },
+		    { x: -28.7856, y: 76.813, z: -52.3584 },
+		    { x: -30.7553, y: 77.704, z: -53.3129 },
+		    { x: -31.9371, y: 77.9213, z: -53.7021 },
+		    { x: -33.1739, y: 78.009, z: -54.1446 },
+		    { x: -35.8559, y: 78.5608, z: -54.6517 },
+		    { x: -37.2329, y: 79.2864, z: -54.6918 },
+		    { x: -38.7037, y: 79.857, z: -55.1998 },
+		    { x: -40.4209, y: 81.1816, z: -55.6038 },
+		    { x: -41.9229, y: 84.2739, z: -55.4127 },
+		    { x: -43.2009, y: 84.6211, z: -55.4607 },
+		    { x: -44.5149, y: 84.8597, z: -55.8676 },
+		    { x: -46.1803, y: 87.7731, z: -56.5086 },
+		    { x: -47.1463, y: 87.2833, z: -57.2353 },
+		    { x: -49.1912, y: 89.6201, z: -58.2051 },
+		    { x: -50.0338, y: 91.2818, z: -59.2972 },
+		    { x: -50.5539, y: 90.6703, z: -60.1185 },
+		    { x: -50.8872, y: 90.5123, z: -60.5789 },
+		    { x: -51.0433, y: 90.5353, z: -61.1551 },
+		    { x: -51.2109, y: 90.5912, z: -61.3644 },
+		    { x: -51.2842, y: 90.8316, z: -61.3846 },
+		    { x: -51.3753, y: 90.9938, z: -61.3771 },
+		    { x: -51.4847, y: 91.0737, z: -61.3648 },
+		    { x: -51.6759, y: 90.9364, z: -61.3517 },
+		    { x: -51.6818, y: 90.952, z: -61.3127 },
+		    { x: -51.6389, y: 90.3421, z: -61.2818 },
+		    { x: -51.5493, y: 90.2819, z: -61.1741 },
+		    { x: -51.3372, y: 90.2178, z: -61.0274 },
+		    { x: -50.8463, y: 90.3974, z: -60.6749 },
+		    { x: -50.8463, y: 90.3974, z: -60.6749 },
+		    { x: -50.5881, y: 89, z: -60.8915 },
+		    { x: -50.5881, y: 89, z: -60.8915 },
+		    { x: -50.5881, y: 89, z: -60.8915 },
+		    { x: -49.8075, y: 87.1127, z: -61.703 },
+		    { x: -49.5945, y: 86.5781, z: -61.9598 },
+		    { x: -48.8514, y: 87.7866, z: -61.9179 },
+		    { x: -48.8514, y: 87.7866, z: -61.9179 },
+		    { x: -48.3121, y: 86.8545, z: -61.9627 },
+		    { x: -48.2048, y: 88.0795, z: -61.6152 },
+		    { x: -48.1737, y: 89.1062, z: -61.4791 },
+		    { x: -48.2017, y: 87.9752, z: -61.8712 },
+		    { x: -48.2017, y: 87.9752, z: -61.8712 },
+		    { x: -48.1094, y: 87.0158, z: -62.2042 },
+		    { x: -48.0376, y: 86.4249, z: -62.4599 },
+		    { x: -48.0896, y: 85.8689, z: -62.6122 },
+		    { x: -48.1072, y: 85.5348, z: -62.7236 },
+		    { x: -48.1466, y: 85.3426, z: -62.7953 },
+		    { x: -48.2214, y: 85.1987, z: -62.8477 },
+		    { x: -48.2214, y: 85.1987, z: -62.8477 },
+		    { x: -48.2972, y: 85.0929, z: -62.8878 },
+		    { x: -48.3704, y: 85.0646, z: -62.9274 },
+		    { x: -48.4455, y: 85.0676, z: -62.9545 },
+		    { x: -48.4625, y: 85.0066, z: -62.9903 },
+		    { x: -48.4625, y: 85.0066, z: -62.9903 },
+		    { x: -48.4495, y: 84.9227, z: -63.0261 },
+		    { x: -48.3898, y: 84.811, z: -63.0831 },
+		    { x: -48.2757, y: 84.6831, z: -63.1693 },
+		    { x: -48.1904, y: 84.5874, z: -63.2383 },
+		    { x: -48.1904, y: 84.5874, z: -63.2383 },
+		    { x: -48.1398, y: 84.5308, z: -63.2909 },
+		    { x: -48.1282, y: 84.5222, z: -63.3284 },
+		    { x: -48.1708, y: 84.5345, z: -63.3405 },
+		    { x: -48.1708, y: 84.5345, z: -63.3405 },
+		    { x: -48.2472, y: 84.5802, z: -63.3359 },
+		    { x: -48.3124, y: 84.6115, z: -63.3719 },
+		    { x: -48.3124, y: 84.6115, z: -63.3719 },
+		    { x: -48.4441, y: 84.7456, z: -63.3769 },
+		    { x: -48.596, y: 84.9497, z: -63.389 },
+		    { x: -48.7445, y: 85.2296, z: -63.3745 } ]
+    },
+    {
+	"name": "7",
+	"points": [ { x: -85.5092, y: 101.689, z: -46.6896 },
+		    { x: -85.5843, y: 101.601, z: -46.6291 },
+		    { x: -85.6104, y: 101.606, z: -46.5894 },
+		    { x: -85.613, y: 101.605, z: -46.5504 },
+		    { x: -85.613, y: 101.605, z: -46.5504 },
+		    { x: -85.5477, y: 101.616, z: -46.5203 },
+		    { x: -85.3564, y: 101.804, z: -46.3878 },
+		    { x: -85.2614, y: 101.763, z: -46.3498 },
+		    { x: -85.2614, y: 101.763, z: -46.3498 },
+		    { x: -85.1114, y: 101.791, z: -46.2952 },
+		    { x: -84.7404, y: 101.716, z: -46.3029 },
+		    { x: -84.7404, y: 101.716, z: -46.3029 },
+		    { x: -84.1632, y: 101.612, z: -46.1475 },
+		    { x: -83.4771, y: 101.233, z: -46.074 },
+		    { x: -83.4771, y: 101.233, z: -46.074 },
+		    { x: -82.6561, y: 100.726, z: -46.1619 },
+		    { x: -80.8458, y: 99.9634, z: -46.3545 },
+		    { x: -78.209, y: 97.92, z: -46.7581 },
+		    { x: -78.209, y: 97.92, z: -46.7581 },
+		    { x: -74.6166, y: 95.0982, z: -47.2643 },
+		    { x: -70.8752, y: 91.8398, z: -47.9321 },
+		    { x: -66.4481, y: 89.3535, z: -47.4765 },
+		    { x: -62.6843, y: 87.2825, z: -46.5117 },
+		    { x: -60.2101, y: 84.432, z: -46.1399 },
+		    { x: -53.8945, y: 79.0101, z: -44.817 },
+		    { x: -53.8945, y: 79.0101, z: -44.817 },
+		    { x: -53.8945, y: 79.0101, z: -44.817 },
+		    { x: -45.1307, y: 74.3465, z: -39.2238 },
+		    { x: -43.5933, y: 73.5215, z: -39.1131 },
+		    { x: -42.4082, y: 72.4528, z: -39.259 },
+		    { x: -38.6775, y: 68.3412, z: -40.0197 },
+		    { x: -36.711, y: 66.3831, z: -40.2752 },
+		    { x: -32.2597, y: 62.7827, z: -40.4872 },
+		    { x: -29.5525, y: 60.2261, z: -38.7675 },
+		    { x: -23.3518, y: 56.8335, z: -36.0886 },
+		    { x: -20.5919, y: 55.5249, z: -34.4765 },
+		    { x: -16.7489, y: 54.3894, z: -32.7713 },
+		    { x: -12.2496, y: 53.572, z: -31.1035 },
+		    { x: -10.2994, y: 53.2009, z: -30.5316 },
+		    { x: -6.74097, y: 52.3267, z: -29.5629 },
+		    { x: -2.76891, y: 51.3747, z: -28.3529 },
+		    { x: -0.94607, y: 50.7326, z: -27.9259 },
+		    { x: 2.20323, y: 49.4134, z: -27.0126 },
+		    { x: 5.42493, y: 47.2885, z: -26.1392 },
+		    { x: 7.09844, y: 46.1628, z: -25.7004 },
+		    { x: 10.5939, y: 45.1901, z: -25.0979 },
+		    { x: 12.5737, y: 45.6967, z: -25.4171 },
+		    { x: 16.371, y: 45.7385, z: -23.6568 },
+		    { x: 18.3687, y: 43.9894, z: -20.7278 },
+		    { x: 19.8687, y: 43.1773, z: -19.9422 },
+		    { x: 22.3252, y: 42.4557, z: -19.4199 },
+		    { x: 24.5245, y: 41.9144, z: -19.0156 },
+		    { x: 25.5892, y: 41.5773, z: -18.7597 },
+		    { x: 27.7894, y: 40.843, z: -18.414 },
+		    { x: 29.6044, y: 40.1729, z: -17.7244 },
+		    { x: 31.4074, y: 39.4177, z: -17.1158 },
+		    { x: 32.1182, y: 39.0339, z: -16.9339 },
+		    { x: 33.8512, y: 38.4467, z: -16.1548 },
+		    { x: 35.3647, y: 37.7997, z: -15.407 },
+		    { x: 36.0169, y: 37.4676, z: -15.0567 },
+		    { x: 37.2167, y: 36.7772, z: -14.2894 },
+		    { x: 38.3185, y: 36.0595, z: -13.5659 },
+		    { x: 38.7611, y: 35.8477, z: -13.1622 },
+		    { x: 39.5684, y: 35.5147, z: -12.3975 },
+		    { x: 39.9339, y: 35.4093, z: -12.0409 } ]
+    },
+    {
+	"name": "7'",
+	"points": [ { x: -45.4844, y: 18.6702, z: -11.7634 },
+		    { x: -19.4762, y: 44.4766, z: -35.43 },
+		    { x: -11.2632, y: 48.4675, z: -40.281 },
+		    { x: -10.2727, y: 49.2435, z: -41.0792 },
+		    { x: -10.1767, y: 49.9261, z: -41.5931 },
+		    { x: -10.6276, y: 51.8328, z: -43.1726 },
+		    { x: -12.0202, y: 53.6563, z: -45.1698 },
+		    { x: -13.4866, y: 54.0712, z: -46.226 },
+		    { x: -15.1733, y: 55.2635, z: -47.1833 },
+		    { x: -17.0326, y: 56.6542, z: -47.9644 },
+		    { x: -20.2092, y: 63.5948, z: -47.2679 },
+		    { x: -21.9987, y: 64.4533, z: -48.5808 },
+		    { x: -24.4624, y: 64.0086, z: -49.6895 },
+		    { x: -26.3907, y: 64.1536, z: -51.2955 },
+		    { x: -28.0642, y: 64.9007, z: -52.736 },
+		    { x: -29.6023, y: 65.7189, z: -53.5487 },
+		    { x: -31.4171, y: 67.1865, z: -54.1675 },
+		    { x: -33.5484, y: 68.7901, z: -55.1742 },
+		    { x: -34.6401, y: 69.85, z: -56.0004 },
+		    { x: -36.6412, y: 71.815, z: -56.8206 },
+		    { x: -37.6648, y: 73.141, z: -57.1331 },
+		    { x: -38.7745, y: 74.7272, z: -57.4229 },
+		    { x: -41.5184, y: 78.1921, z: -58.6092 },
+		    { x: -42.9943, y: 79.671, z: -58.9557 },
+		    { x: -46.5316, y: 83.0873, z: -60.3345 },
+		    { x: -48.6202, y: 84.4143, z: -61.1635 },
+		    { x: -52.9625, y: 88.3042, z: -63.7161 },
+		    { x: -54.8253, y: 89.5367, z: -65.4502 },
+		    { x: -56.6368, y: 90.9141, z: -66.4106 },
+		    { x: -58.7557, y: 92.2595, z: -67.351 },
+		    { x: -62.3616, y: 94.7022, z: -67.5421 },
+		    { x: -64.8275, y: 96.3097, z: -67.8077 },
+		    { x: -67.4916, y: 97.552, z: -68.1436 },
+		    { x: -72.0475, y: 102.04, z: -68.6897 },
+		    { x: -74.7465, y: 104.005, z: -69.5615 },
+		    { x: -77.2558, y: 104.771, z: -70.2313 },
+		    { x: -80.1491, y: 106.167, z: -70.3902 },
+		    { x: -81.748, y: 106.909, z: -70.5466 },
+		    { x: -83.2678, y: 107.526, z: -70.5668 },
+		    { x: -86.2251, y: 109.339, z: -71.3553 },
+		    { x: -88.4311, y: 110.625, z: -72.1031 },
+		    { x: -89.9199, y: 112.028, z: -72.9382 },
+		    { x: -92.1978, y: 114.291, z: -74.5544 },
+		    { x: -93.167, y: 115.216, z: -75.3724 },
+		    { x: -94.2934, y: 116.2, z: -76.3645 },
+		    { x: -96.3002, y: 118.341, z: -77.5275 },
+		    { x: -97.6107, y: 120.123, z: -77.6947 },
+		    { x: -97.9658, y: 117.962, z: -78.271 },
+		    { x: -99.153, y: 119.396, z: -78.5037 },
+		    { x: -101.735, y: 122.017, z: -80.0816 },
+		    { x: -102.923, y: 124.124, z: -80.3065 },
+		    { x: -104.256, y: 125.906, z: -80.4242 },
+		    { x: -105.419, y: 127.154, z: -80.4578 },
+		    { x: -109.295, y: 129.712, z: -81.9372 },
+		    { x: -111.302, y: 131.443, z: -82.6286 },
+		    { x: -112.542, y: 130.756, z: -83.2258 },
+		    { x: -113.867, y: 131.947, z: -83.6108 },
+		    { x: -114.395, y: 130.222, z: -83.8758 },
+		    { x: -115.968, y: 132.187, z: -83.6016 },
+		    { x: -117.156, y: 131.636, z: -83.4592 },
+		    { x: -118.383, y: 132.282, z: -83.3261 },
+		    { x: -123.043, y: 140.807, z: -83.4296 },
+		    { x: -124.062, y: 141.615, z: -83.7034 },
+		    { x: -125.261, y: 139.658, z: -84.8478 },
+		    { x: -128.879, y: 146.467, z: -84.3199 },
+		    { x: -130.848, y: 147.957, z: -84.7867 },
+		    { x: -132.193, y: 149.064, z: -85.172 },
+		    { x: -134.631, y: 151.731, z: -86.416 },
+		    { x: -138.496, y: 157.925, z: -87.1547 },
+		    { x: -142.659, y: 163.38, z: -88.3123 },
+		    { x: -142.352, y: 163.674, z: -88.2871 },
+		    { x: -144.987, y: 166.394, z: -87.1919 },
+		    { x: -148.446, y: 174.938, z: -88.2125 },
+		    { x: -150.951, y: 177.323, z: -87.1419 },
+		    { x: -153.41, y: 181.399, z: -86.1374 },
+		    { x: -156.141, y: 183.046, z: -86.281 },
+		    { x: -159.358, y: 187.038, z: -86.6562 },
+		    { x: -162.396, y: 192.41, z: -86.5119 },
+		    { x: -163.055, y: 193.956, z: -87.1814 },
+		    { x: -169.922, y: 196.038, z: -87.1707 },
+		    { x: -171.173, y: 195.826, z: -87.4904 },
+		    { x: -171.173, y: 195.826, z: -87.4904 },
+		    { x: -176.668, y: 197.026, z: -86.0811 },
+		    { x: -178.101, y: 198.394, z: -86.0696 },
+		    { x: -180.904, y: 198.531, z: -85.8284 },
+		    { x: -181.398, y: 197.391, z: -86.3227 },
+		    { x: -183.58, y: 196.232, z: -86.8484 },
+		    { x: -186.72, y: 199.599, z: -85.705 },
+		    { x: -187.498, y: 200.107, z: -85.5261 },
+		    { x: -189.992, y: 197.92, z: -85.6503 },
+		    { x: -190.813, y: 197.141, z: -85.7404 },
+		    { x: -193.454, y: 199.983, z: -86.1907 },
+		    { x: -194.765, y: 200.055, z: -86.2159 },
+		    { x: -195.61, y: 198.288, z: -87.0214 },
+		    { x: -195.571, y: 196.795, z: -88.2098 },
+		    { x: -195.491, y: 195.065, z: -89.5961 },
+		    { x: -195.45, y: 194.265, z: -90.2346 },
+		    { x: -195.377, y: 193.875, z: -90.6202 },
+		    { x: -194.739, y: 193.696, z: -91.8206 },
+		    { x: -194.125, y: 194.683, z: -93.3086 },
+		    { x: -194.352, y: 196.016, z: -93.3771 },
+		    { x: -194.267, y: 197.344, z: -93.3601 },
+		    { x: -193.981, y: 197.771, z: -93.3676 },
+		    { x: -193.382, y: 197.953, z: -93.4161 },
+		    { x: -192.475, y: 196.611, z: -94.0548 },
+		    { x: -191.758, y: 197.088, z: -93.7619 },
+		    { x: -190.946, y: 196.782, z: -94.5111 },
+		    { x: -190.365, y: 196.566, z: -95.0393 },
+		    { x: -189.925, y: 196.405, z: -95.4346 },
+		    { x: -189.066, y: 195.972, z: -96.4077 },
+		    { x: -188.416, y: 195.651, z: -97.13 },
+		    { x: -187.941, y: 195.42, z: -97.6521 },
+		    { x: -154.675, y: 145.604, z: -92.6532 },
+		    { x: -130.086, y: 128.052, z: -65.0264 },
+		    { x: -121.465, y: 123.514, z: -57.038 },
+		    { x: -117.413, y: 122.31, z: -54.2516 },
+		    { x: -116.407, y: 126.943, z: -44.3505 },
+		    { x: -116.483, y: 128.166, z: -42.1588 },
+		    { x: -116.492, y: 128.614, z: -41.5535 },
+		    { x: -116.534, y: 128.949, z: -41.3378 },
+		    { x: -116.553, y: 129.077, z: -41.2717 },
+		    { x: -116.668, y: 129.434, z: -41.3011 },
+		    { x: -116.957, y: 130.012, z: -41.9721 },
+		    { x: -117.03, y: 130.387, z: -42.581 },
+		    { x: -117.303, y: 131.124, z: -44.0181 },
+		    { x: -117.339, y: 131.374, z: -44.4529 },
+		    { x: -117.384, y: 131.697, z: -44.9723 },
+		    { x: -117.513, y: 132.018, z: -45.371 },
+		    { x: -117.498, y: 132.386, z: -46.0764 },
+		    { x: -117.395, y: 132.792, z: -46.7971 },
+		    { x: -117.44, y: 132.675, z: -48.8057 },
+		    { x: -117.475, y: 132.661, z: -49.3161 },
+		    { x: -117.622, y: 132.664, z: -49.8566 },
+		    { x: -117.703, y: 132.659, z: -50.0413 },
+		    { x: -117.81, y: 132.67, z: -50.2939 },
+		    { x: -117.976, y: 132.687, z: -50.4957 },
+		    { x: -118.186, y: 132.705, z: -50.6716 },
+		    { x: -118.403, y: 132.72, z: -50.8349 },
+		    { x: -118.564, y: 132.721, z: -51.0024 },
+		    { x: -119.094, y: 132.554, z: -51.6803 },
+		    { x: -119.267, y: 132.504, z: -51.8956 },
+		    { x: -119.398, y: 132.467, z: -52.0562 },
+		    { x: -119.5, y: 132.44, z: -52.1791 },
+		    { x: -119.585, y: 132.418, z: -52.2787 },
+		    { x: -119.801, y: 132.272, z: -52.5795 },
+		    { x: -120.195, y: 132.026, z: -52.9343 } ]
+    },
+    {
+	"name": "8",
+	"points": [ { x: -110.137, y: 158.729, z: -65.7815 },
+		    { x: -110.137, y: 158.729, z: -65.7815 },
+		    { x: -110.07, y: 158.767, z: -65.8038 },
+		    { x: -109.996, y: 158.72, z: -65.8401 },
+		    { x: -109.996, y: 158.72, z: -65.8401 },
+		    { x: -109.933, y: 158.706, z: -65.8703 },
+		    { x: -109.882, y: 158.697, z: -65.8958 },
+		    { x: -109.882, y: 158.697, z: -65.8958 },
+		    { x: -109.829, y: 158.735, z: -65.9105 },
+		    { x: -109.782, y: 158.768, z: -65.9249 },
+		    { x: -109.735, y: 158.795, z: -65.9406 },
+		    { x: -109.735, y: 158.795, z: -65.9406 },
+		    { x: -109.684, y: 158.81, z: -65.9598 },
+		    { x: -109.63, y: 158.81, z: -65.9806 },
+		    { x: -109.63, y: 158.81, z: -65.9806 },
+		    { x: -109.58, y: 158.806, z: -65.9995 },
+		    { x: -109.548, y: 158.768, z: -66.0057 },
+		    { x: -109.548, y: 158.768, z: -66.0057 },
+		    { x: -109.492, y: 158.719, z: -66.0189 },
+		    { x: -109.097, y: 158.37, z: -66.0252 },
+		    { x: -109.097, y: 158.37, z: -66.0252 },
+		    { x: -107.96, y: 157.446, z: -65.9552 },
+		    { x: -106.227, y: 155.871, z: -66.1945 },
+		    { x: -106.227, y: 155.871, z: -66.1945 },
+		    { x: -103.91, y: 153.184, z: -67.3414 },
+		    { x: -100.956, y: 150.877, z: -68.8985 },
+		    { x: -96.935, y: 147.471, z: -70.4326 },
+		    { x: -92.6211, y: 144.337, z: -71.4583 },
+		    { x: -88.6322, y: 141.318, z: -71.9439 },
+		    { x: -84.8233, y: 137.698, z: -71.76 },
+		    { x: -78.6788, y: 129.229, z: -72.5475 },
+		    { x: -75.8236, y: 125.081, z: -72.5843 },
+		    { x: -72.7372, y: 122.135, z: -72.585 },
+		    { x: -66.3057, y: 118.197, z: -70.1103 },
+		    { x: -64.4689, y: 115.397, z: -69.6647 },
+		    { x: -61.8954, y: 109.177, z: -68.9648 },
+		    { x: -59.952, y: 102.382, z: -68.1283 },
+		    { x: -58.8558, y: 98.7514, z: -67.3739 },
+		    { x: -57.1611, y: 90.7868, z: -65.7313 },
+		    { x: -56.2046, y: 87.6743, z: -64.8318 },
+		    { x: -56.1228, y: 80.9809, z: -60.0111 },
+		    { x: -56.0858, y: 75.9532, z: -56.497 },
+		    { x: -56.3221, y: 74.2976, z: -55.5775 },
+		    { x: -57.8056, y: 70.1901, z: -52.9357 },
+		    { x: -59.4895, y: 67.8551, z: -48.5105 },
+		    { x: -60.1755, y: 66.8793, z: -47.3725 },
+		    { x: -61.8379, y: 65.703, z: -46.0637 },
+		    { x: -63.8104, y: 64.5179, z: -44.8861 },
+		    { x: -64.8187, y: 64.4646, z: -44.4959 },
+		    { x: -68.333, y: 64.4986, z: -45.0687 },
+		    { x: -69.7234, y: 64.9291, z: -45.5087 },
+		    { x: -72.3508, y: 66.8934, z: -45.6309 },
+		    { x: -74.7897, y: 70.2821, z: -45.9373 },
+		    { x: -76.9119, y: 74.9254, z: -46.4092 },
+		    { x: -78.1163, y: 77.3759, z: -47.3308 },
+		    { x: -79.4589, y: 81.782, z: -50.6453 },
+		    { x: -80.2715, y: 84.9206, z: -51.3218 },
+		    { x: -79.8091, y: 88.8507, z: -53.6815 },
+		    { x: -79.1628, y: 90.8976, z: -54.9034 },
+		    { x: -78.0968, y: 92.6612, z: -56.7845 },
+		    { x: -75.9021, y: 95.9254, z: -60.7491 },
+		    { x: -74.5406, y: 96.361, z: -62.5745 },
+		    { x: -71.8366, y: 98.5076, z: -65.1621 },
+		    { x: -69.7922, y: 98.1084, z: -66.0964 },
+		    { x: -68.0923, y: 97.5958, z: -67.0655 },
+		    { x: -63.3491, y: 95.3635, z: -69.3049 },
+		    { x: -61.0539, y: 94.1631, z: -69.9921 },
+		    { x: -54.208, y: 93.0651, z: -70.7235 },
+		    { x: -50.6099, y: 93.0795, z: -70.4418 },
+		    { x: -47.0392, y: 91.5624, z: -69.5823 },
+		    { x: -40.4394, y: 89.3446, z: -67.7096 },
+		    { x: -37.4866, y: 87.9765, z: -66.6852 },
+		    { x: -32.0063, y: 83.5076, z: -64.3845 },
+		    { x: -29.7405, y: 80.9941, z: -62.8718 },
+		    { x: -25.2518, y: 75.9204, z: -56.4234 },
+		    { x: -23.8891, y: 74.7925, z: -55.2726 },
+		    { x: -20.2715, y: 70.7695, z: -54.0364 },
+		    { x: -18.5449, y: 67.8084, z: -53.3196 },
+		    { x: -15.673, y: 62.2287, z: -51.8641 },
+		    { x: -12.8455, y: 56.8153, z: -50.3629 },
+		    { x: -11.6101, y: 54.5515, z: -49.3936 },
+		    { x: -10.1368, y: 51.4581, z: -47.4491 },
+		    { x: -9.38578, y: 49.2634, z: -45.166 },
+		    { x: -9.37633, y: 48.4623, z: -44.229 },
+		    { x: -9.87954, y: 47.0459, z: -42.5532 },
+		    { x: -10.4529, y: 46.493, z: -41.8196 },
+		    { x: -12.2018, y: 45.8159, z: -40.4625 },
+		    { x: -16.3321, y: 46.4242, z: -37.5618 },
+		    { x: -18.2777, y: 46.5668, z: -36.2949 },
+		    { x: -22.2209, y: 45.8615, z: -34.4307 },
+		    { x: -22.2209, y: 45.8615, z: -34.4307 },
+		    { x: -30.7385, y: 46.0588, z: -32.1195 },
+		    { x: -32.6677, y: 46.036, z: -31.6617 },
+		    { x: -34.9459, y: 46.0251, z: -31.3501 },
+		    { x: -40.4685, y: 46.1559, z: -31.5212 },
+		    { x: -43.0579, y: 46.319, z: -31.7699 },
+		    { x: -45.3386, y: 46.3647, z: -32.0631 },
+		    { x: -49.1962, y: 47.2423, z: -33.0065 },
+		    { x: -50.9613, y: 48.025, z: -34.1055 },
+		    { x: -52.2491, y: 48.7294, z: -35.3856 },
+		    { x: -53.1352, y: 48.8224, z: -37.1897 },
+		    { x: -53.0131, y: 50.6081, z: -40.1996 },
+		    { x: -52.7172, y: 51.4971, z: -41.7853 },
+		    { x: -51.9638, y: 52.2469, z: -43.6214 },
+		    { x: -49.362, y: 54.21, z: -47.3533 },
+		    { x: -47.4477, y: 55.2154, z: -49.2324 },
+		    { x: -45.0316, y: 56.2554, z: -51.1767 },
+		    { x: -38.8722, y: 58.1091, z: -55.19 },
+		    { x: -36.1377, y: 58.3408, z: -56.6164 },
+		    { x: -32.8464, y: 58.4328, z: -58.2913 },
+		    { x: -29.141, y: 58.6653, z: -58.8113 },
+		    { x: -23.0082, y: 56.8429, z: -57.9974 },
+		    { x: -21.2003, y: 55.9621, z: -58.2983 },
+		    { x: -19.4147, y: 55.6585, z: -58.3482 },
+		    { x: -14.4773, y: 54.4318, z: -58.1901 },
+		    { x: -11.5156, y: 53.751, z: -57.3571 },
+		    { x: -5.65042, y: 51.6628, z: -54.7403 },
+		    { x: -3.45253, y: 50.2196, z: -53.6556 },
+		    { x: -0.992406, y: 48.9567, z: -51.6446 },
+		    { x: 2.76909, y: 46.6586, z: -49.1284 },
+		    { x: 4.36474, y: 45.5131, z: -47.4222 },
+		    { x: 7.01805, y: 43.1317, z: -44.4915 },
+		    { x: 8.14417, y: 41.8718, z: -42.9316 },
+		    { x: 10.3211, y: 39.101, z: -39.958 },
+		    { x: 11.4391, y: 37.9488, z: -38.5718 },
+		    { x: 14.7539, y: 36.6232, z: -36.8762 },
+		    { x: 16.9273, y: 36.5577, z: -35.3244 },
+		    { x: 17.2026, y: 36.3842, z: -34.8706 },
+		    { x: 18.1365, y: 35.7146, z: -33.6216 },
+		    { x: 18.6809, y: 35.3104, z: -33.1098 },
+		    { x: 19.5985, y: 34.3137, z: -32.3627 },
+		    { x: 20.2124, y: 33.671, z: -31.9883 },
+		    { x: 21.7532, y: 32.664, z: -31.0531 },
+		    { x: 22.7209, y: 32.0233, z: -30.4503 },
+		    { x: 24.7422, y: 30.6433, z: -29 },
+		    { x: 25.5731, y: 29.8914, z: -28.4806 },
+		    { x: 27.092, y: 28.522, z: -27.245 },
+		    { x: 27.877, y: 27.7435, z: -26.5113 },
+		    { x: 29.6576, y: 26.129, z: -24.2971 },
+		    { x: 30.1757, y: 25.5841, z: -23.4515 },
+		    { x: 30.5561, y: 24.9013, z: -22.7931 },
+		    { x: 31.1583, y: 24.1597, z: -21.0364 },
+		    { x: 31.439, y: 24.2215, z: -20.4663 },
+		    { x: 31.6318, y: 24.3428, z: -20.112 },
+		    { x: 31.9524, y: 24.5706, z: -19.5439 },
+		    { x: 32.081, y: 24.6634, z: -19.2282 },
+		    { x: 32.246, y: 24.7525, z: -18.9723 },
+		    { x: 32.5481, y: 24.8082, z: -18.6128 },
+		    { x: 32.792, y: 24.7526, z: -18.3743 },
+		    { x: 33.2835, y: 24.629, z: -17.9803 },
+		    { x: 33.5974, y: 24.5028, z: -17.6759 },
+		    { x: 33.8956, y: 23.5578, z: -17.0956 },
+		    { x: 33.9402, y: 23.0112, z: -16.8193 },
+		    { x: 32.5896, y: 19.9502, z: -15.3819 },
+		    { x: 30.6987, y: 16.2575, z: -13.5129 },
+		    { x: 30.6987, y: 16.2575, z: -13.5129 },
+		    { x: 29.1965, y: 13.1497, z: -12.0598 } ]
+    },
+    {
+	"name": "8'",
+	"points": [ { x: 2.67005, y: 43.5962, z: -51.0118 },
+		    { x: 1.16318, y: 43.7258, z: -49.8527 },
+		    { x: 0.372104, y: 43.6476, z: -49.3151 },
+		    { x: -2.32822, y: 43.759, z: -47.933 },
+		    { x: -4.64429, y: 43.7949, z: -46.9022 },
+		    { x: -7.04994, y: 43.7394, z: -46.0425 },
+		    { x: -9.39222, y: 43.4049, z: -45.1837 },
+		    { x: -11.5508, y: 42.6719, z: -44.3497 },
+		    { x: -11.5508, y: 42.6719, z: -44.3497 },
+		    { x: -13.3242, y: 41.7025, z: -43.6752 },
+		    { x: -14.7976, y: 40.8353, z: -43.0821 },
+		    { x: -16.0201, y: 39.917, z: -42.5301 },
+		    { x: -16.9109, y: 38.8912, z: -42.0764 },
+		    { x: -17.5211, y: 37.8514, z: -41.7502 },
+		    { x: -18.5067, y: 35.8517, z: -41.136 },
+		    { x: -18.9725, y: 34.5459, z: -40.792 },
+		    { x: -19.4477, y: 33.5251, z: -40.5131 },
+		    { x: -19.9508, y: 32.4347, z: -40.1673 },
+		    { x: -20.611, y: 31.3727, z: -39.7016 },
+		    { x: -21.2203, y: 30.381, z: -39.1343 },
+		    { x: -21.8252, y: 29.3406, z: -38.5294 },
+		    { x: -22.4786, y: 28.2664, z: -37.7302 },
+		    { x: -23.3124, y: 27.0131, z: -36.5286 },
+		    { x: -23.9366, y: 25.6973, z: -35.2958 },
+		    { x: -24.3785, y: 24.5286, z: -34.1092 },
+		    { x: -24.6863, y: 23.5436, z: -33.3167 },
+		    { x: -22.5742, y: 16.3823, z: -30.5788 },
+		    { x: -22.0658, y: 15.8103, z: -29.7039 },
+		    { x: -21.2175, y: 15.1475, z: -29.295 },
+		    { x: -19.1258, y: 13.4519, z: -28.5987 },
+		    { x: -18.3743, y: 12.9025, z: -28.2457 },
+		    { x: -18.3743, y: 12.9025, z: -28.2457 },
+		    { x: -15.3392, y: 11.5846, z: -28.0308 },
+		    { x: -14.299, y: 11.2131, z: -28.1495 },
+		    { x: -10.8763, y: 10.0475, z: -28.8198 },
+		    { x: -9.13981, y: 9.74701, z: -29.34 },
+		    { x: -7.30792, y: 9.54312, z: -30.0648 },
+		    { x: -3.73283, y: 9.59188, z: -31.9105 },
+		    { x: -2.00643, y: 9.8681, z: -33.0458 },
+		    { x: 1.26699, y: 10.7368, z: -35.3057 },
+		    { x: 2.77297, y: 11.3041, z: -36.6417 },
+		    { x: 3.93917, y: 12.0416, z: -37.8787 },
+		    { x: 5.49408, y: 13.5965, z: -40.1197 },
+		    { x: 5.37066, y: 16.8555, z: -41.771 },
+		    { x: 6.01684, y: 20.4441, z: -44.3097 },
+		    { x: 9.3482, y: 20.0121, z: -45.5458 },
+		    { x: 11.1662, y: 20.5513, z: -46.2516 },
+		    { x: 11.6822, y: 21.5745, z: -46.6869 },
+		    { x: 13.7345, y: 25.2596, z: -48.851 },
+		    { x: 12.3083, y: 27.1084, z: -50.4451 },
+		    { x: 10.9071, y: 29.0759, z: -52.0883 },
+		    { x: 8.36038, y: 33.4018, z: -54.6525 },
+		    { x: 7.05657, y: 35.6033, z: -55.9294 },
+		    { x: 5.27483, y: 38.0963, z: -57.3615 },
+		    { x: 1.40102, y: 42.4439, z: -60.97 },
+		    { x: -0.616151, y: 44.4736, z: -62.0632 },
+		    { x: -4.31557, y: 48.1763, z: -63.9626 },
+		    { x: -5.29251, y: 50.1816, z: -63.9867 },
+		    { x: -3.60772, y: 51.935, z: -64.4282 },
+		    { x: -1.54214, y: 59.073, z: -64.6045 },
+		    { x: -1.42942, y: 61.7325, z: -65.0954 },
+		    { x: -3.09238, y: 67.261, z: -66.5498 },
+		    { x: -5.9046, y: 69.8031, z: -67.2821 },
+		    { x: -9.24068, y: 70.7617, z: -67.8724 },
+		    { x: -13.6336, y: 71.393, z: -68.3437 },
+		    { x: -18.1816, y: 72.5949, z: -68.9616 },
+		    { x: -20.3689, y: 73.1907, z: -69.1595 },
+		    { x: -24.5135, y: 74.3944, z: -69.4625 },
+		    { x: -28.3926, y: 75.3915, z: -69.5098 },
+		    { x: -32.0513, y: 76.1926, z: -69.1429 },
+		    { x: -35.5612, y: 76.4024, z: -68.3162 },
+		    { x: -38.8498, y: 76.4278, z: -67.2415 },
+		    { x: -42.0703, y: 76.329, z: -66.0107 },
+		    { x: -43.5921, y: 76.1634, z: -65.3444 },
+		    { x: -46.5827, y: 75.3542, z: -63.8187 },
+		    { x: -49.3742, y: 74.4768, z: -62.1427 },
+		    { x: -52.408, y: 72.7471, z: -60.0102 },
+		    { x: -53.9098, y: 71.3799, z: -58.5266 },
+		    { x: -55.2664, y: 69.8525, z: -57.0019 },
+		    { x: -57.6584, y: 66.6306, z: -53.7946 },
+		    { x: -58.5423, y: 64.979, z: -52.4047 },
+		    { x: -59.1066, y: 62.9959, z: -50.8101 },
+		    { x: -59.4488, y: 61.1362, z: -49.2302 },
+		    { x: -58.9012, y: 56.9141, z: -46.2045 },
+		    { x: -58.2292, y: 55.0038, z: -44.7555 },
+		    { x: -57.3074, y: 53.2192, z: -43.5894 },
+		    { x: -54.5004, y: 49.4499, z: -41.7421 },
+		    { x: -52.6808, y: 47.6977, z: -41.0563 },
+		    { x: -50.7559, y: 46.034, z: -40.5668 },
+		    { x: -46.3888, y: 43.1655, z: -40.1446 },
+		    { x: -43.9686, y: 41.734, z: -40.1951 },
+		    { x: -41.5033, y: 40.6334, z: -40.5091 },
+		    { x: -38.9778, y: 39.7591, z: -41.0944 },
+		    { x: -33.6093, y: 38.1162, z: -42.9225 },
+		    { x: -30.9382, y: 37.6024, z: -44.0794 },
+		    { x: -28.3944, y: 37.404, z: -45.239 },
+		    { x: -24.0081, y: 38.1848, z: -47.7462 },
+		    { x: -22.0516, y: 39.0868, z: -48.9798 },
+		    { x: -20.3046, y: 40.4523, z: -50.5288 },
+		    { x: -17.3405, y: 43.2448, z: -54.1431 },
+		    { x: -16.2962, y: 44.6241, z: -55.6697 },
+		    { x: -15.5006, y: 46.1901, z: -57.1769 },
+		    { x: -14.627, y: 49.7783, z: -60.7786 },
+		    { x: -14.6934, y: 51.3949, z: -62.526 },
+		    { x: -15.0653, y: 53.2629, z: -64.114 },
+		    { x: -16.7845, y: 57.1839, z: -67.6978 },
+		    { x: -18.113, y: 59.258, z: -69.7887 },
+		    { x: -19.6878, y: 60.8992, z: -71.1795 },
+		    { x: -23.898, y: 65.1298, z: -74.9738 },
+		    { x: -26.293, y: 66.9236, z: -76.4778 },
+		    { x: -28.9388, y: 68.5862, z: -77.9991 },
+		    { x: -34.5694, y: 71.4208, z: -80.5779 },
+		    { x: -37.566, y: 73.0612, z: -82.0329 },
+		    { x: -43.8834, y: 76.0567, z: -84.3512 },
+		    { x: -46.9966, y: 77.1359, z: -85.877 },
+		    { x: -50.1539, y: 78.9293, z: -86.8252 },
+		    { x: -57.1518, y: 83.9155, z: -89.7908 },
+		    { x: -60.0972, y: 85.9148, z: -90.4002 },
+		    { x: -66.7708, y: 89.4124, z: -91.5354 },
+		    { x: -66.7708, y: 89.4124, z: -91.5354 },
+		    { x: -76.8859, y: 97.8406, z: -92.4591 },
+		    { x: -82.3491, y: 104.05, z: -94.1703 },
+		    { x: -84.5047, y: 106.03, z: -94.4432 },
+		    { x: -88.2648, y: 109.992, z: -93.8572 },
+		    { x: -93.0041, y: 115.93, z: -94.1689 },
+		    { x: -94.7341, y: 116.893, z: -94.3747 },
+		    { x: -98.9683, y: 121.909, z: -94.9907 },
+		    { x: -101.024, y: 123.375, z: -95.6377 },
+		    { x: -104.541, y: 130.408, z: -94.854 },
+		    { x: -108.171, y: 134.876, z: -95.7901 },
+		    { x: -109.76, y: 136.711, z: -95.5767 },
+		    { x: -112.195, y: 140.158, z: -94.0863 },
+		    { x: -114.877, y: 145.912, z: -91.4797 },
+		    { x: -115.828, y: 147.418, z: -91.1829 },
+		    { x: -117.609, y: 152.48, z: -89.5809 },
+		    { x: -119.087, y: 156.795, z: -88.1184 },
+		    { x: -119.762, y: 158.754, z: -87.2883 },
+		    { x: -120.36, y: 160.865, z: -87.4988 },
+		    { x: -120.696, y: 161.152, z: -87.9679 },
+		    { x: -120.948, y: 163.24, z: -89.1686 },
+		    { x: -120.954, y: 163.832, z: -89.3909 },
+		    { x: -121.133, y: 165.08, z: -89.015 },
+		    { x: -121.039, y: 165.141, z: -89.0375 },
+		    { x: -120.684, y: 165.54, z: -89.2433 },
+		    { x: -120.428, y: 165.529, z: -89.635 },
+		    { x: -120.087, y: 165.424, z: -90.2381 },
+		    { x: -119.682, y: 165.231, z: -90.9782 },
+		    { x: -119.387, y: 165.533, z: -90.9874 },
+		    { x: -119.033, y: 165.806, z: -91.0398 },
+		    { x: -118.322, y: 165.213, z: -91.5371 },
+		    { x: -117.953, y: 164.785, z: -91.7673 },
+		    { x: -117.286, y: 163.827, z: -91.9073 },
+		    { x: -116.828, y: 163.165, z: -91.9585 },
+		    { x: -116.396, y: 162.296, z: -92.0096 },
+		    { x: -116.01, y: 161.603, z: -91.8898 },
+		    { x: -116.01, y: 161.603, z: -91.8898 },
+		    { x: -115.68, y: 160.555, z: -91.9883 },
+		    { x: -115.405, y: 159.853, z: -91.9789 },
+		    { x: -115.251, y: 159.261, z: -91.9923 },
+		    { x: -115.13, y: 158.951, z: -91.9231 },
+		    { x: -115.014, y: 158.866, z: -91.7398 },
+		    { x: -114.899, y: 158.82, z: -91.5637 },
+		    { x: -114.899, y: 158.82, z: -91.5637 },
+		    { x: -114.773, y: 158.717, z: -91.4373 },
+		    { x: -114.664, y: 158.63, z: -91.3555 },
+		    { x: -114.561, y: 158.442, z: -91.3019 },
+		    { x: -114.493, y: 158.327, z: -91.2126 },
+		    { x: -114.493, y: 158.327, z: -91.2126 },
+		    { x: -114.434, y: 158.243, z: -91.1124 },
+		    { x: -114.374, y: 158.264, z: -90.9371 },
+		    { x: -114.334, y: 158.172, z: -90.8022 },
+		    { x: -114.311, y: 158.035, z: -90.7202 },
+		    { x: -114.311, y: 158.035, z: -90.7202 },
+		    { x: -114.292, y: 157.814, z: -90.6876 },
+		    { x: -114.306, y: 157.627, z: -90.6354 },
+		    { x: -114.336, y: 157.558, z: -90.5688 },
+		    { x: -114.336, y: 157.558, z: -90.5688 },
+		    { x: -114.414, y: 157.66, z: -90.3455 },
+		    { x: -114.521, y: 157.609, z: -90.1646 },
+		    { x: -114.521, y: 157.609, z: -90.1646 },
+		    { x: -114.62, y: 157.833, z: -90.1041 },
+		    { x: -114.7, y: 157.936, z: -90.0688 },
+		    { x: -114.7, y: 157.936, z: -90.0688 },
+		    { x: -114.802, y: 157.545, z: -89.8496 },
+		    { x: -114.676, y: 156.957, z: -89.6945 },
+		    { x: -114.632, y: 156.379, z: -89.6005 },
+		    { x: -114.632, y: 156.379, z: -89.6005 } ]
+    },
+    {    "name": "9",
+	 "points": [ { x: -98.1576, y: 156.263, z: -30.7876 },
+		     { x: -98.1576, y: 156.263, z: -30.7876 },
+		     { x: -98.2102, y: 156.251, z: -30.7323 },
+		     { x: -98.2785, y: 156.256, z: -30.6766 },
+		     { x: -98.4204, y: 156.22, z: -30.5722 },
+		     { x: -98.4204, y: 156.22, z: -30.5722 },
+		     { x: -98.6216, y: 156.137, z: -30.4205 },
+		     { x: -98.8951, y: 156.052, z: -30.2336 },
+		     { x: -98.8951, y: 156.052, z: -30.2336 },
+		     { x: -99.1785, y: 155.9, z: -30.0237 },
+		     { x: -99.3738, y: 155.675, z: -29.8869 },
+		     { x: -99.3738, y: 155.675, z: -29.8869 },
+		     { x: -99.588, y: 155.297, z: -29.6822 },
+		     { x: -99.7327, y: 154.662, z: -29.3824 },
+		     { x: -99.7327, y: 154.662, z: -29.3824 },
+		     { x: -99.6954, y: 153.762, z: -29.1828 },
+		     { x: -99.601, y: 152.859, z: -29.2546 },
+		     { x: -99.601, y: 152.859, z: -29.2546 },
+		     { x: -98.9273, y: 151.121, z: -29.3937 },
+		     { x: -97.5926, y: 148.845, z: -29.9586 },
+		     { x: -95.3944, y: 145.255, z: -30.6058 },
+		     { x: -92.8772, y: 141.416, z: -31.0791 },
+		     { x: -90.0364, y: 137.509, z: -31.4691 },
+		     { x: -87.3817, y: 134.435, z: -31.8251 },
+		     { x: -84.7066, y: 131.067, z: -32.2495 },
+		     { x: -83.5798, y: 129.373, z: -32.9587 },
+		     { x: -81.4612, y: 126.629, z: -34.5237 },
+		     { x: -78.2202, y: 123.141, z: -35.2042 },
+		     { x: -72.0665, y: 117.697, z: -36.1661 },
+		     { x: -69.5069, y: 115.198, z: -36.5579 },
+		     { x: -67.2172, y: 112.881, z: -37.2621 },
+		     { x: -61.8053, y: 107.777, z: -38.191 },
+		     { x: -59.4747, y: 105.446, z: -38.5403 },
+		     { x: -54.229, y: 99.8649, z: -39.5221 },
+		     { x: -52.5192, y: 98.0206, z: -40.1145 },
+		     { x: -48.5268, y: 93.9992, z: -40.9315 },
+		     { x: -46.8143, y: 92.053, z: -40.8796 },
+		     { x: -43.7078, y: 88.528, z: -40.8212 },
+		     { x: -42.4142, y: 86.9731, z: -40.7601 },
+		     { x: -40.2739, y: 84.5053, z: -40.8642 },
+		     { x: -39.3594, y: 83.4476, z: -40.9471 },
+		     { x: -37.9384, y: 82.1277, z: -41.0582 },
+		     { x: -37.9384, y: 82.1277, z: -41.0582 },
+		     { x: -36.6687, y: 79.6591, z: -41.2459 },
+		     { x: -36.255, y: 78.8021, z: -41.2718 },
+		     { x: -35.8919, y: 77.9769, z: -41.3574 },
+		     { x: -35.3115, y: 76.4875, z: -41.4514 },
+		     { x: -35.0169, y: 76.0157, z: -41.5389 },
+		     { x: -34.3513, y: 75.5918, z: -41.8152 },
+		     { x: -34.0198, y: 75.6626, z: -42.0272 },
+		     { x: -33.6624, y: 75.8477, z: -42.2794 },
+		     { x: -33.1223, y: 76.166, z: -42.6169 },
+		     { x: -33.1223, y: 76.166, z: -42.6169 },
+		     { x: -29.7664, y: 78.1042, z: -44.5791 },
+		     { x: -28.7298, y: 78.8055, z: -45.1305 },
+		     { x: -27.2616, y: 79.5621, z: -45.9629 },
+		     { x: -25.5257, y: 80.3185, z: -46.8102 },
+		     { x: -23.8016, y: 80.8574, z: -47.5965 },
+		     { x: -22.0002, y: 81.5008, z: -48.4453 },
+		     { x: -20.312, y: 81.5459, z: -49.0114 },
+		     { x: -19.3061, y: 81.194, z: -49.3068 },
+		     { x: -18.6728, y: 79.8506, z: -49.2751 },
+		     { x: -18.884, y: 73.989, z: -48.478 },
+		     { x: -20.3269, y: 70, z: -47.589 },
+		     { x: -22.048, y: 66.0028, z: -46.2825 },
+		     { x: -24.2221, y: 62.459, z: -45.1285 },
+		     { x: -26.4233, y: 59.7899, z: -43.5371 },
+		     { x: -28.656, y: 57.0106, z: -42.1796 },
+		     { x: -32.913, y: 53.2958, z: -37.9568 },
+		     { x: -34.8863, y: 51.9446, z: -36.0409 },
+		     { x: -36.8593, y: 50.8212, z: -34.1697 },
+		     { x: -39.7397, y: 49.0565, z: -31.1083 },
+		     { x: -40.8422, y: 48.3924, z: -29.939 },
+		     { x: -41.6566, y: 47.8983, z: -29.0359 },
+		     { x: -42.407, y: 47.3023, z: -28.0064 },
+		     { x: -42.5114, y: 47.2116, z: -27.7449 },
+		     { x: -42.4728, y: 47.2021, z: -27.6037 },
+		     { x: -41.5348, y: 47.8189, z: -27.5997 },
+		     { x: -40.1294, y: 48.728, z: -28.1668 },
+		     { x: -37.9733, y: 50.115, z: -29.3758 },
+		     { x: -31.5936, y: 53.1615, z: -33.0866 },
+		     { x: -28.1009, y: 54.475, z: -35.5093 },
+		     { x: -24.2845, y: 55.7188, z: -37.4284 },
+		     { x: -15.6345, y: 57.3398, z: -40.798 },
+		     { x: -11.3721, y: 58.6898, z: -42.2037 },
+		     { x: -7.96, y: 59.5133, z: -43.8721 },
+		     { x: -1.13531, y: 61.4376, z: -46.0815 },
+		     { x: 1.77215, y: 62.307, z: -46.9007 },
+		     { x: 4.31656, y: 62.7666, z: -47.482 },
+		     { x: 8.33275, y: 63.5507, z: -48.9533 },
+		     { x: 9.72436, y: 63.869, z: -49.6312 },
+		     { x: 12.4831, y: 65.4351, z: -50.9692 },
+		     { x: 13.4065, y: 65.6398, z: -51.3519 },
+		     { x: 14.8839, y: 65.5848, z: -51.7192 },
+		     { x: 15.2863, y: 65.3531, z: -51.7574 },
+		     { x: 15.5854, y: 63.9265, z: -51.3609 },
+		     { x: 15.1662, y: 62.665, z: -50.8788 },
+		     { x: 14.3045, y: 61.0042, z: -49.6729 },
+		     { x: 10.7089, y: 56.3581, z: -47.3654 },
+		     { x: 8.56073, y: 54.1351, z: -45.8866 },
+		     { x: 5.98149, y: 51.2532, z: -44.0359 },
+		     { x: 1.26807, y: 45.4656, z: -39.0526 },
+		     { x: -0.93229, y: 43.0198, z: -36.2112 },
+		     { x: -3.39035, y: 40.6884, z: -33.8074 },
+		     { x: -8.17232, y: 36.8811, z: -30.3159 },
+		     { x: -10.9247, y: 35.3282, z: -29.089 },
+		     { x: -15.1522, y: 32.1912, z: -26.9863 },
+		     { x: -16.7958, y: 30.7755, z: -25.7471 },
+		     { x: -18.1091, y: 29.7738, z: -24.8237 },
+		     { x: -19.9251, y: 28.7854, z: -23.2131 },
+		     { x: -20.9737, y: 28.2006, z: -22.0571 },
+		     { x: -21.7301, y: 27.4459, z: -20.623 },
+		     { x: -21.9034, y: 27.0486, z: -20.1542 },
+		     { x: -20.9183, y: 27.0081, z: -20.3273 },
+		     { x: -20.213, y: 26.9434, z: -20.5505 },
+		     { x: -17.7738, y: 27.0809, z: -22.0928 },
+		     { x: -15.9223, y: 27.3531, z: -23.3216 },
+		     { x: -11.6004, y: 27.7162, z: -25.6958 },
+		     { x: -8.67875, y: 27.7945, z: -27.1755 },
+		     { x: -5.70895, y: 27.956, z: -28.5262 },
+		     { x: 1.09099, y: 28.0477, z: -31.6543 },
+		     { x: 4.62736, y: 27.9494, z: -33.1244 },
+		     { x: 12.3158, y: 28.5127, z: -36.3368 },
+		     { x: 15.8748, y: 28.7314, z: -37.6242 },
+		     { x: 19.6161, y: 29.4954, z: -38.8448 },
+		     { x: 26.1662, y: 30.9386, z: -40.5974 },
+		     { x: 31.1106, y: 32.1617, z: -42.4993 },
+		     { x: 33.2858, y: 32.5855, z: -43.3996 },
+		     { x: 37.5303, y: 33.0417, z: -44.7466 },
+		     { x: 39.2937, y: 32.9819, z: -45.115 },
+		     { x: 41.1807, y: 32.9521, z: -45.4906 },
+		     { x: 43.0027, y: 33.6972, z: -46.1365 },
+		     { x: 43.8036, y: 34.0115, z: -46.3944 },
+		     { x: 45.0592, y: 34.4746, z: -46.7892 },
+		     { x: 45.5933, y: 34.6833, z: -46.9165 },
+		     { x: 45.6138, y: 34.7101, z: -46.8542 },
+		     { x: 44.8876, y: 34.5056, z: -46.2608 },
+		     { x: 43.7854, y: 34.1465, z: -45.5818 },
+		     { x: 40.2081, y: 33.1613, z: -43.6558 },
+		     { x: 37.8841, y: 32.3647, z: -42.4841 },
+		     { x: 33.5721, y: 30.7683, z: -40.5621 },
+		     { x: 30.6223, y: 30.0849, z: -38.9982 },
+		     { x: 27.4484, y: 29.1726, z: -37.4047 },
+		     { x: 22.8023, y: 28.0173, z: -35.6584 },
+		     { x: 22.8023, y: 28.0173, z: -35.6584 },
+		     { x: 22.8023, y: 28.0173, z: -35.6584 },
+		     { x: 13.743, y: 25.8681, z: -31.4227 },
+		     { x: 9.80265, y: 25.365, z: -29.6604 },
+		     { x: 8.15287, y: 24.934, z: -28.8707 },
+		     { x: 6.675, y: 24.4033, z: -28.1675 },
+		     { x: 4.88088, y: 23.6689, z: -26.9692 },
+		     { x: 3.08683, y: 23.0482, z: -25.7013 },
+		     { x: 2.67364, y: 22.9674, z: -25.3922 },
+		     { x: 2.29491, y: 22.8441, z: -25.116 },
+		     { x: 1.80578, y: 22.4976, z: -24.8194 },
+		     { x: 1.80578, y: 22.4976, z: -24.8194 },
+		     { x: 1.61097, y: 22.1633, z: -24.6155 },
+		     { x: 1.42562, y: 21.8706, z: -24.5169 },
+		     { x: 1.25952, y: 21.6746, z: -24.4686 },
+		     { x: 1.14327, y: 21.4959, z: -24.4302 },
+		     { x: 1.20376, y: 21.4409, z: -24.4475 },
+		     { x: 1.28331, y: 21.4346, z: -24.4372 },
+		     { x: 1.3492, y: 21.4114, z: -24.4593 },
+		     { x: 1.33704, y: 21.4728, z: -24.4556 },
+		     { x: 1.32371, y: 21.5458, z: -24.4392 },
+		     { x: 1.41313, y: 21.6334, z: -24.4555 },
+		     { x: 1.56633, y: 21.9869, z: -24.5233 },
+		     { x: 1.62597, y: 22.0958, z: -24.5531 },
+		     { x: 1.67713, y: 22.3339, z: -24.5731 },
+		     { x: 1.55798, y: 22.5238, z: -24.5487 },
+		     { x: 1.55937, y: 22.6578, z: -24.5371 },
+		     { x: 1.5639, y: 22.9143, z: -24.4542 },
+		     { x: 1.30123, y: 23.0548, z: -24.3595 },
+		     { x: 1.40128, y: 23.1833, z: -24.681 },
+		     { x: 1.59232, y: 23.1524, z: -24.826 },
+		     { x: 1.37798, y: 23.0562, z: -24.7281 },
+		     { x: 1.37798, y: 23.0562, z: -24.7281 },
+		     { x: 1.25743, y: 22.8576, z: -24.8514 },
+		     { x: 0.930129, y: 22.2901, z: -25.1175 },
+		     { x: -0.0379273, y: 22.4171, z: -24.8651 },
+		     { x: -0.104971, y: 22.4401, z: -24.8551 },
+		     { x: -0.386874, y: 22.3806, z: -24.8298 },
+		     { x: -0.585744, y: 22.3475, z: -24.8151 },
+		     { x: -0.734717, y: 22.3419, z: -24.7988 },
+		     { x: -0.846237, y: 22.3539, z: -24.7864 },
+		     { x: -0.980846, y: 22.3462, z: -24.7809 },
+		     { x: -1.08484, y: 22.3445, z: -24.7905 },
+		     { x: -1.18148, y: 22.3591, z: -24.7853 },
+		     { x: -1.23843, y: 22.3958, z: -24.7796 },
+		     { x: -1.58153, y: 22.6268, z: -24.7417 },
+		     { x: -1.7019, y: 22.7542, z: -24.7336 },
+		     { x: -1.8002, y: 22.8668, z: -24.7258 },
+		     { x: -1.85851, y: 22.9625, z: -24.7304 },
+		     { x: -1.90272, y: 23.0346, z: -24.7384 },
+		     { x: -1.94082, y: 23.1031, z: -24.745 },
+		     { x: -1.97411, y: 23.1665, z: -24.7516 },
+		     { x: -2.10276, y: 23.3413, z: -24.7809 },
+		     { x: -2.10276, y: 23.3413, z: -24.7809 },
+		     { x: -2.20274, y: 23.4816, z: -24.8025 },
+		     { x: -2.26602, y: 23.5955, z: -24.8179 },
+		     { x: -2.27961, y: 23.6679, z: -24.8375 },
+		     { x: -2.29746, y: 23.7414, z: -24.8605 },
+		     { x: -2.29746, y: 23.7414, z: -24.8605 },
+		     { x: -2.31095, y: 23.7964, z: -24.879 },
+		     { x: -2.31241, y: 23.8323, z: -24.8924 },
+		     { x: -2.31241, y: 23.8323, z: -24.8924 },
+		     { x: -2.31044, y: 23.8952, z: -24.9284 },
+		     { x: -2.29993, y: 23.9734, z: -24.9806 },
+		     { x: -2.29094, y: 24.0539, z: -25.0279 },
+		     { x: -2.29094, y: 24.0539, z: -25.0279 },
+		     { x: -2.28876, y: 24.1053, z: -25.0529 } ]
+    },
+    {
+	"name": "9'",
+	"points": [ { x: 15.8609, y: 81.6491, z: -21.6201 },
+		    { x: 14.6082, y: 84.0639, z: -22.3515 },
+		    { x: 14.6082, y: 84.0639, z: -22.3515 },
+		    { x: 11.6966, y: 89.7062, z: -22.2966 },
+		    { x: 9.16407, y: 92.2939, z: -22.3372 },
+		    { x: 7.22688, y: 94.4999, z: -22.5118 },
+		    { x: 5.51602, y: 96.1625, z: -22.6312 },
+		    { x: 3.65239, y: 96.6477, z: -22.8386 },
+		    { x: 0.682271, y: 97.3661, z: -22.1978 },
+		    { x: -0.497227, y: 97.5332, z: -22.0752 },
+		    { x: -1.52561, y: 97.7456, z: -22.1065 },
+		    { x: -2.5608, y: 97.7686, z: -22.1362 },
+		    { x: -3.84744, y: 96.7871, z: -22.7912 },
+		    { x: -4.88907, y: 95.5432, z: -23.4043 },
+		    { x: -6.14934, y: 92.4724, z: -23.5585 },
+		    { x: -6.51942, y: 90.708, z: -23.5743 },
+		    { x: -6.74719, y: 88.8338, z: -23.6349 },
+		    { x: -6.99445, y: 86.989, z: -23.5958 },
+		    { x: -7.33264, y: 84.7404, z: -23.6777 },
+		    { x: -8.15771, y: 78.7005, z: -23.5787 },
+		    { x: -8.88486, y: 75.4335, z: -23.7088 },
+		    { x: -10.2308, y: 71.3441, z: -22.4731 },
+		    { x: -11.4758, y: 67.8055, z: -20.719 },
+		    { x: -12.877, y: 64.4047, z: -19.0891 },
+		    { x: -14.7325, y: 58.1425, z: -16.6877 },
+		    { x: -15.5118, y: 55.3026, z: -15.9459 },
+		    { x: -16.2124, y: 52.2359, z: -14.8409 },
+		    { x: -17.0481, y: 45.4914, z: -12.6125 },
+		    { x: -17.4393, y: 42.5255, z: -12.1347 },
+		    { x: -17.6758, y: 39.9036, z: -11.6149 },
+		    { x: -18.624, y: 36.415, z: -10.8013 },
+		    { x: -18.959, y: 35.1806, z: -10.5168 },
+		    { x: -19.7149, y: 32.6751, z: -9.94048 },
+		    { x: -20.1434, y: 31.6954, z: -9.77677 },
+		    { x: -20.9844, y: 30.3403, z: -9.61163 },
+		    { x: -21.3799, y: 29.9929, z: -9.61432 },
+		    { x: -22.4462, y: 28.1239, z: -9.72646 },
+		    { x: -23.4079, y: 29.4685, z: -10.2 },
+		    { x: -24.0215, y: 30.3787, z: -10.5058 },
+		    { x: -28.2974, y: 36.4573, z: -12.339 },
+		    { x: -29.1078, y: 38.1214, z: -13.022 },
+		    { x: -30.1249, y: 40.4325, z: -13.5749 },
+		    { x: -32.5739, y: 46.4355, z: -15.4948 },
+		    { x: -33.8163, y: 49.6214, z: -16.6448 },
+		    { x: -36.4916, y: 54.7409, z: -19.4826 },
+		    { x: -38.0645, y: 56.094, z: -20.4 },
+		    { x: -39.5823, y: 60.1965, z: -21.3907 },
+		    { x: -43.9558, y: 66.5911, z: -24.8913 },
+		    { x: -48.4576, y: 74.4239, z: -25.0323 },
+		    { x: -48.4576, y: 74.4239, z: -25.0323 },
+		    { x: -53.8619, y: 86.0352, z: -23.0533 },
+		    { x: -57.0859, y: 91.303, z: -22.1858 },
+		    { x: -59.0887, y: 94.38, z: -22.5769 },
+		    { x: -62.7065, y: 99.0311, z: -22.2997 },
+		    { x: -66.0293, y: 103.825, z: -20.5349 },
+		    { x: -67.9405, y: 106.099, z: -20.1531 },
+		    { x: -71.0909, y: 110.315, z: -18.9826 },
+		    { x: -74.2618, y: 113.872, z: -17.3479 },
+		    { x: -75.9286, y: 114.829, z: -16.5403 },
+		    { x: -79.8672, y: 115.946, z: -14.9397 },
+		    { x: -84.0757, y: 116.759, z: -13.0072 },
+		    { x: -88.438, y: 116.282, z: -11.1329 },
+		    { x: -90.426, y: 115.714, z: -10.0863 },
+		    { x: -94.4462, y: 113.917, z: -7.5216 },
+		    { x: -96.5983, y: 112.908, z: -6.01724 },
+		    { x: -100.662, y: 109.608, z: -3.23544 },
+		    { x: -102.267, y: 107.95, z: -1.99468 },
+		    { x: -106.087, y: 102.899, z: 0.457086 },
+		    { x: -107.843, y: 100.403, z: 2.03576 },
+		    { x: -111.017, y: 95.3718, z: 4.97163 },
+		    { x: -112.336, y: 93.0575, z: 5.65364 },
+		    { x: -114.254, y: 88.8228, z: 7.96409 },
+		    { x: -114.683, y: 86.7357, z: 9.23428 },
+		    { x: -115.419, y: 82.4597, z: 11.6485 },
+		    { x: -115.652, y: 80.5071, z: 12.6244 },
+		    { x: -115.871, y: 77.6245, z: 13.9593 },
+		    { x: -115.941, y: 76.1164, z: 14.6211 },
+		    { x: -115.563, y: 73.3073, z: 14.7428 },
+		    { x: -115.501, y: 72.3311, z: 14.8719 },
+		    { x: -115.412, y: 71.6012, z: 14.4684 },
+		    { x: -115.05, y: 70.3787, z: 13.7377 },
+		    { x: -114.771, y: 69.71, z: 12.9368 },
+		    { x: -114.304, y: 69.084, z: 11.2327 },
+		    { x: -114.199, y: 69.3309, z: 9.94139 },
+		    { x: -114.318, y: 69.8126, z: 8.46045 },
+		    { x: -114.702, y: 70.4246, z: 7.22911 },
+		    { x: -115.944, y: 73.1939, z: 4.85294 },
+		    { x: -117.506, y: 75.1141, z: 3.26472 },
+		    { x: -118.755, y: 76.0897, z: 1.55374 },
+		    { x: -122.612, y: 78.4203, z: -0.983332 },
+		    { x: -124.323, y: 80.1952, z: -1.76648 },
+		    { x: -129.888, y: 84.1037, z: -3.56421 },
+		    { x: -135.563, y: 91.0054, z: -3.1957 },
+		    { x: -138.084, y: 94.0259, z: -2.52862 },
+		    { x: -145.164, y: 102.757, z: -0.630813 },
+		    { x: -148.674, y: 107.097, z: -0.540719 },
+		    { x: -155.872, y: 115.539, z: 2.26248 },
+		    { x: -163.015, y: 124.031, z: 5.37148 },
+		    { x: -169.445, y: 131.133, z: 8.07722 },
+		    { x: -173.174, y: 136.136, z: 10.2575 },
+		    { x: -179.644, y: 143.874, z: 13.62 },
+		    { x: -185.785, y: 152.151, z: 16.8348 },
+		    { x: -191.9, y: 159.915, z: 21.242 },
+		    { x: -197.258, y: 168.151, z: 26.0585 },
+		    { x: -197.258, y: 168.151, z: 26.0585 },
+		    { x: -206.551, y: 179.152, z: 29.9113 },
+		    { x: -211.272, y: 184.889, z: 34.2854 },
+		    { x: -214.961, y: 190.862, z: 39.1056 },
+		    { x: -218.334, y: 196.069, z: 43.7456 },
+		    { x: -222.079, y: 201.335, z: 47.9257 },
+		    { x: -225.332, y: 206.216, z: 51.5258 },
+		    { x: -227.146, y: 208.042, z: 53.7734 },
+		    { x: -229.805, y: 211.005, z: 56.9887 },
+		    { x: -231.925, y: 213.582, z: 59.465 },
+		    { x: -234.514, y: 216.518, z: 62.0505 },
+		    { x: -235.712, y: 218.487, z: 64.7773 },
+		    { x: -237.386, y: 220.705, z: 68.4681 },
+		    { x: -239.132, y: 223.596, z: 73.2217 },
+		    { x: -239.472, y: 224.32, z: 74.0113 },
+		    { x: -240.675, y: 227.208, z: 75.9683 },
+		    { x: -241.387, y: 228.176, z: 77.1219 },
+		    { x: -241.474, y: 229.863, z: 80.0157 },
+		    { x: -241.464, y: 229.229, z: 78.909 },
+		    { x: -241.209, y: 229.32, z: 78.3717 },
+		    { x: -240.876, y: 230.26, z: 78.2537 },
+		    { x: -240.493, y: 229.824, z: 78.1468 },
+		    { x: -239.356, y: 228.513, z: 78.021 },
+		    { x: -236.775, y: 226.423, z: 76.5859 },
+		    { x: -236.277, y: 225.749, z: 76.0534 },
+		    { x: -233.598, y: 223.428, z: 75.6694 },
+		    { x: -232.166, y: 222.114, z: 74.927 },
+		    { x: -231.452, y: 221.317, z: 74.1995 },
+		    { x: -229.803, y: 219.798, z: 72.7098 },
+		    { x: -229.031, y: 219.318, z: 71.9077 },
+		    { x: -228.436, y: 219.121, z: 71.2794 },
+		    { x: -227.61, y: 218.763, z: 70.6796 },
+		    { x: -227.304, y: 218.723, z: 70.4155 },
+		    { x: -225.473, y: 218.319, z: 69.9495 },
+		    { x: -224.086, y: 217.553, z: 69.9123 },
+		    { x: -223.091, y: 217.287, z: 69.556 },
+		    { x: -222.495, y: 216.041, z: 67.8796 },
+		    { x: -221.998, y: 215.672, z: 68.2327 },
+		    { x: -221.852, y: 215.319, z: 67.917 },
+		    { x: -221.07, y: 214.895, z: 67.0774 },
+		    { x: -221.296, y: 215.083, z: 67.4412 },
+		    { x: -221.677, y: 215.414, z: 67.9645 },
+		    { x: -221.762, y: 215.466, z: 67.9391 },
+		    { x: -222.074, y: 215.703, z: 67.8424 },
+		    { x: -222.041, y: 215.485, z: 67.7834 },
+		    { x: -222.038, y: 215.284, z: 67.6959 },
+		    { x: -222.057, y: 215.344, z: 67.6279 },
+		    { x: -222.548, y: 215.9, z: 67.9165 },
+		    { x: -222.936, y: 216.084, z: 68.055 },
+		    { x: -223.349, y: 216.172, z: 68.163 },
+		    { x: -223.75, y: 216.083, z: 68.2913 },
+		    { x: -224.103, y: 216.323, z: 68.5655 },
+		    { x: -224.321, y: 216.076, z: 68.6045 },
+		    { x: -224.543, y: 215.982, z: 68.7417 },
+		    { x: -224.746, y: 215.765, z: 68.9667 },
+		    { x: -224.675, y: 214.936, z: 69.0595 },
+		    { x: -224.675, y: 214.936, z: 69.0595 },
+		    { x: -224.815, y: 214.784, z: 69.6092 },
+		    { x: -225.414, y: 215.402, z: 70.3848 },
+		    { x: -225.47, y: 215.345, z: 70.7708 },
+		    { x: -225.584, y: 215.117, z: 71.0076 },
+		    { x: -225.584, y: 215.117, z: 71.0076 },
+		    { x: -225.705, y: 214.923, z: 71.1575 },
+		    { x: -225.827, y: 214.979, z: 71.3137 },
+		    { x: -225.829, y: 214.545, z: 71.0619 },
+		    { x: -225.781, y: 214.266, z: 70.9197 },
+		    { x: -225.8, y: 214.225, z: 70.8377 } ]
+    },
+    {
+	"name": "10",
+	"points": [ { x: -104.631, y: 130.145, z: -27.0498 },
+		    { x: -104.631, y: 130.145, z: -27.0498 },
+		    { x: -104.822, y: 129.853, z: -26.8789 },
+		    { x: -105.02, y: 129.594, z: -26.7229 },
+		    { x: -105.02, y: 129.594, z: -26.7229 },
+		    { x: -105.198, y: 129.409, z: -26.593 },
+		    { x: -105.381, y: 129.135, z: -26.4525 },
+		    { x: -105.381, y: 129.135, z: -26.4525 },
+		    { x: -105.741, y: 128.586, z: -26.2892 },
+		    { x: -106.085, y: 128.182, z: -26.1394 },
+		    { x: -106.085, y: 128.182, z: -26.1394 },
+		    { x: -106.501, y: 127.643, z: -26.0669 },
+		    { x: -106.929, y: 127.134, z: -25.984 },
+		    { x: -106.929, y: 127.134, z: -25.984 },
+		    { x: -107.495, y: 126.554, z: -25.8917 },
+		    { x: -108.004, y: 126.126, z: -25.7779 },
+		    { x: -108.004, y: 126.126, z: -25.7779 },
+		    { x: -108.427, y: 125.855, z: -25.4131 },
+		    { x: -108.744, y: 125.624, z: -25.1355 },
+		    { x: -109.021, y: 125.412, z: -24.912 },
+		    { x: -109.021, y: 125.412, z: -24.912 },
+		    { x: -109.291, y: 125.2, z: -24.7066 },
+		    { x: -109.431, y: 125.009, z: -24.6621 },
+		    { x: -109.409, y: 124.933, z: -24.7453 },
+		    { x: -109.409, y: 124.933, z: -24.7453 },
+		    { x: -109.05, y: 124.917, z: -25.1693 },
+		    { x: -107.801, y: 125.449, z: -26.4278 },
+		    { x: -105.477, y: 126.725, z: -28.1487 },
+		    { x: -105.477, y: 126.725, z: -28.1487 },
+		    { x: -102.554, y: 128.118, z: -30.1299 },
+		    { x: -99.3417, y: 129.768, z: -31.8662 },
+		    { x: -96.5146, y: 131.197, z: -33.2804 },
+		    { x: -94.0554, y: 132.32, z: -34.2263 },
+		    { x: -91.8276, y: 133.317, z: -35.0698 },
+		    { x: -89.7716, y: 135.155, z: -35.6314 },
+		    { x: -87.5945, y: 136.518, z: -36.1847 },
+		    { x: -84.1898, y: 137.061, z: -36.8505 },
+		    { x: -82.3454, y: 137.406, z: -36.7756 },
+		    { x: -80.2759, y: 137.43, z: -36.7431 },
+		    { x: -80.2759, y: 137.43, z: -36.7431 },
+		    { x: -80.2759, y: 137.43, z: -36.7431 },
+		    { x: -75.144, y: 135.027, z: -36.2386 },
+		    { x: -72.9164, y: 132.949, z: -35.6098 },
+		    { x: -71.6029, y: 130.875, z: -35.0545 },
+		    { x: -70.8701, y: 128.806, z: -34.5323 },
+		    { x: -70.723, y: 126.33, z: -33.9739 },
+		    { x: -71.0675, y: 123.555, z: -32.7203 },
+		    { x: -71.4777, y: 121.069, z: -31.6235 },
+		    { x: -73.1731, y: 114.533, z: -27.5916 },
+		    { x: -74.1753, y: 111.402, z: -25.961 },
+		    { x: -75.3675, y: 108.585, z: -23.4004 },
+		    { x: -77.6422, y: 103.335, z: -20.2693 },
+		    { x: -78.4451, y: 100.968, z: -18.6791 },
+		    { x: -78.816, y: 97.6593, z: -17.4972 },
+		    { x: -80.1015, y: 92.8493, z: -11.6153 },
+		    { x: -81.8002, y: 91.8834, z: -8.9278 },
+		    { x: -82.9278, y: 91.0085, z: -7.77491 },
+		    { x: -83.8687, y: 90.0611, z: -6.95702 },
+		    { x: -84.1256, y: 89.6484, z: -6.65721 },
+		    { x: -84.2904, y: 89.0437, z: -6.29171 },
+		    { x: -84.1752, y: 88.9798, z: -6.17649 },
+		    { x: -83.9158, y: 88.9257, z: -6.43471 },
+		    { x: -81.7919, y: 88.6158, z: -9.60911 },
+		    { x: -79.7141, y: 88.7283, z: -11.3925 },
+		    { x: -76.4871, y: 89.7436, z: -14.3392 },
+		    { x: -74.2341, y: 90.8096, z: -15.7993 },
+		    { x: -71.2911, y: 92.0329, z: -17.6431 },
+		    { x: -65.4483, y: 94.6133, z: -21.5508 },
+		    { x: -62.7231, y: 95.543, z: -24.1183 },
+		    { x: -57.148, y: 98.6226, z: -27.8712 },
+		    { x: -54.376, y: 99.887, z: -29.139 },
+		    { x: -51.6523, y: 100.958, z: -30.3742 },
+		    { x: -45.6881, y: 103.41, z: -32.2628 },
+		    { x: -42.5263, y: 104.398, z: -32.8879 },
+		    { x: -39.2639, y: 105.568, z: -32.4648 },
+		    { x: -34.2527, y: 106.931, z: -33.0328 },
+		    { x: -31.8098, y: 107.131, z: -33.3154 },
+		    { x: -26.9677, y: 107.941, z: -33.2879 },
+		    { x: -24.8775, y: 108.329, z: -33.3797 },
+		    { x: -21.5636, y: 107.617, z: -33.1107 },
+		    { x: -20.3663, y: 106.956, z: -32.9033 },
+		    { x: -19.1641, y: 105.98, z: -32.3806 },
+		    { x: -17.1734, y: 103.527, z: -31.1693 },
+		    { x: -16.6146, y: 101.397, z: -30.8316 },
+		    { x: -16.2745, y: 97.3258, z: -28.916 },
+		    { x: -16.4157, y: 95.3877, z: -27.9671 },
+		    { x: -16.8809, y: 92.3492, z: -26.4185 },
+		    { x: -18.5701, y: 86.957, z: -23.4249 },
+		    { x: -19.9365, y: 84.884, z: -22.0749 },
+		    { x: -24.2064, y: 79.1336, z: -17.7317 },
+		    { x: -27.0872, y: 76.0418, z: -14.5118 },
+		    { x: -29.9212, y: 72.5981, z: -12.0203 },
+		    { x: -34.6721, y: 68.6996, z: -8.00406 },
+		    { x: -36.8295, y: 66.8501, z: -5.68368 },
+		    { x: -38.7751, y: 64.9565, z: -3.67008 },
+		    { x: -42.2167, y: 61.5265, z: -0.651471 },
+		    { x: -43.6292, y: 60.5601, z: 0.851294 },
+		    { x: -46.596, y: 57.742, z: 2.48222 },
+		    { x: -47.6545, y: 57.6159, z: 2.95785 },
+		    { x: -47.76, y: 55.3905, z: 4.38272 },
+		    { x: -47.7581, y: 55.0424, z: 4.72372 },
+		    { x: -47.7462, y: 54.8913, z: 4.77096 },
+		    { x: -47.7102, y: 54.9877, z: 4.5766 },
+		    { x: -47.0789, y: 55.5626, z: 3.66394 },
+		    { x: -46.2542, y: 56.4321, z: 2.91952 },
+		    { x: -42.979, y: 59.2764, z: 0.223338 },
+		    { x: -41.2555, y: 60.487, z: -1.20372 },
+		    { x: -38.8986, y: 61.9955, z: -2.97135 },
+		    { x: -33.1751, y: 65.1595, z: -6.22199 },
+		    { x: -30.5333, y: 66.7569, z: -7.44849 },
+		    { x: -27.9541, y: 67.6918, z: -9.38099 },
+		    { x: -22.7824, y: 69.3378, z: -12.622 },
+		    { x: -19.325, y: 70.3372, z: -14.6888 },
+		    { x: -15.9072, y: 71.4762, z: -16.2408 },
+		    { x: -8.55343, y: 73.9335, z: -18.9894 },
+		    { x: -5.35746, y: 74.3855, z: -20.1945 },
+		    { x: -2.9409, y: 74.6275, z: -20.9572 },
+		    { x: 2.35476, y: 74.7243, z: -22.3656 },
+		    { x: 5.09041, y: 74.8359, z: -22.9367 },
+		    { x: 9.82589, y: 75.6117, z: -23.4312 },
+		    { x: 11.362, y: 75.3314, z: -23.4607 },
+		    { x: 13.7102, y: 73.0886, z: -22.939 },
+		    { x: 14.7588, y: 71.9496, z: -22.5532 },
+		    { x: 15.6741, y: 70.9344, z: -22.0907 },
+		    { x: 17.2523, y: 67.8422, z: -20.5612 },
+		    { x: 17.9918, y: 65.7947, z: -19.1475 },
+		    { x: 18.4424, y: 61.4581, z: -15.9311 },
+		    { x: 18.1662, y: 59.444, z: -14.4414 },
+		    { x: 17.5524, y: 56.8468, z: -12.2087 },
+		    { x: 14.5514, y: 51.2238, z: -8.5285 },
+		    { x: 12.8461, y: 48.9561, z: -6.03881 },
+		    { x: 11.1919, y: 47.1758, z: -3.89069 },
+		    { x: 11.1919, y: 47.1758, z: -3.89069 },
+		    { x: 4.03463, y: 40.1927, z: 5.13436 },
+		    { x: 1.69874, y: 37.1249, z: 6.81615 },
+		    { x: 1.04433, y: 36.7606, z: 7.80225 },
+		    { x: -0.131761, y: 36.2636, z: 9.12235 },
+		    { x: -1.44693, y: 35.1858, z: 10.451 },
+		    { x: -2.49782, y: 34.4333, z: 11.3138 },
+		    { x: -4.77166, y: 32.6616, z: 12.5154 },
+		    { x: -5.45515, y: 32.1655, z: 12.7789 },
+		    { x: -6.12954, y: 31.5581, z: 12.8044 },
+		    { x: -6.58562, y: 31.532, z: 12.9679 },
+		    { x: -6.76101, y: 31.9273, z: 12.8996 },
+		    { x: -6.81279, y: 32.699, z: 12.2613 },
+		    { x: -6.32982, y: 34.8511, z: 10.7426 },
+		    { x: -5.94181, y: 36.1175, z: 9.81272 },
+		    { x: -5.29804, y: 38.1688, z: 7.82135 },
+		    { x: -4.27862, y: 41.0122, z: 5.68344 },
+		    { x: -2.40153, y: 47.2194, z: 0.60397 },
+		    { x: -1.05991, y: 49.9743, z: -2.1641 },
+		    { x: 0.308091, y: 52.1572, z: -3.88758 },
+		    { x: 1.92739, y: 54.2277, z: -5.69496 },
+		    { x: 5.79012, y: 58.2106, z: -8.95206 },
+		    { x: 8.13742, y: 59.7133, z: -10.0373 },
+		    { x: 11.1079, y: 61.3908, z: -11.6223 },
+		    { x: 16.5024, y: 63.7445, z: -14.3424 },
+		    { x: 19.1131, y: 64.4731, z: -15.5587 },
+		    { x: 21.744, y: 65.1727, z: -16.7962 },
+		    { x: 26.7986, y: 65.7615, z: -18.9314 },
+		    { x: 28.9586, y: 65.5807, z: -19.7326 },
+		    { x: 31.0672, y: 65.3352, z: -20.5576 },
+		    { x: 35.4376, y: 64.6318, z: -21.61 },
+		    { x: 37.516, y: 64.1492, z: -22.2722 },
+		    { x: 41.5837, y: 63.2536, z: -23.2773 },
+		    { x: 43.6434, y: 62.5176, z: -23.5408 },
+		    { x: 47.4493, y: 60.2604, z: -22.2365 },
+		    { x: 48.9675, y: 58.9852, z: -21.653 },
+		    { x: 50.6221, y: 57.5021, z: -20.4278 },
+		    { x: 53.1888, y: 54.1126, z: -18.2672 },
+		    { x: 53.852, y: 52.0553, z: -17.7731 },
+		    { x: 54.0253, y: 49.409, z: -18.0454 },
+		    { x: 54.1098, y: 45.8439, z: -17.0919 },
+		    { x: 54.1295, y: 44.631, z: -16.1481 },
+		    { x: 53.2592, y: 41.5479, z: -13.4859 },
+		    { x: 52.8247, y: 40.1332, z: -12.2343 },
+		    { x: 50.7234, y: 36.6963, z: -9.07683 },
+		    { x: 49.0662, y: 34.2232, z: -7.483 },
+		    { x: 47.8723, y: 32.2778, z: -6.22663 },
+		    { x: 47.4168, y: 31.3538, z: -5.6747 } ]
+    },
+    {
+	"name": "10'",
+	"points":  [ { x: 75.5829, y: 46.8849, z: -9.6385 },
+		     { x: 75.049, y: 49.8741, z: -10.4127 },
+		     { x: 75.049, y: 49.8741, z: -10.4127 },
+		     { x: 72.6658, y: 54.218, z: -12.8235 },
+		     { x: 71.0515, y: 57.3175, z: -14.055 },
+		     { x: 69.8047, y: 61.0207, z: -15.1001 },
+		     { x: 68.4234, y: 63.5232, z: -16.0642 },
+		     { x: 67.0551, y: 65.247, z: -16.4357 },
+		     { x: 64.4298, y: 68.0183, z: -16.3557 },
+		     { x: 63.3386, y: 68.7034, z: -16.2712 },
+		     { x: 62.0964, y: 69.3497, z: -15.9768 },
+		     { x: 60.7562, y: 69.5056, z: -15.6838 },
+		     { x: 59.2537, y: 69.6686, z: -15.1036 },
+		     { x: 56.4243, y: 69.2906, z: -14.3194 },
+		     { x: 55.1109, y: 68.8523, z: -13.8441 },
+		     { x: 53.4594, y: 68.115, z: -13.2669 },
+		     { x: 51.5405, y: 67.1986, z: -12.4177 },
+		     { x: 48.0545, y: 63.747, z: -10.8411 },
+		     { x: 46.2835, y: 61.7544, z: -9.80327 },
+		     { x: 44.5727, y: 60.1227, z: -9.22111 },
+		     { x: 41.2553, y: 55.5346, z: -6.74248 },
+		     { x: 39.9797, y: 53.2543, z: -6.34031 },
+		     { x: 38.8667, y: 51.4141, z: -5.59381 },
+		     { x: 37.5291, y: 49.1152, z: -4.97496 },
+		     { x: 34.659, y: 44.0569, z: -4.10434 },
+		     { x: 33.6758, y: 41.9217, z: -3.82666 },
+		     { x: 32.5017, y: 39.3442, z: -3.77326 },
+		     { x: 30.8832, y: 35.0424, z: -3.2263 },
+		     { x: 30.0578, y: 33.004, z: -3.12995 },
+		     { x: 28.393, y: 28.6766, z: -2.55073 },
+		     { x: 27.9988, y: 27.5511, z: -2.51663 },
+		     { x: 27.7447, y: 26.5678, z: -2.6146 },
+		     { x: 27.6939, y: 25.4044, z: -3.04762 },
+		     { x: 27.6556, y: 24.9309, z: -3.20871 },
+		     { x: 27.7044, y: 24.3404, z: -3.44524 },
+		     { x: 27.6961, y: 24.2828, z: -3.5493 },
+		     { x: 27.7186, y: 24.3371, z: -3.73251 },
+		     { x: 28.0274, y: 24.7332, z: -4.66788 },
+		     { x: 28.3202, y: 24.9149, z: -5.4052 },
+		     { x: 28.7473, y: 25.5729, z: -6.8172 },
+		     { x: 29.4585, y: 26.6208, z: -6.10969 },
+		     { x: 30.019, y: 27.8162, z: -7.24852 },
+		     { x: 31.5392, y: 31.6298, z: -9.07608 },
+		     { x: 32.349, y: 34.5412, z: -10.3309 },
+		     { x: 33.9282, y: 40.8859, z: -13.0282 },
+		     { x: 34.3355, y: 43.9622, z: -14.1972 },
+		     { x: 34.5859, y: 47.3564, z: -15.5855 },
+		     { x: 34.748, y: 50.7094, z: -16.6993 },
+		     { x: 34.6405, y: 58.1818, z: -19.415 },
+		     { x: 34.2092, y: 60.8999, z: -20.1736 },
+		     { x: 33.0846, y: 64.254, z: -21.7979 },
+		     { x: 31.0494, y: 71.44, z: -23.5256 },
+		     { x: 31.0494, y: 71.44, z: -23.5256 },
+		     { x: 31.0494, y: 71.44, z: -23.5256 },
+		     { x: 25.5774, y: 84.247, z: -25.062 },
+		     { x: 24.1012, y: 86.7066, z: -25.2991 },
+		     { x: 20.3794, y: 92.5691, z: -27.0084 },
+		     { x: 18.513, y: 95.5511, z: -27.8268 },
+		     { x: 14.8465, y: 100.367, z: -29.5397 },
+		     { x: 12.8826, y: 102.698, z: -30.0979 },
+		     { x: 8.56335, y: 107.091, z: -30.7103 },
+		     { x: 6.13134, y: 108.654, z: -30.8043 },
+		     { x: 0.756285, y: 112.931, z: -30.3018 },
+		     { x: -4.36218, y: 116.574, z: -29.7795 },
+		     { x: -6.67316, y: 117.886, z: -29.2701 },
+		     { x: -11.7061, y: 119.977, z: -27.9341 },
+		     { x: -14.0314, y: 120.422, z: -27.5118 },
+		     { x: -18.494, y: 120.05, z: -25.6769 },
+		     { x: -20.9475, y: 119.818, z: -24.9394 },
+		     { x: -26.1573, y: 118.629, z: -23.0169 },
+		     { x: -28.2701, y: 117.128, z: -21.6777 },
+		     { x: -32.4892, y: 114.496, z: -19.7792 },
+		     { x: -34.862, y: 112.6, z: -18.214 },
+		     { x: -39.2192, y: 108.574, z: -16.0001 },
+		     { x: -42.3483, y: 104.712, z: -13.4761 },
+		     { x: -43.743, y: 102.132, z: -12.34 },
+		     { x: -45.8559, y: 97.4067, z: -10.8544 },
+		     { x: -46.6424, y: 94.8818, z: -10.2001 },
+		     { x: -47.7727, y: 90.5638, z: -9.38774 },
+		     { x: -48.4296, y: 84.5297, z: -8.18495 },
+		     { x: -49.0505, y: 80.054, z: -7.57584 },
+		     { x: -49.0601, y: 79.4098, z: -7.36249 },
+		     { x: -49.2848, y: 76.037, z: -7.27785 },
+		     { x: -49.0855, y: 76.3086, z: -6.9346 },
+		     { x: -48.8088, y: 75.8534, z: -6.87428 },
+		     { x: -48.0491, y: 75.6117, z: -7.07774 },
+		     { x: -47.458, y: 75.8257, z: -7.07283 },
+		     { x: -46.0908, y: 76.7758, z: -9.32718 },
+		     { x: -45.2414, y: 76.6572, z: -10.1533 },
+		     { x: -43.7157, y: 76.708, z: -11.2413 },
+		     { x: -42.7658, y: 76.8959, z: -11.7732 },
+		     { x: -40.2476, y: 77.7185, z: -12.842 },
+		     { x: -38.7134, y: 78.607, z: -13.4613 },
+		     { x: -35.2419, y: 81.2633, z: -14.6779 },
+		     { x: -33.6021, y: 83.4665, z: -15.478 },
+		     { x: -30.8781, y: 86.8558, z: -16.5492 },
+		     { x: -29.5227, y: 88.719, z: -17.6571 },
+		     { x: -28.2183, y: 91.1759, z: -18.9744 },
+		     { x: -25.9458, y: 96.9608, z: -20.9036 },
+		     { x: -25.007, y: 99.4911, z: -21.9662 },
+		     { x: -23.6096, y: 103.99, z: -23.023 },
+		     { x: -23.2451, y: 106.822, z: -23.9074 },
+		     { x: -22.8744, y: 109.155, z: -24.3921 },
+		     { x: -22.6353, y: 113.736, z: -25.4089 },
+		     { x: -22.6554, y: 116.581, z: -25.9267 },
+		     { x: -23.2559, y: 121.058, z: -26.6612 },
+		     { x: -23.667, y: 122.95, z: -26.7717 },
+		     { x: -25.4364, y: 126.272, z: -27.2055 },
+		     { x: -26.6549, y: 127.83, z: -27.4352 },
+		     { x: -29.2574, y: 131.429, z: -27.4666 },
+		     { x: -30.7159, y: 133.071, z: -27.282 },
+		     { x: -34.3718, y: 136.553, z: -27.0171 },
+		     { x: -36.4263, y: 137.676, z: -26.9812 },
+		     { x: -38.5596, y: 138.506, z: -26.7658 },
+		     { x: -43.139, y: 139.296, z: -26.1022 },
+		     { x: -46.4828, y: 139.906, z: -25.9997 },
+		     { x: -52.1666, y: 139.83, z: -25.3136 },
+		     { x: -55.3383, y: 140.154, z: -25.2446 },
+		     { x: -57.3941, y: 139.224, z: -23.7834 },
+		     { x: -63.3741, y: 138.883, z: -22.8558 },
+		     { x: -66.5487, y: 137.911, z: -21.8568 },
+		     { x: -72.8416, y: 134.357, z: -20.3243 },
+		     { x: -75.338, y: 133.645, z: -19.4407 },
+		     { x: -80.4209, y: 129.311, z: -17.5103 },
+		     { x: -83.107, y: 128.719, z: -17.1605 },
+		     { x: -85.9332, y: 127.721, z: -16.0604 },
+		     { x: -89.9888, y: 121.286, z: -13.319 },
+		     { x: -92.0787, y: 119.264, z: -12.6836 },
+		     { x: -94.4365, y: 116.178, z: -11.3151 },
+		     { x: -96.697, y: 112.649, z: -10.0054 },
+		     { x: -97.4413, y: 110.901, z: -8.9817 },
+		     { x: -98.6193, y: 108.144, z: -7.60184 },
+		     { x: -99.0188, y: 106.764, z: -7.35472 },
+		     { x: -99.7317, y: 103.166, z: -7.65108 },
+		     { x: -99.6883, y: 101.786, z: -7.95411 },
+		     { x: -99.9423, y: 97.742, z: -8.61171 },
+		     { x: -98.7306, y: 96.5144, z: -9.05762 },
+		     { x: -98.7306, y: 96.5144, z: -9.05762 },
+		     { x: -95.7664, y: 95.4621, z: -10.29 },
+		     { x: -94.9573, y: 95.4512, z: -11.5087 },
+		     { x: -92.3744, y: 96.2407, z: -16.3409 },
+		     { x: -91.1827, y: 97.6606, z: -18.7904 },
+		     { x: -90.107, y: 99.6772, z: -21.2916 },
+		     { x: -86.8607, y: 102.44, z: -24.5526 },
+		     { x: -85.5897, y: 104.89, z: -26.2713 },
+		     { x: -83.9378, y: 110.294, z: -29.9599 },
+		     { x: -83.2107, y: 112.676, z: -31.9604 },
+		     { x: -82.7183, y: 116.082, z: -33.8161 },
+		     { x: -82.4409, y: 121.734, z: -36.4523 },
+		     { x: -82.748, y: 124.808, z: -37.7105 },
+		     { x: -84.1727, y: 130.647, z: -39.6702 },
+		     { x: -85.7225, y: 133.864, z: -40.3538 },
+		     { x: -88.7312, y: 140.248, z: -41.3561 },
+		     { x: -90.4438, y: 142.932, z: -41.663 },
+		     { x: -95.229, y: 147.729, z: -42.7127 },
+		     { x: -98.0241, y: 151.669, z: -42.1982 },
+		     { x: -103.142, y: 156.543, z: -42.6121 },
+		     { x: -109.104, y: 162.075, z: -41.5821 },
+		     { x: -111.316, y: 163.878, z: -40.1399 },
+		     { x: -116.698, y: 167.158, z: -37.8018 },
+		     { x: -122.958, y: 170.484, z: -35.7106 },
+		     { x: -126.553, y: 173.063, z: -34.7524 },
+		     { x: -132.732, y: 173.782, z: -29.9023 },
+		     { x: -138.393, y: 175.923, z: -27.2039 },
+		     { x: -141.736, y: 178.034, z: -26.5215 },
+		     { x: -147.05, y: 178.333, z: -23.1627 },
+		     { x: -153.968, y: 180.719, z: -21.7373 },
+		     { x: -157.274, y: 181.327, z: -21.129 },
+		     { x: -162.401, y: 180.888, z: -18.6173 },
+		     { x: -165.917, y: 177.051, z: -13.7202 },
+		     { x: -170.267, y: 174.528, z: -10.3724 },
+		     { x: -172.106, y: 173.588, z: -9.25301 },
+		     { x: -175.307, y: 170.107, z: -4.76177 },
+		     { x: -180.347, y: 169.401, z: -5.21372 },
+		     { x: -182.337, y: 168.241, z: -4.87182 },
+		     { x: -185.569, y: 164.958, z: -3.35368 },
+		     { x: -188.771, y: 162.756, z: -1.90554 },
+		     { x: -192.4, y: 160.135, z: -0.541541 },
+		     { x: -193.827, y: 158.859, z: 0.447863 },
+		     { x: -196.207, y: 155.483, z: 1.91273 },
+		     { x: -197.275, y: 153.869, z: 2.18946 },
+		     { x: -200.142, y: 151.012, z: 3.55447 },
+		     { x: -201.223, y: 148.595, z: 4.4758 },
+		     { x: -200.838, y: 147.453, z: 5.04884 },
+		     { x: -201.58, y: 146.192, z: 5.64657 },
+		     { x: -202.597, y: 145.253, z: 5.3586 },
+		     { x: -202.17, y: 142.902, z: 5.19806 },
+		     { x: -201.92, y: 142.546, z: 4.75463 },
+		     { x: -201.734, y: 142.119, z: 4.02854 },
+		     { x: -201.414, y: 141.991, z: 3.93973 },
+		     { x: -201.899, y: 141.924, z: 3.60378 },
+		     { x: -201.814, y: 141.829, z: 3.40841 },
+		     { x: -201.781, y: 141.667, z: 3.6848 },
+		     { x: -200.729, y: 140.384, z: 3.6199 },
+		     { x: -199.675, y: 139.363, z: 3.59756 },
+		     { x: -198.766, y: 138.276, z: 3.70933 },
+		     { x: -198.816, y: 137.95, z: 2.99503 },
+		     { x: -198.617, y: 137.662, z: 2.32567 },
+		     { x: -197.955, y: 137.537, z: 1.50783 },
+		     { x: -197.639, y: 137.647, z: 1.31538 },
+		     { x: -197.308, y: 137.871, z: 1.18853 },
+		     { x: -196.673, y: 138.203, z: 0.85756 },
+		     { x: -196.283, y: 138.727, z: 0.475924 },
+		     { x: -195.599, y: 140.038, z: -1.33436 },
+		     { x: -194.83, y: 139.861, z: -1.60068 },
+		     { x: -194.04, y: 139.657, z: -1.11458 },
+		     { x: -193.454, y: 139.88, z: -0.978403 },
+		     { x: -193.881, y: 140.799, z: -1.60343 },
+		     { x: -194.287, y: 141.489, z: -2.11328 },
+		     { x: -195.538, y: 142.963, z: -2.72804 },
+		     { x: -195.158, y: 143.641, z: -3.01564 },
+		     { x: -194.806, y: 144.254, z: -3.10805 },
+		     { x: -194.248, y: 144.613, z: -3.23137 },
+		     { x: -194.517, y: 145.049, z: -3.3281 },
+		     { x: -194.483, y: 145.703, z: -3.42576 },
+		     { x: -194.593, y: 146.306, z: -3.41337 },
+		     { x: -194.255, y: 146.649, z: -3.48564 },
+		     { x: -193.89, y: 147.006, z: -3.52341 },
+		     { x: -193.323, y: 147.37, z: -3.82136 },
+		     { x: -192.632, y: 147.845, z: -3.8046 },
+		     { x: -192.231, y: 148.262, z: -3.83967 },
+		     { x: -191.944, y: 148.276, z: -3.6829 },
+		     { x: -191.791, y: 148.277, z: -3.60759 } ]
+    },
+    {
+	"name": "11",
+	"points":  [ { x: 7.44908, y: 99.9709, z: -87.3742 },
+		     { x: 7.65108, y: 99.901, z: -87.4365 },
+		     { x: 7.65108, y: 99.901, z: -87.4365 },
+		     { x: 7.79089, y: 99.8577, z: -87.4849 },
+		     { x: 7.90951, y: 99.7909, z: -87.5256 },
+		     { x: 7.90951, y: 99.7909, z: -87.5256 },
+		     { x: 8.05491, y: 99.7157, z: -87.5625 },
+		     { x: 8.18253, y: 99.6442, z: -87.5903 },
+		     { x: 8.18253, y: 99.6442, z: -87.5903 },
+		     { x: 8.32443, y: 99.5094, z: -87.6203 },
+		     { x: 8.49169, y: 99.3122, z: -87.6292 },
+		     { x: 8.49169, y: 99.3122, z: -87.6292 },
+		     { x: 8.63034, y: 99.0579, z: -87.6116 },
+		     { x: 8.63057, y: 98.6935, z: -87.6479 },
+		     { x: 8.63057, y: 98.6935, z: -87.6479 },
+		     { x: 8.62646, y: 98.424, z: -87.6507 },
+		     { x: 8.49024, y: 97.7199, z: -87.5167 },
+		     { x: 8.49024, y: 97.7199, z: -87.5167 },
+		     { x: 8.17635, y: 96.2762, z: -87.0293 },
+		     { x: 7.6179, y: 94.0137, z: -86.549 },
+		     { x: 6.67696, y: 91.7163, z: -85.6209 },
+		     { x: 6.67696, y: 91.7163, z: -85.6209 },
+		     { x: 5.51458, y: 89.4092, z: -84.5039 },
+		     { x: 4.55569, y: 86.3018, z: -83.8372 },
+		     { x: 2.98632, y: 82.8333, z: -82.9169 },
+		     { x: 1.91386, y: 79.5622, z: -81.5996 },
+		     { x: 1.17234, y: 76.7716, z: -80.8432 },
+		     { x: 0.852255, y: 74.0359, z: -79.8578 },
+		     { x: -0.328618, y: 70.6882, z: -80.0349 },
+		     { x: -1.07761, y: 68.0332, z: -79.9684 },
+		     { x: -2.25773, y: 62.5363, z: -79.4257 },
+		     { x: -3.05132, y: 59.2712, z: -79.5548 },
+		     { x: -4.11262, y: 55.7333, z: -79.9226 },
+		     { x: -4.69265, y: 53.5656, z: -80.1167 },
+		     { x: -5.11696, y: 49.3569, z: -80.1618 },
+		     { x: -5.14581, y: 44.7688, z: -79.9254 },
+		     { x: -7.14982, y: 41.8191, z: -76.8324 },
+		     { x: -7.97902, y: 39.9839, z: -74.8072 },
+		     { x: -7.58152, y: 38.5977, z: -73.7412 },
+		     { x: -7.51978, y: 36.9749, z: -73.2823 },
+		     { x: -7.13022, y: 34.8192, z: -72.7689 },
+		     { x: -6.98515, y: 33.9973, z: -72.5418 },
+		     { x: -6.67294, y: 33.4995, z: -72.4014 },
+		     { x: -5.90509, y: 32.5484, z: -72.2361 },
+		     { x: -5.51878, y: 31.9301, z: -72.1616 },
+		     { x: -5.21389, y: 31.3752, z: -72.0408 },
+		     { x: -4.72527, y: 30.1088, z: -71.7112 },
+		     { x: -4.48379, y: 29.5973, z: -71.5855 },
+		     { x: -4.15158, y: 29.2498, z: -71.4727 },
+		     { x: -3.35828, y: 28.8567, z: -71.3687 },
+		     { x: -2.84355, y: 29.2523, z: -71.1054 },
+		     { x: -2.29854, y: 29.705, z: -70.8771 },
+		     { x: -1.82636, y: 30.0908, z: -70.7178 },
+		     { x: -1.27962, y: 30.4165, z: -70.6085 },
+		     { x: -0.642706, y: 31.2598, z: -70.3161 },
+		     { x: -0.149144, y: 32.0184, z: -70.1158 },
+		     { x: 0.274286, y: 32.715, z: -69.9982 },
+		     { x: 0.682866, y: 33.4269, z: -69.9588 },
+		     { x: 0.682866, y: 33.4269, z: -69.9588 },
+		     { x: 1.07952, y: 34.78, z: -69.8691 },
+		     { x: 1.13453, y: 35.8269, z: -69.5114 },
+		     { x: 1.41517, y: 36.5252, z: -69.0515 },
+		     { x: 1.76509, y: 36.9074, z: -68.6583 },
+		     { x: 1.96009, y: 37.3751, z: -68.5003 },
+		     { x: 1.96009, y: 37.3751, z: -68.5003 },
+		     { x: 2.08951, y: 37.6912, z: -68.437 },
+		     { x: 2.13257, y: 37.8796, z: -68.3914 },
+		     { x: 2.15838, y: 38.105, z: -68.3876 },
+		     { x: 2.21055, y: 38.2771, z: -68.3876 },
+		     { x: 2.05523, y: 38.5018, z: -68.1395 },
+		     { x: 2.05523, y: 38.5018, z: -68.1395 },
+		     { x: 2.09623, y: 38.635, z: -67.9723 },
+		     { x: 2.10974, y: 39.043, z: -67.9444 },
+		     { x: 2.23047, y: 39.4185, z: -67.9603 },
+		     { x: 2.44307, y: 39.5892, z: -68.4966 },
+		     { x: 2.55017, y: 39.9305, z: -68.901 },
+		     { x: 2.55017, y: 39.9305, z: -68.901 },
+		     { x: 2.58416, y: 40.2402, z: -69.1636 },
+		     { x: 2.41449, y: 40.3741, z: -69.4324 },
+		     { x: 2.26896, y: 40.5534, z: -69.6702 },
+		     { x: 2.26896, y: 40.5534, z: -69.6702 },
+		     { x: 2.15131, y: 40.78, z: -69.8669 },
+		     { x: 2.06958, y: 41.0041, z: -70.0295 },
+		     { x: 2.00355, y: 41.2146, z: -70.1684 },
+		     { x: 2.00355, y: 41.2146, z: -70.1684 },
+		     { x: 2.03105, y: 41.4403, z: -70.3216 } ]
+    },
+    {
+	"name": "11'",
+	"points":  [ { x: -45.4143, y: 31.7005, z: -20.7345 },
+		     { x: -45.3399, y: 31.4889, z: -18.9577 },
+		     { x: -45.5717, y: 31.7544, z: -17.7975 },
+		     { x: -46.5488, y: 33.0066, z: -16.2471 },
+		     { x: -48.3915, y: 37.7314, z: -14.0554 },
+		     { x: -49.2533, y: 40.2848, z: -13.5428 },
+		     { x: -50.3298, y: 43.1535, z: -12.713 },
+		     { x: -51.6133, y: 46.0844, z: -12.0524 },
+		     { x: -52.6075, y: 49.0092, z: -11.486 },
+		     { x: -53.3835, y: 51.9756, z: -11.1562 },
+		     { x: -53.731, y: 55.3358, z: -11.3178 },
+		     { x: -54.0572, y: 59.5764, z: -12.9262 },
+		     { x: -54.5888, y: 61.247, z: -13.2858 },
+		     { x: -55.132, y: 63.0685, z: -12.7798 },
+		     { x: -55.439, y: 64.6252, z: -12.5154 },
+		     { x: -56.4676, y: 67.9264, z: -13.0425 },
+		     { x: -57.1054, y: 69.7084, z: -15.1436 },
+		     { x: -58.4202, y: 72.699, z: -18.0671 },
+		     { x: -59.5631, y: 77.2905, z: -18.2125 },
+		     { x: -60.9506, y: 82.4588, z: -16.9096 },
+		     { x: -62.4915, y: 89.1437, z: -15.0573 },
+		     { x: -63.2903, y: 93.1534, z: -15.7743 },
+		     { x: -63.7863, y: 97.0712, z: -16.4728 },
+		     { x: -65.9483, y: 102.002, z: -22.276 },
+		     { x: -66.5152, y: 103.599, z: -23.055 },
+		     { x: -66.7371, y: 104.671, z: -21.1596 },
+		     { x: -66.7371, y: 104.671, z: -21.1596 },
+		     { x: -67.0294, y: 109.306, z: -21.5141 },
+		     { x: -67.0784, y: 110.071, z: -22.0111 },
+		     { x: -67.1246, y: 111.454, z: -22.4819 },
+		     { x: -67.1749, y: 112.107, z: -22.5795 },
+		     { x: -67.174, y: 112.698, z: -22.7611 },
+		     { x: -67.1191, y: 114.006, z: -23.3046 },
+		     { x: -67.1037, y: 114.369, z: -23.4138 },
+		     { x: -67.0472, y: 115.307, z: -23.1477 },
+		     { x: -67.0178, y: 115.521, z: -23.1095 },
+		     { x: -66.738, y: 116.656, z: -22.333 },
+		     { x: -66.4854, y: 117.465, z: -21.8428 },
+		     { x: -66.2462, y: 118.138, z: -21.5414 },
+		     { x: -65.9634, y: 119.772, z: -21.0894 },
+		     { x: -65.8478, y: 120.211, z: -21.0178 },
+		     { x: -65.7191, y: 120.601, z: -20.9881 },
+		     { x: -65.61, y: 120.88, z: -20.9923 },
+		     { x: -65.3813, y: 121.442, z: -20.8514 },
+		     { x: -65.2439, y: 122.132, z: -20.3629 },
+		     { x: -65.0962, y: 122.631, z: -20.0863 },
+		     { x: -64.9011, y: 123.097, z: -19.6477 },
+		     { x: -64.6882, y: 123.493, z: -19.4434 },
+		     { x: -64.2834, y: 124.093, z: -19.4111 },
+		     { x: -64.0953, y: 124.333, z: -19.4638 },
+		     { x: -63.9436, y: 124.513, z: -19.5181 },
+		     { x: -63.8193, y: 124.636, z: -19.5788 },
+		     { x: -63.7002, y: 124.729, z: -19.656 },
+		     { x: -63.5877, y: 124.804, z: -19.737 },
+		     { x: -63.4926, y: 124.86, z: -19.8096 },
+		     { x: -63.3997, y: 124.942, z: -19.8648 },
+		     { x: -63.3091, y: 124.995, z: -19.9397 },
+		     { x: -63.2283, y: 125.028, z: -20.0144 },
+		     { x: -63.2283, y: 125.028, z: -20.0144 },
+		     { x: -63.1393, y: 125.054, z: -20.1056 },
+		     { x: -63.0514, y: 125.056, z: -20.2104 },
+		     { x: -62.9672, y: 125.04, z: -20.3234 },
+		     { x: -62.9672, y: 125.04, z: -20.3234 },
+		     { x: -62.9715, y: 124.845, z: -20.2408 },
+		     { x: -62.963, y: 124.677, z: -20.1866 },
+		     { x: -62.9394, y: 124.541, z: -20.1858 },
+		     { x: -62.9394, y: 124.541, z: -20.1858 },
+		     { x: -62.9087, y: 124.426, z: -20.2085 },
+		     { x: -62.8696, y: 124.314, z: -20.2641 },
+		     { x: -62.8336, y: 124.203, z: -20.328 },
+		     { x: -62.8336, y: 124.203, z: -20.328 },
+		     { x: -62.8206, y: 124.135, z: -20.3452 },
+		     { x: -62.8071, y: 124.057, z: -20.376 },
+		     { x: -62.7998, y: 124.011, z: -20.3835 },
+		     { x: -62.7998, y: 124.011, z: -20.3835 },
+		     { x: -62.8001, y: 123.991, z: -20.3673 },
+		     { x: -62.8022, y: 123.969, z: -20.353 },
+		     { x: -62.8047, y: 123.948, z: -20.3403 },
+		     { x: -62.8047, y: 123.948, z: -20.3403 },
+		     { x: -62.8068, y: 123.931, z: -20.3278 },
+		     { x: -62.9167, y: 123.863, z: -20.228 },
+		     { x: -62.9167, y: 123.863, z: -20.228 },
+		     { x: -63.2566, y: 123.688, z: -19.9973 },
+		     { x: -63.4993, y: 123.561, z: -19.8266 },
+		     { x: -63.7295, y: 123.573, z: -19.5675 },
+		     { x: -63.7295, y: 123.573, z: -19.5675 },
+		     { x: -63.8848, y: 123.54, z: -19.4199 } ]
+    },
+    {
+	"name": "12",
+	"points":  [ { x: -6.91777, y: 166.093, z: -65.7583 },
+		     { x: -6.67177, y: 166.2, z: -65.7398 },
+		     { x: -6.57747, y: 166.099, z: -65.7118 },
+		     { x: -6.57747, y: 166.099, z: -65.7118 },
+		     { x: -7.05601, y: 164.512, z: -67.1802 },
+		     { x: -6.59844, y: 164.585, z: -67.8568 },
+		     { x: -6.49202, y: 163.861, z: -68.7817 },
+		     { x: -6.42109, y: 163.312, z: -69.3967 },
+		     { x: -6.49327, y: 162.425, z: -70.0066 },
+		     { x: -6.49327, y: 162.425, z: -70.0066 },
+		     { x: -6.47819, y: 161.914, z: -70.333 },
+		     { x: -6.45596, y: 161.318, z: -70.7751 },
+		     { x: -6.44163, y: 160.445, z: -71.2787 },
+		     { x: -6.85692, y: 158.415, z: -72.0011 },
+		     { x: -7.09917, y: 156.758, z: -72.389 },
+		     { x: -7.02231, y: 155.1, z: -72.6786 },
+		     { x: -7.02231, y: 155.1, z: -72.6786 },
+		     { x: -7.0575, y: 153.01, z: -72.9084 },
+		     { x: -7.45546, y: 150.195, z: -73.0359 },
+		     { x: -7.84788, y: 146.465, z: -73.1673 },
+		     { x: -9.44274, y: 144.048, z: -72.5477 },
+		     { x: -10.9271, y: 140.748, z: -72.3946 },
+		     { x: -11.9086, y: 135.795, z: -72.3371 },
+		     { x: -13.0925, y: 130.925, z: -72.0563 },
+		     { x: -13.6706, y: 123.501, z: -70.0289 },
+		     { x: -13.823, y: 120.023, z: -69.8103 },
+		     { x: -13.3964, y: 116.212, z: -69.828 },
+		     { x: -11.1735, y: 113.088, z: -64.2646 },
+		     { x: -11.8049, y: 108.855, z: -65.0968 },
+		     { x: -12.1298, y: 104.747, z: -65.9181 },
+		     { x: -12.4775, y: 97.7794, z: -65.9099 },
+		     { x: -12.1384, y: 95.0279, z: -66.0868 },
+		     { x: -11.7801, y: 89.7722, z: -66.0267 },
+		     { x: -11.4842, y: 87.2808, z: -65.8262 },
+		     { x: -10.6354, y: 82.2277, z: -65.7958 },
+		     { x: -10.1968, y: 80.1005, z: -65.4085 },
+		     { x: -9.72417, y: 76.5319, z: -65.2534 },
+		     { x: -9.38629, y: 74.8637, z: -64.9548 },
+		     { x: -9.01277, y: 73.1047, z: -64.6266 },
+		     { x: -8.35977, y: 69.7352, z: -63.6823 },
+		     { x: -8.24191, y: 68.1565, z: -63.0454 },
+		     { x: -8.18555, y: 65.514, z: -62.5336 },
+		     { x: -8.14057, y: 64.0858, z: -62.4352 },
+		     { x: -5.29192, y: 63.448, z: -63.7289 },
+		     { x: -5.95458, y: 61.8222, z: -62.7983 },
+		     { x: -6.31855, y: 60.7141, z: -62.3121 },
+		     { x: -6.31855, y: 60.7141, z: -62.3121 },
+		     { x: -6.31233, y: 57.4583, z: -61.0064 },
+		     { x: -6.28904, y: 56.6746, z: -60.5943 },
+		     { x: -6.25989, y: 55.8407, z: -60.0714 },
+		     { x: -5.95068, y: 54.3533, z: -58.9151 },
+		     { x: -6.17049, y: 53.8635, z: -58.753 },
+		     { x: -6.06234, y: 53.5705, z: -58.5547 },
+		     { x: -5.91902, y: 53.1827, z: -58.3499 },
+		     { x: -5.76231, y: 52.8794, z: -58.212 },
+		     { x: -5.26985, y: 51.8576, z: -57.7385 },
+		     { x: -4.8912, y: 51.0346, z: -57.432 },
+		     { x: -4.55256, y: 50.2403, z: -57.0713 },
+		     { x: -4.26385, y: 49.4892, z: -56.638 },
+		     { x: -3.98893, y: 48.6003, z: -55.9471 },
+		     { x: -3.82607, y: 47.7899, z: -55.1981 },
+		     { x: -3.79733, y: 47.3446, z: -54.2787 },
+		     { x: -3.95259, y: 46.9999, z: -53.4187 },
+		     { x: -4.19633, y: 46.8029, z: -52.9747 },
+		     { x: -4.43948, y: 46.7229, z: -52.5989 },
+		     { x: -4.59388, y: 46.7275, z: -52.5715 },
+		     { x: -4.66546, y: 46.4166, z: -52.3627 },
+		     { x: -4.72654, y: 46.0183, z: -52.0369 },
+		     { x: -4.79269, y: 45.2677, z: -51.641 },
+		     { x: -4.86876, y: 44.6693, z: -51.2771 },
+		     { x: -4.99467, y: 43.9634, z: -50.9307 },
+		     { x: -5.20343, y: 43.0902, z: -50.5828 },
+		     { x: -5.47185, y: 41.9962, z: -50.2024 },
+		     { x: -5.9011, y: 40.4063, z: -49.3905 },
+		     { x: -6.2882, y: 38.9172, z: -48.7229 },
+		     { x: -6.65381, y: 37.2456, z: -48.0107 },
+		     { x: -6.8464, y: 35.9499, z: -47.5402 },
+		     { x: -6.91103, y: 34.6292, z: -46.9936 },
+		     { x: -6.73663, y: 32.6037, z: -44.752 },
+		     { x: -6.77145, y: 31.0402, z: -42.513 },
+		     { x: -6.85893, y: 30.3487, z: -41.453 },
+		     { x: -6.75641, y: 29.7002, z: -40.1573 },
+		     { x: -6.59717, y: 29.3396, z: -39.4649 },
+		     { x: -6.34712, y: 29.0259, z: -38.7167 },
+		     { x: -6.17829, y: 28.8594, z: -38.164 },
+		     { x: -5.94851, y: 28.5814, z: -37.0481 },
+		     { x: -6, y: 28.3892, z: -36.6165 },
+		     { x: -6.0435, y: 28.171, z: -36.1685 },
+		     { x: -6.05266, y: 27.8911, z: -35.72 },
+		     { x: -5.96022, y: 26.9103, z: -34.5161 },
+		     { x: -5.89314, y: 26.5235, z: -34.0957 },
+		     { x: -5.70858, y: 25.9698, z: -33.4881 },
+		     { x: -5.54008, y: 25.5116, z: -33.0053 },
+		     { x: -5.39549, y: 24.9831, z: -32.5202 },
+		     { x: -5.33239, y: 24.6636, z: -32.3267 },
+		     { x: -5.25684, y: 24.2612, z: -32.107 },
+		     { x: -5.27338, y: 23.896, z: -31.9049 },
+		     { x: -5.29764, y: 23.4663, z: -31.5422 },
+		     { x: -5.32051, y: 22.8996, z: -31.1537 },
+		     { x: -5.3289, y: 22.2422, z: -30.7518 },
+		     { x: -5.37822, y: 21.4915, z: -30.1606 },
+		     { x: -5.70236, y: 20.8119, z: -29.6402 },
+		     { x: -6.03689, y: 20.0597, z: -29.1488 },
+		     { x: -6.32784, y: 19.2676, z: -28.6649 },
+		     { x: -6.65167, y: 18.6379, z: -28.2115 },
+		     { x: -6.65167, y: 18.6379, z: -28.2115 },
+		     { x: -6.95057, y: 18.0067, z: -27.6841 },
+		     { x: -7.17721, y: 17.4976, z: -27.2148 },
+		     { x: -7.38125, y: 17.0609, z: -26.7865 },
+		     { x: -7.5846, y: 16.6959, z: -26.4297 },
+		     { x: -7.78584, y: 16.4239, z: -26.0789 },
+		     { x: -7.78584, y: 16.4239, z: -26.0789 },
+		     { x: -8.02074, y: 16.2708, z: -25.7322 },
+		     { x: -8.20533, y: 16.1422, z: -25.4566 },
+		     { x: -8.35472, y: 16.0369, z: -25.239 },
+		     { x: -8.49726, y: 15.9841, z: -25.0564 },
+		     { x: -8.65253, y: 15.9478, z: -24.8968 },
+		     { x: -8.65253, y: 15.9478, z: -24.8968 },
+		     { x: -8.78892, y: 15.9128, z: -24.7594 },
+		     { x: -8.92419, y: 15.8683, z: -24.627 },
+		     { x: -9.07048, y: 15.8398, z: -24.4925 },
+		     { x: -9.07048, y: 15.8398, z: -24.4925 },
+		     { x: -9.21284, y: 15.8407, z: -24.362 },
+		     { x: -9.32797, y: 15.8356, z: -24.2521 },
+		     { x: -9.40204, y: 15.8117, z: -24.1633 },
+		     { x: -9.40204, y: 15.8117, z: -24.1633 },
+		     { x: -9.35102, y: 15.7561, z: -24.0441 },
+		     { x: -9.30755, y: 15.6876, z: -23.9235 },
+		     { x: -9.29214, y: 15.6493, z: -23.8154 },
+		     { x: -9.29214, y: 15.6493, z: -23.8154 },
+		     { x: -9.25518, y: 15.6015, z: -23.707 },
+		     { x: -9.22967, y: 15.5569, z: -23.6223 },
+		     { x: -9.21058, y: 15.5173, z: -23.5512 },
+		     { x: -9.21058, y: 15.5173, z: -23.5512 },
+		     { x: -9.18507, y: 15.4786, z: -23.4814 },
+		     { x: -9.15679, y: 15.4395, z: -23.4169 },
+		     { x: -9.15679, y: 15.4395, z: -23.4169 },
+		     { x: -9.21157, y: 15.3831, z: -23.3062 },
+		     { x: -9.23633, y: 15.3372, z: -23.2181 },
+		     { x: -9.24273, y: 15.3023, z: -23.148 },
+		     { x: -9.24273, y: 15.3023, z: -23.148 },
+		     { x: -9.2396, y: 15.2747, z: -23.086 },
+		     { x: -9.23419, y: 15.2551, z: -23.0348 },
+		     { x: -9.23419, y: 15.2551, z: -23.0348 },
+		     { x: -9.233, y: 15.2457, z: -22.9875 },
+		     { x: -9.22935, y: 15.2384, z: -22.9309 },
+		     { x: -9.22935, y: 15.2384, z: -22.9309 },
+		     { x: -9.21904, y: 15.2369, z: -22.8699 },
+		     { x: -9.2097, y: 15.2412, z: -22.8079 },
+		     { x: -9.2097, y: 15.2412, z: -22.8079 },
+		     { x: -9.20194, y: 15.2394, z: -22.721 },
+		     { x: -9.19835, y: 15.2239, z: -22.6281 },
+		     { x: -9.19504, y: 15.219, z: -22.5531 },
+		     { x: -9.19504, y: 15.219, z: -22.5531 },
+		     { x: -9.19119, y: 15.2429, z: -22.4822 } ]
+    },
+    {
+	"name": "12'",
+	"points":  [ { x: -47.4992, y: 16.4636, z: -17.1548 },
+		     { x: -46.613, y: 14.3183, z: -17.4878 },
+		     { x: -46.1331, y: 15.8349, z: -17.698 },
+		     { x: -46.2521, y: 19.5935, z: -16.1098 },
+		     { x: -46.3312, y: 22.8035, z: -14.4543 },
+		     { x: -46.1594, y: 27.1189, z: -12.0052 },
+		     { x: -13.8923, y: -18.5565, z: 15.6279 },
+		     { x: -13.9784, y: -19.0678, z: 19.1167 },
+		     { x: -14.6738, y: -18.0304, z: 19.7043 },
+		     { x: -16.7822, y: -15.1978, z: 20.5645 },
+		     { x: -17.5906, y: -13.0629, z: 20.2771 },
+		     { x: -17.8253, y: -10.7939, z: 18.1166 },
+		     { x: -17.9422, y: -9.35601, z: 14.4445 },
+		     { x: -18.3202, y: -7.24094, z: 11.0348 },
+		     { x: -18.8312, y: -4.59727, z: 7.12322 },
+		     { x: -19.3272, y: -3.45411, z: 6.20955 },
+		     { x: -15.1721, y: 0.789552, z: 22.8176 },
+		     { x: -11.1617, y: 3.75795, z: 35.0688 },
+		     { x: -9.96118, y: 5.46423, z: 38.2423 },
+		     { x: -10.6919, y: 12.9819, z: 46.0922 },
+		     { x: -10.3277, y: 16.279, z: 49.4963 },
+		     { x: -22.1636, y: 44.3749, z: -2.62599 },
+		     { x: -22.1636, y: 44.3749, z: -2.62599 },
+		     { x: -19.8492, y: 67.0746, z: -30.0376 },
+		     { x: -18.1949, y: 72.0197, z: -34.3138 },
+		     { x: -17.7734, y: 75.282, z: -36.1417 },
+		     { x: -17.7708, y: 77.8738, z: -37.2737 },
+		     { x: -17.8122, y: 80.3347, z: -38.1298 },
+		     { x: -17.8227, y: 81.5336, z: -38.6279 },
+		     { x: -18.2673, y: 84.0295, z: -39.6734 },
+		     { x: -18.1297, y: 86.136, z: -39.8881 },
+		     { x: -17.8111, y: 88.5282, z: -39.8451 },
+		     { x: -17.9006, y: 91.2703, z: -40.5959 },
+		     { x: -18.1174, y: 94.5194, z: -41.7448 },
+		     { x: -18.0848, y: 96.2254, z: -41.7239 },
+		     { x: -18.4863, y: 100.092, z: -43.9854 },
+		     { x: -18.5366, y: 102.929, z: -44.6476 },
+		     { x: -22.4124, y: 98.4535, z: -45.581 },
+		     { x: -22.2482, y: 106.731, z: -47.58 },
+		     { x: -22.1531, y: 111.701, z: -48.1552 },
+		     { x: -22.4613, y: 114.129, z: -49.0649 },
+		     { x: -22.8539, y: 117.209, z: -50.282 },
+		     { x: -22.9877, y: 121.979, z: -51.4085 },
+		     { x: -22.9249, y: 124.383, z: -52.6893 },
+		     { x: -22.8574, y: 127.946, z: -53.6049 },
+		     { x: -22.9342, y: 129.78, z: -53.9541 },
+		     { x: -23.2007, y: 133.491, z: -55.0349 },
+		     { x: -23.5948, y: 135.76, z: -55.355 },
+		     { x: -24.1191, y: 139.94, z: -56.1398 },
+		     { x: -24.519, y: 142.277, z: -56.7112 },
+		     { x: -24.7449, y: 145.74, z: -57.4516 },
+		     { x: -24.8491, y: 147.403, z: -57.9224 },
+		     { x: -25.4035, y: 152.033, z: -59.3384 },
+		     { x: -25.625, y: 154.102, z: -60.2576 },
+		     { x: -26.197, y: 159.186, z: -61.2889 },
+		     { x: -26.2155, y: 161.205, z: -61.6212 },
+		     { x: -26.3571, y: 163.212, z: -61.9636 },
+		     { x: -26.4454, y: 166.299, z: -62.1276 },
+		     { x: -26.4348, y: 167.816, z: -61.8699 },
+		     { x: -26.4058, y: 169.172, z: -61.5654 },
+		     { x: -26.5544, y: 173.156, z: -60.934 },
+		     { x: -26.7391, y: 175.23, z: -60.6441 },
+		     { x: -26.7391, y: 175.23, z: -60.6441 },
+		     { x: -27.1809, y: 182.35, z: -61.8589 },
+		     { x: -27.4431, y: 186.032, z: -64.0831 },
+		     { x: -27.8075, y: 187.773, z: -65.0452 },
+		     { x: -28.0821, y: 190.306, z: -66.4857 },
+		     { x: -28.628, y: 194.101, z: -67.7282 },
+		     { x: -28.7301, y: 195.698, z: -67.9936 },
+		     { x: -28.5003, y: 198.06, z: -68.2528 },
+		     { x: -28.8244, y: 200.479, z: -68.5956 },
+		     { x: -28.7279, y: 202.169, z: -68.5284 },
+		     { x: -28.9414, y: 203.828, z: -68.6972 },
+		     { x: -29.4177, y: 207.458, z: -69.3102 },
+		     { x: -29.6714, y: 209.326, z: -69.8678 },
+		     { x: -29.718, y: 210.727, z: -69.829 },
+		     { x: -29.8626, y: 213.039, z: -69.4556 },
+		     { x: -30.0191, y: 213.66, z: -69.333 },
+		     { x: -29.8742, y: 215.992, z: -68.8784 },
+		     { x: -29.6184, y: 218.653, z: -68.4698 },
+		     { x: -29.4451, y: 223.173, z: -67.4107 },
+		     { x: -29.3372, y: 225.572, z: -66.6819 },
+		     { x: -29.1087, y: 228.182, z: -65.4972 },
+		     { x: -28.8838, y: 230.013, z: -65.1095 },
+		     { x: -28.5642, y: 234.549, z: -64.1441 },
+		     { x: -28.1528, y: 237.655, z: -63.7249 },
+		     { x: -28.2354, y: 239.703, z: -63.5454 },
+		     { x: -27.6996, y: 241.531, z: -63.7692 },
+		     { x: -27.352, y: 244.513, z: -63.2321 },
+		     { x: -28.2032, y: 250.701, z: -61.8362 },
+		     { x: -27.9726, y: 251.684, z: -62.526 },
+		     { x: -27.7506, y: 253.101, z: -62.5486 },
+		     { x: -27.6916, y: 254.86, z: -62.2457 },
+		     { x: -27.8624, y: 267.769, z: -58.2524 },
+		     { x: -28.3191, y: 270.046, z: -56.8466 },
+		     { x: -28.4993, y: 270.85, z: -56.3404 },
+		     { x: -28.6819, y: 271.411, z: -55.9641 },
+		     { x: -28.7322, y: 271.554, z: -55.8673 },
+		     { x: -28.7671, y: 271.675, z: -55.7886 },
+		     { x: -28.8271, y: 271.989, z: -55.3995 },
+		     { x: -28.8255, y: 272.209, z: -55.1333 },
+		     { x: -28.8225, y: 272.432, z: -54.8639 },
+		     { x: -28.8238, y: 272.491, z: -54.79 },
+		     { x: -28.8459, y: 272.505, z: -54.7413 },
+		     { x: -28.8883, y: 272.511, z: -54.6744 },
+		     { x: -28.9071, y: 272.508, z: -54.6506 },
+		     { x: -28.918, y: 272.52, z: -54.6253 },
+		     { x: -13.866, y: 221.507, z: 2.18759 },
+		     { x: -12.0777, y: 216.685, z: 10.4878 },
+		     { x: -12.9158, y: 216.211, z: 9.86127 },
+		     { x: -12.9158, y: 216.211, z: 9.86127 },
+		     { x: -17.426, y: 213.072, z: 7.42101 },
+		     { x: -21.05, y: 206.543, z: 7.25876 },
+		     { x: -23.0742, y: 204.572, z: 9.04634 },
+		     { x: -23.4827, y: 204.268, z: 9.33822 },
+		     { x: -23.5856, y: 204.191, z: 9.3927 },
+		     { x: -23.727, y: 204.088, z: 9.45875 },
+		     { x: -23.8156, y: 204.04, z: 9.49978 },
+		     { x: -23.8625, y: 204.048, z: 9.53534 },
+		     { x: -23.8991, y: 204.107, z: 9.51393 },
+		     { x: -23.908, y: 204.192, z: 9.50978 },
+		     { x: -23.9118, y: 204.27, z: 9.50603 },
+		     { x: -23.909, y: 204.309, z: 9.50706 },
+		     { x: -23.8854, y: 204.396, z: 9.4617 },
+		     { x: -24.0164, y: 204.53, z: 8.84184 },
+		     { x: -24.0769, y: 204.621, z: 8.516 },
+		     { x: -24.0919, y: 204.662, z: 8.42147 },
+		     { x: -25.2739, y: 204.877, z: 6.8711 },
+		     { x: -26.3813, y: 205.154, z: 5.59293 },
+		     { x: -26.5615, y: 205.258, z: 5.42079 },
+		     { x: -27.1634, y: 205.764, z: 4.7238 },
+		     { x: -29.6384, y: 208.858, z: 1.92252 },
+		     { x: -31.4042, y: 211.235, z: -0.0236034 },
+		     { x: -32.4468, y: 212.527, z: -1.2346 },
+		     { x: -33.1204, y: 213.178, z: -2.12288 },
+		     { x: -33.4565, y: 213.28, z: -2.7197 },
+		     { x: -33.4224, y: 212.513, z: -3.20627 },
+		     { x: -32.8426, y: 210.913, z: -3.77644 },
+		     { x: -32.4603, y: 209.825, z: -3.86625 },
+		     { x: -32.2267, y: 209.168, z: -3.92999 },
+		     { x: -32.077, y: 208.75, z: -3.97753 },
+		     { x: -31.9685, y: 208.426, z: -4.04105 },
+		     { x: -31.6101, y: 207.496, z: -3.99463 },
+		     { x: -30.908, y: 206.178, z: -3.71502 },
+		     { x: -30.2759, y: 204.903, z: -3.51147 },
+		     { x: -29.2366, y: 203.054, z: -2.32263 },
+		     { x: -28.996, y: 202.619, z: -2.02235 },
+		     { x: -28.7804, y: 202.302, z: -1.61143 },
+		     { x: -28.6677, y: 202.121, z: -1.3357 },
+		     { x: -28.5691, y: 201.978, z: -1.1365 },
+		     { x: -28.4969, y: 201.849, z: -0.981609 },
+		     { x: -28.483, y: 201.836, z: -0.932947 },
+		     { x: -28.483, y: 201.836, z: -0.932947 },
+		     { x: -28.4782, y: 201.835, z: -0.904278 },
+		     { x: -28.4738, y: 201.833, z: -0.879297 },
+		     { x: -28.4356, y: 201.782, z: -0.850186 } ]
+    }
+];
